### 2025-01-21 #681 @antaldaniel
dataset: Create Data Frames that are Easier to Exchange and Reuse
Submitting Author Name: Daniel Antal
Submitting Author Github Handle: @antaldaniel
Repository:  https://github.com/dataobservatory-eu/dataset
Version submitted:  0.3.4002
Submission type: Standard
Reviewers: @maurolepore

Version accepted: TBD

---

```
Package: dataset
Title: Create Data Frames that are Easier to Exchange and Reuse
Version: 0.3.4002
DOI: 10.32614/CRAN.package.dataset
Language: en-GB
    c(person(given = "Daniel", family = "Antal",
           email = "daniel.antal@dataobservatory.eu",
           comment = c(ORCID = "0000-0001-7513-6760")
           ),
      person(given = "Marcelo", family =  "Perlin",
             role = c("rev"),
             comment = c(ORCID = "0000-0002-9839-4268")
             )
      )
Maintainer: Daniel Antal <daniel.antal@dataobservatory.eu>
Description: The aim of the 'dataset' package is to make tidy datasets easier to release,
    exchange and reuse. It organizes and formats data frame 'R' objects into well-referenced,
    well-described, interoperable datasets into release and reuse ready form.
    assertthat,
    cli,
    haven,
    ISOcodes,
    labelled,
    pillar,
    RefManageR,
    vctrs (>= 0.5.2)
    rdflib,
```

The package works with various semantic interoperability standards, therefore it allows the users to retrieve RDF annotated, rich and platform-independent data and reconstruct it as an R data.frame with rich metadata attributes, or to release interoperable, RDF annotated datasets on linked open data platforms from native R objects.

Production-side statisticans. Scientists who want to update their sources from various data repositories and exchanges. Scientists and research data managers who want to release new scientific or professional datasets that follow modern interoperability standards.

The package aimst to complement the rdflib and the datapsice package.

-   (If applicable) Does your package comply with our [guidance around _Ethics, Data Privacy and Human Subjects Research_](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)? Not applicable.

- [ x I have read the [author guide](https://devdevguide.netlify.app/authors-guide.html) and I expect to maintain this package for at least 2 years or to find a replacement.

- [xú contains a [README with instructions for installing the development version](https://ropensci.github.io/dev_guide/building.html#readme).
