### 2025-01-21 #681 @antaldaniel
Thank you @maelle and I will look in to why the coverage is not updating... However, do you have an explicit coverage target?
