### 2025-01-21 dataobservatory-eu/dataset#20 @antaldaniel
Incluse contributor policies
[Editor comments](https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603):

-  [ ] The contributing guide does not seem customized. It mentions a possible "src" folder which is not present.
-  [ ] You might want to rename the master branch to main as some people can be offended by the word "master", see https://www.tidyverse.org/blog/2021/10/renaming-default-branch/ that includes links to context, and practical advice on renaming the default branch.
- [ ] The iris dataset is very well-known, but it is also infamous because of its eugenics links.
Since having a good example dataset is very important, would you consider replacing it with another one, like maybe the [palmerpenguins](https://allisonhorst.github.io/penguins_paper_distill/rjarticle/penguins.html) one, even if it comes at the cost of adding a (possibly optional) dependency?
