### 2025-01-21 dataobservatory-eu/dataset#23 @antaldaniel
installation
[Editor's comment:](https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603)

I'd recommend documenting the two methods of installation (CRAN and GitHub) in distinct chunks so readers could copy-paste the entire code chunk of interest.

Instead of devtools you could recommend using pak.

install.packages("pak")
pak::pak("dataobservatory-eu/dataset")
