### 2025-01-21 dataobservatory-eu/dataset#24 @antaldaniel
Remove AppVeyor
[Editor comments](https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603): If AppVeyor is not used any more, please remove the related contributing file.
