### 2025-01-21 dataobservatory-eu/dataset#25 @antaldaniel
More consistent spacing
[Editor comments](https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603): I'd recommend running styler (on R scripts including tests) to make spacing more consistent.
