### 2025-01-21 dataobservatory-eu/dataset#26 @antaldaniel
Simpler inheritance checks
[Editor comments](https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603): The code could be simplified so that reviewers might more easily follow the logic. is_person <- function(p) ifelse (inherits(p, "person"), TRUE, FALSE) in https://github.com/dataobservatory-eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/R/as_dublincore.R#L11
