### 2025-01-21 #681 @antaldaniel
@maelle Thank you agian for your useful comments and your PRs.  I reorganised the issues and created a new milestone. The milestone currently breaks up your review comments to 7 issues, but they could of course be broken down to more.  I set myself a deadline of 16 February to resolve them, though it may happen earlier. I will tag you in the issues when they are ready for review and will also add a new comment here.
