### 2025-01-21 dataobservatory-eu/dataset#28 @antaldaniel
Discontinue the use of the iris dataset
[Editor comments](https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603):
The iris dataset is very well-known, but it is also infamous because of its eugenics links.
Since having a good example dataset is very important, would you consider replacing it with another one, like maybe the [palmerpenguins](https://allisonhorst.github.io/penguins_paper_distill/rjarticle/penguins.html) one, even if it comes at the cost of adding a (possibly optional) dependency?
