### 2025-01-21 dataobservatory-eu/dataset#28 @maelle
Fair enough!

Having something else than iris in user-facing docs would be great, thank you for considering it.
