### 2025-02-02 dataobservatory-eu/dataset#25 @antaldaniel
https://styler.r-lib.org/
https://style.tidyverse.org/
