### 2025-02-02 #681 @maurolepore
Dear @antaldaniel this is to mark the start of my EiC rotation. I'm reviewing all open issues and noting what I see:

* @maelle provided detailed editor checks.
* @antaldaniel plans to follow up by 16 February.
* Until then the submission is holding.
