+++ R/subject.R
@@ -14,11 +14,12 @@
 #' @examples
 #' # To set the subject of a dataset_df object:
 #' subject(iris_dataset) <- subject_create(
-#'          term  = "Irises (plants)",
-#'          schemeURI = "http://id.loc.gov/authorities/subjects",
-#'          valueURI = "https://id.loc.gov/authorities/subjects/sh85068079",
-#'          subjectScheme = "LCCH",
-#'          prefix = "lcch:")
+#'   term = "Irises (plants)",
+#'   schemeURI = "http://id.loc.gov/authorities/subjects",
+#'   valueURI = "https://id.loc.gov/authorities/subjects/sh85068079",
+#'   subjectScheme = "LCCH",
+#'   prefix = "lcch:"
+#' )
 #'
 #' # To retrieve the subject with its subproperties:
 #' subject(iris_dataset)
@@ -31,11 +32,12 @@
 #' @rdname subject
 subject <- function(x) {
   assertthat::assert_that(is.dataset_df(x),
-              msg = "subject(x): x must be a dataset_df object created with dataset_df() or as_dataset_df().")
+    msg = "subject(x): x must be a dataset_df object created with dataset_df() or as_dataset_df()."
+  )
 
-  if("subject" %in% names(attributes(x)))  {
+  if ("subject" %in% names(attributes(x))) {
     attr(x, "subject")
-  } else if(!is.null(get_bibentry(x)$subject)) {
+  } else if (!is.null(get_bibentry(x)$subject)) {
     get_bibentry(x)$subject
   } else {
     message("No subject is recorded.")
@@ -63,30 +65,33 @@ subject_create <- function(term,
                            valueURI = NULL,
                            prefix = NULL,
                            subjectScheme = NULL,
-                           classificationCode = NULL ) {
-
+                           classificationCode = NULL) {
   if (is.null(term)) term <- ":tba"
 
-  #if (! all.equal(length(heading), length(subjectScheme))) {
+  # if (! all.equal(length(heading), length(subjectScheme))) {
   #  stop("You must provide exactly one subjectSchemes, URIs and Codes for each heading.")
-  #}
+  # }
 
-  if (length(term)>1) {
-    dataset_subject <-  lapply (1:seq_along(term), function(x) new_Subject(term[x],
-                                                                subjectScheme = subjectScheme[x],
-                                                                schemeURI = schemeURI[x],
-                                                                classificationCode = classificationCode[x],
-                                                                prefix = prefix[x])
-    )
+  if (length(term) > 1) {
+    dataset_subject <- lapply(1:seq_along(term), function(x) {
+      new_Subject(term[x],
+        subjectScheme = subjectScheme[x],
+        schemeURI = schemeURI[x],
+        classificationCode = classificationCode[x],
+        prefix = prefix[x]
+      )
+    })
     # this is not nice
     class(dataset_subject) <- c("subject", class(subject))
   } else {
-    dataset_subject <- new_Subject(term=term,
-                                   subjectScheme=subjectScheme,
-                                   schemeURI=schemeURI,
-                                   valueURI=valueURI,
-                                   classificationCode = classificationCode,
-                                   prefix = prefix)
+    dataset_subject <- new_Subject(
+      term = term,
+      subjectScheme = subjectScheme,
+      schemeURI = schemeURI,
+      valueURI = valueURI,
+      classificationCode = classificationCode,
+      prefix = prefix
+    )
   }
 
   dataset_subject
@@ -99,27 +104,29 @@ new_Subject <- function(term,
                         valueURI = NULL,
                         prefix = NULL,
                         subjectScheme = NULL,
-                        classificationCode = NULL)
- {
-
[…]
