+++ tests/testthat/test-var_label.R
@@ -1,10 +1,9 @@
-
 test_that("var_label() works", {
   expect_equal(var_label(iris_dataset$Sepal.Length), "Length of the sepal in cm")
-  expect_equal(class(var_label(iris_dataset, unlist=TRUE)), "character")
-  test_df <- dataset_df(a = 1:2, b=3:4)
-  expect_equal(var_label(test_df, unlist=TRUE, null_action = "fill" ), c(rowid="rowid", a = "a", b="b"))
-  #expect_equal(label_attribute(iris_dataset$Species), "Taxon name within the Iris genus")
+  expect_equal(class(var_label(iris_dataset, unlist = TRUE)), "character")
+  test_df <- dataset_df(a = 1:2, b = 3:4)
+  expect_equal(var_label(test_df, unlist = TRUE, null_action = "fill"), c(rowid = "rowid", a = "a", b = "b"))
+  # expect_equal(label_attribute(iris_dataset$Species), "Taxon name within the Iris genus")
 })
 
 test_that("var_label()<-  works", {
@@ -16,20 +15,20 @@ test_that("var_label()<-  works", {
 
 
 test_that("var_label() throws error", {
-  test_df <- dataset_df(a = 1:2, b=3:4)
+  test_df <- dataset_df(a = 1:2, b = 3:4)
   expect_error(var_label(test_df) <- c("A", "B", "C", "E"))
 })
 
 
 test_that("var_label.dataset_df() works", {
-  expect_equal(length(var_label(iris_dataset, unlist=TRUE)), 6)
-  expect_true(is.character(var_label(iris_dataset, unlist=TRUE)))
+  expect_equal(length(var_label(iris_dataset, unlist = TRUE)), 6)
+  expect_true(is.character(var_label(iris_dataset, unlist = TRUE)))
 })
 
 test_that("var_label.dataset_df() works", {
-  d <- defined(x=1:2)
+  d <- defined(x = 1:2)
   var_label(d) <- "test"
   expect_equal(var_label(d), "test")
 })
 
-dataset_df(mtcars, identifier = c(mt="http://mtcars.com/"))
+dataset_df(mtcars, identifier = c(mt = "http://mtcars.com/"))
