+++ data-raw/prepare_iris_dataset.R
@@ -4,32 +4,38 @@ iris_doi <- "10.5281/zenodo.10396807"
 
 iris_dataset <- dataset_df(
   rowid = defined(paste0("#", row.names(iris)),
-                  label = "ID in the iris dataset",
-                  namespace=c("#"= "10.5281/zenodo.10396807")),
+    label = "ID in the iris dataset",
+    namespace = c("#" = "10.5281/zenodo.10396807")
+  ),
   Sepal.Length = defined(eg_iris$Sepal.Length,
-                         label = "Length of the sepal in cm",
-                         unit = "centimeter",
-                         definition = "https://www.wikidata.org/wiki/Property:P2043"),
+    label = "Length of the sepal in cm",
+    unit = "centimeter",
+    definition = "https://www.wikidata.org/wiki/Property:P2043"
+  ),
   Petal.Length = defined(eg_iris$Petal.Length,
-                         label = "Length of the petal in cm",
-                         unit = "centimeter",
-                         definition = "https://www.wikidata.org/wiki/Property:P2043"),
+    label = "Length of the petal in cm",
+    unit = "centimeter",
+    definition = "https://www.wikidata.org/wiki/Property:P2043"
+  ),
   Sepal.Width = defined(eg_iris$Sepal.Width,
-                        label = "Width of the sepal in cm",
-                        unit = "centimeter",
-                        definition = "https://www.wikidata.org/wiki/Property:P2049"),
+    label = "Width of the sepal in cm",
+    unit = "centimeter",
+    definition = "https://www.wikidata.org/wiki/Property:P2049"
+  ),
   Petal.Width = defined(eg_iris$Petal.Width,
-                        label = "Width of the petal in cm",
-                        unit = "centimeter",
-                        definition  = "https://www.wikidata.org/wiki/Property:P2049"),
+    label = "Width of the petal in cm",
+    unit = "centimeter",
+    definition = "https://www.wikidata.org/wiki/Property:P2049"
+  ),
   Species = defined(eg_iris$Species,
-                    label = "Taxon name within the Iris genus",
-                    definition = "https://npgsweb.ars-grin.gov/gringlobal/taxon/taxonomygenus?id=6074",
-                    namespace = "Iris" ),
-  dataset_bibentry =  dublincore(
+    label = "Taxon name within the Iris genus",
+    definition = "https://npgsweb.ars-grin.gov/gringlobal/taxon/taxonomygenus?id=6074",
+    namespace = "Iris"
+  ),
+  dataset_bibentry = dublincore(
     title = "Iris Dataset",
-    creator = person(given="Edgar", family="Anderson", role = "cre", comment=c(VIAF="http://viaf.org/viaf/6440526")),
-    contributor = person(given="Antal", family="Daniel", role = "dtm"),
+    creator = person(given = "Edgar", family = "Anderson", role = "cre", comment = c(VIAF = "http://viaf.org/viaf/6440526")),
+    contributor = person(given = "Antal", family = "Daniel", role = "dtm"),
     identifier = "https://doi.org/10.5281/zenodo.10396807",
     publisher = "American Iris Society",
     datasource = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
@@ -37,11 +43,13 @@ iris_dataset <- dataset_df(
     language = "en",
     description = "The famous (Fisher's or Anderson's) iris data set."
   ),
[…]
