### 2025-02-02 dataobservatory-eu/dataset#25 @antaldaniel
This [commit](https://github.com/dataobservatory-eu/dataset/commit/cc622c002e80933580fd4d41e38c25e3a4781bed) changes the code styling to the tidyverse style guide.
