### 2025-02-02 dataobservatory-eu/dataset#20 @antaldaniel
- [This commit](https://github.com/dataobservatory-eu/dataset/commit/cc622c002e80933580fd4d41e38c25e3a4781bed#diff-eca12c0a30e25b4b46522ebf89465a03ba72a03f540796c979137931d8f92055) adds a new contributor statement that is extended.
- [ ] the `src` is the standard location and it should be either used or explained that it should be used in the future.
