### 2025-02-02 #681 @antaldaniel
Hi @maurolepore,

I just solved three issues with new commits, so their evolution and solution is on the dataset package place.  I was hoping and I did receive some broader and useful review, as this is the early stages of the development of a package family.  But responding to those is perhaps best placed in an accompanying paper, not a package, for the purpose of this review, I add them as a vignette but it is a bit too philsophical for a normal vignette.

I will note when all are solved, although I have a very serious objection to one.

I think that the use of the `iris` dataset is a choice of the R-core team, and any criticism should be aimed there. This package wants to extend the usability of a central, core, base R object, i.e., the data.frame itself, which is taught, explained, tested in literally millions of use cases with the `iris` dataset.  If there was a debate on ditching the dataset on the R-core team, I would pitch in defense of it, and accept if it would be replaced with something else; however, I do not think that this is a valid point in the review of a package that extends the core R system.
