### 2025-02-07 #681 @maelle
Dear @antaldaniel,

We provide feedback on packages following our packaging guide, our and our reviewers' expertise as package developers, users, and contributors. Our goal in this process is to advance the R ecosystem in many dimensions, including maintainability, accessibility, and inclusivity. Many of our requirements and recommendations are greater or different from requirements from R-Core or CRAN.  Recommendations to the language of documentation to expand its accessibility, or to use a dataset that makes more users or contributors comfortable, are entirely within the scope of our review and its goals. We hope you can be receptive to these recommendations with the shared goal of making your package the best resource it can be for the R community.  Suggestions from editors and reviewers are suggestions, but we do expect authors to be receptive to this feedback so we can make good use of reviewers' time.

On the topic of the iris dataset in particular we understand that changing the dataset for the unit tests is a lot of work, and while the dataset may still be useful from a technical perspective, using an alternative that doesn't have this background would make the documentation more inclusive and aligned with ethical research practices. This inspired us to strive to better document the choice of data in our guidelines https://github.com/ropensci/dev_guide/issues/868 Our guidelines improve over time as we learn from submissions, so thank you!
