+++ README.md
@@ -16,8 +16,8 @@ Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci
 [![DOI](https://zenodo.org/badge/DOI/10.32614/CRAN.package.dataset.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
 [![devel-version](https://img.shields.io/badge/devel%20version-0.3.4002-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
-[![Coveralls test
-coverage](https://coveralls.io/repos/github/dataobservatory-eu/dataset/badge.svg)](https://coveralls.io/r/dataobservatory-eu/dataset?branch=master)
+[![Codecov test
+coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset)
 <!-- badges: end -->
 
 The aim of the *dataset* package is to make tidy datasets easier to
@@ -27,23 +27,23 @@ into release and reuse ready form.
 
 1.  **Increase FAIR use of your datasets**: Offer a way to better
     utilise the `utils:bibentry` bibliographic entry objects and working
-    with the ROpenSci package
+    with the rOpenSci package
     [RefManageR](https://docs.ropensci.org/RefManageR/) extending their
     fields of the Dublin Core and DataCite standards, and making them
     detachable from the data, i.e., including the bibliographic entries
     into the attributes of a data frame-like object. See for more
     information the [Bibentry for FAIR
     datasets](https://dataset.dataobservatory.eu/articles/bibentry.html)
-    vignette. 2.**Interoperability outside R**: Extending the
-    `haven_labelled` class of the `tidyverse` for consistently labelled
-    categorical variables with linked (standard) definitions and units
-    of measures in our
+    vignette.
+2.  **Interoperability outside R**: Extending the `haven_labelled` class
+    of the `tidyverse` for consistently labelled categorical variables
+    with linked (standard) definitions and units of measures in our
     [defined](https://dataset.dataobservatory.eu/articles/defined.html)
     class; this enables to share exact definitions, units of measures
     across computers and systems, and increasing the interoperability of
     the data set from an R data.frame to any standardised statistical or
     library system.
-2.  **Tidy data tidier, richer**: Offering a new data frame format,
+3.  **Tidy data tidier, richer**: Offering a new data frame format,
     `dataset_df` that extends tibbles with semantically rich metadata,
     ready to be shared on open data exchange platforms and in data
     repositories. This s3 class is aimed at developers and we are
@@ -53,11 +53,11 @@ into release and reuse ready form.
     Exchange and
     Extend](https://dataset.dataobservatory.eu/articles/dataset_df.html)
     vignette.
-3.  **R+RDF=global interoperability**: The [From R to
+4.  **R+RDF=global interoperability**: The [From R to
     RDF](https://dataset.dataobservatory.eu/articles/rdf.html) vignette
     shows how to leverage the capabilities of the *dataset* package with
     [rdflib](https://docs.ropensci.org/rdflib/index.html), an
-    R-user-friendly wrapper on ROpenSci to work with the *redland*
[…]
