### 2025-02-11 7f69578
Good improvements with lintr
