+++ tests/testthat/test-identifier.R
@@ -37,3 +37,4 @@ test_that("identifier()<- assignment works", {
   identifier(iris_dataset_2) <- "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
   expect_equal(identifier(iris_dataset_2), "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x")
 })
+
