+++ R/description.R
@@ -1,16 +1,19 @@
 #' @title Get/set the Description of the object.
-#' @description Get/set the optional \code{Description} property as an attribute to an R object.
-#' @details The \code{Description} is recommended for discovery in DataCite. All additional information that does not
-#' fit in any of the other categories. May be used for technical information. A free text.
-#' Similar to
-#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/}{dct:description}.
+#' @description Get/set the optional \code{Description} property as an attribute
+#'   to an R object.
+#' @details The \code{Description} is recommended for discovery in DataCite. All
+#'   additional information that does not fit in any of the other categories.
+#'   May be used for technical information. A free text. Similar to
+#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/}{dct:description}.
 #' @param x A dataset object created with \code{dataset::\link{dataset_df}} or
-#' \code{dataset::\link{as_dataset_df}}.
+#'   \code{dataset::\link{as_dataset_df}}.
 #' @param value The \code{Description} as a character set.
-#' @param overwrite If the \code{Description} attribute should be overwritten. In case it is set to \code{FALSE},
-#' it gives a message with the current \code{Description} property instead of overwriting it.
-#' Defaults to \code{FALSE}, when it gives a warning at an accidental overwrite attempt.
-#' @return The \code{Description} attribute as a character of length 1 is added to \code{x}.
+#' @param overwrite If the \code{Description} attribute should be overwritten.
+#'   In case it is set to \code{FALSE}, it gives a message with the current
+#'   \code{Description} property instead of overwriting it. Defaults to
+#'   \code{FALSE}, when it gives a warning at an accidental overwrite attempt.
+#' @return The \code{Description} attribute as a character of length 1 is added
+#'   to \code{x}.
 #' @examples
 #' description(iris_dataset, overwrite = TRUE) <- "The famous iris dataset used
 #' in R language examples."

[…]
