### 2025-02-14 e35516d
0.3.4005 fix dublincore date problem
