+++ tests/testthat/test-dublincore.R
@@ -25,7 +25,8 @@ test_that("fix_contributors() works", {
 
 test_that("new_dublincore() works", {
   expect_equal(
-    new_dublincore(title = "Test", creator = person("Jane", "Doe", role = "cre"))$author,
+    new_dublincore(title = "Test",
+                   creator = person("Jane", "Doe", role = "cre"))$author,
     person("Jane", "Doe", role = "cre")
   )
   expect_equal(
@@ -63,7 +64,7 @@ test_that("new_dublincore() works", {
     new_dublincore(
       title = "Test",
       creator = person("Jane", "Doe", role = "cre"),
-      publication_date = 1935
+      dataset_date = 1935
     )$date,
     "1935"
   )
@@ -138,7 +139,8 @@ test_that("dublincore() works", {
 test_that("as_dublincore() works", {
   expect_true(is.dublincore(as_dublincore(x = iris_dataset)))
   expect_true(is.list(as_dublincore(x = iris_dataset, type = "list")))
-  expect_equal(as_dublincore(iris_dataset, type = "list")$contributor, "Antal Daniel [dtm]")
+  expect_equal(as_dublincore(iris_dataset, type = "list")$contributor,
+               "Antal Daniel [dtm]")
   expect_equal(as_dublincore(iris_dataset)$date, "1935")
   expect_equal(as_dublincore(iris_dataset)$description, "The famous (Fisher's or Anderson's) iris data set.")
   expect_equal(as_dublincore(iris_dataset)$rights, ":tba")
