### 2025-02-14 4d0f9e4
improved handling of exception messages in tests, and some text improvements
