+++ tests/testthat/test-geolocation.R
@@ -2,7 +2,8 @@ test_that("geolocation() works", {
   iris_dataset_2 <- iris_dataset
   geolocation(iris_dataset_2) <- "US"
   expect_equal(geolocation(iris_dataset_2), "US")
-  expect_message(geolocation(iris_dataset_2, overwrite = FALSE) <- "GB")
+  expect_message(geolocation(iris_dataset_2, overwrite = FALSE) <- "GB",
+                 regexp = "dataset has already an Geolocation")
 })
 
 test_that("geolocation(x) <- NULL works", {
