+++ R/dublincore.R
@@ -380,6 +380,7 @@ new_dublincore <- function(title,
     bibtype = "Misc",
     title = title,
     author = creator,
+    year = year,
     identifier = identifier,
     publisher = publisher,
     contributor = contributor,
