+++ vignettes/Motivation.Rmd
@@ -22,9 +22,69 @@ library(dataset)
 
 The `dataset` package aims to enhance the semantic richness of R datasets, particularly tabular data derived from the `data.frame` class. Its goal is to create "tidy" datasets from the outset, promoting reusability and interoperability. Through several iterations, it became clear that this goal is intrinsically linked to the dataset's intended purpose. Consequently, `dataset` has evolved from a single package into a family of related packages.
 
+## Problem statement
+
+```{r example-ambiguity}
+data.frame(
+   geo=c("LI", "SM"), 
+   CPI = c("0.8", "0.9"), 
+   GNI= c("8976", "9672"))
+```
+This dataset is tidy. But it certainly could be improved!
+
+- `geo`: you may figure out that geo means something to do with geography (maybe countries, "LI" standing for Lichtenstein and "SM" for San Marino), but even that, who knows? If you add Greece, should use use "EL" like Eurostat or "GR" like the World Bank?
+
+- `CPI` can stand for "consumer price index" or "corruption perceptions index". Or anything else! 
+
+- `GNI` can stand for "Gross National Income" and "Global Nutrition Index", often mentioned in the same context. If there are numbers, do they mean a physical quantity or a currency? Is the currency dollars or euros?
+
+
+```{r smallcountrydataset}
+options(sciphen = 4)
+small_country_dataset <- dataset_df(
+  country_name = defined(
+    c("AD", "LI"),
+    definition = "http://data.europa.eu/bna/c_6c2bb82d",
+    namespace = "https://www.geonames.org/countries/$1/"
+  ),
+  gdp = defined(
+    c(3897, 7365),
+    label = "Gross Domestic Product",
+    unit = "million dollars",
+    definition = "http://data.europa.eu/83i/aa/GDP"
+  ),
+  population = defined(
+    c(77543, 40015),
+    label = "Population",
+    definition = "http://data.europa.eu/bna/c_f2b50efd"
+  ),
+  dataset_bibentry = dublincore(
+    creator = person(given = "Jane", family = "Doe"),
+    title = "Small Country Dataset",
+    publisher = "Reprex"
+  )
+)
+```
+
+```{r percapita}
+small_country_dataset$gdp_capita <- defined(
+  small_country_dataset$gdp * 1e6 / small_country_dataset$population,
+  unit = "dollar",
+  label = "GDP Per Capita"
+)
+```
+
+Interoperability and future (re)usability depends on the amount and quality of the metadata that was generated, recorded, and released together with the data. The `dataset` package aims to collect such metadata and record them in the least possible intrusive way.
+
+-    [x]  `Descriptive metadata`, about the dataset as a whole, to improve findability and accessibility; in this regard, follow the DCTERMS and DataCite standards.
+-    [x]  Metadata about the `structural elements` within the dataset; better definitions of the row identifier (instead of local, encourage global IDs), and utilise global definitions for variables. Include unit of measure to disallow semantically incorrect joins.
+-    [x]  Add `provenance metadata`, so the next user can understand where each data elements came, and what happened with them.
+
+## Limites on reusability
+
 Let's take a look at the limitations of two, tidy datasets from an interoperability and reusability point of view:
 
-```{r}
+```{r gdpexmampel}
 library(tibble)
 library(dplyr)
 library(tidyr)
@@ -60,7 +120,7 @@ Our work is informed by the European Interoperability Framework (EIF) and its pr
 
 Although we envision diverse future applications for these enhanced datasets, the current development focuses on a key distinction: whether the data will be reused in a tabular (relational database) or graph format. These two reuse scenarios have distinct metadata requirements, necessitating somewhat specialized approaches.
 
-## Tidy data and graphs
+### Tidy data and graphs
 
[…]
