+++ vignettes/Motivation.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-> "A dataset is a structured, coherent collection of related data, organised for a particular purpose." ISO 19115 and ISO/IEC 11179
+> "A dataset is an identifiable collection of data available for access or download in one or more formats." ISO/IEC 20546
 
 ```{r setup}
 library(dataset)
@@ -30,14 +30,14 @@ data.frame(
    CPI = c("0.8", "0.9"), 
    GNI= c("8976", "9672"))
 ```
+
 This dataset is tidy. But it certainly could be improved!
 
-- `geo`: you may figure out that geo means something to do with geography (maybe countries, "LI" standing for Lichtenstein and "SM" for San Marino), but even that, who knows? If you add Greece, should use use "EL" like Eurostat or "GR" like the World Bank?
+-   `geo`: you may figure out that geo means something to do with geography (maybe countries, "LI" standing for Lichtenstein and "SM" for San Marino), but even that, who knows? If you add Greece, should use use "EL" like Eurostat or "GR" like the World Bank?
 
-- `CPI` can stand for "consumer price index" or "corruption perceptions index". Or anything else! 
+-   `CPI` can stand for "consumer price index" or "corruption perceptions index". Or anything else!
 
-- `GNI` can stand for "Gross National Income" and "Global Nutrition Index", often mentioned in the same context. If there are numbers, do they mean a physical quantity or a currency? Is the currency dollars or euros?
-
+-   `GNI` can stand for "Gross National Income" and "Global Nutrition Index", often mentioned in the same context. If there are numbers, do they mean a physical quantity or a currency? Is the currency dollars or euros?
 
 ```{r smallcountrydataset}
 options(sciphen = 4)
@@ -76,11 +76,11 @@ small_country_dataset$gdp_capita <- defined(
 
 Interoperability and future (re)usability depends on the amount and quality of the metadata that was generated, recorded, and released together with the data. The `dataset` package aims to collect such metadata and record them in the least possible intrusive way.
 
--    [x]  `Descriptive metadata`, about the dataset as a whole, to improve findability and accessibility; in this regard, follow the DCTERMS and DataCite standards.
--    [x]  Metadata about the `structural elements` within the dataset; better definitions of the row identifier (instead of local, encourage global IDs), and utilise global definitions for variables. Include unit of measure to disallow semantically incorrect joins.
--    [x]  Add `provenance metadata`, so the next user can understand where each data elements came, and what happened with them.
+-   [x] `Descriptive metadata`, about the dataset as a whole, to improve findability and accessibility; in this regard, follow the DCTERMS and DataCite standards.
+-   [x] Metadata about the `structural elements` within the dataset; better definitions of the row identifier (instead of local, encourage global IDs), and utilise global definitions for variables. Include unit of measure to disallow semantically incorrect joins.
+-   [x] Add `provenance metadata`, so the next user can understand where each data elements came, and what happened with them.
 
-## Limites on reusability
+## Limits on reusability
 
 Let's take a look at the limitations of two, tidy datasets from an interoperability and reusability point of view:
 
@@ -126,9 +126,9 @@ While the semantic and descriptive metadata needs of graph and tabular data aren
 
 -   `datacube`: This planned downstream package prioritizes interoperability with the Statistical Data and Metadata Exchange (SDMX) standard. It enforces stricter adherence to the datacube model when working with `data.frame`s or `tibble`s. The target user is an analyst working with SDMX data sources (e.g., national statistical offices, Eurostat, World Bank). The goal is to enable seamless data transformation, extension, and mutation within R, followed by interoperable export back to an SDMX resource (e.g., the EU Open Data portal).
 
[…]
