+++ vignettes/Motivation.Rmd
@@ -30,6 +30,7 @@ data.frame(
    CPI = c("0.8", "0.9"), 
    GNI= c("8976", "9672"))
 ```
+
 This dataset is tidy. But it certainly could be improved!
 
 - `geo`: you may figure out that geo means something to do with geography (maybe countries, "LI" standing for Lichtenstein and "SM" for San Marino), but even that, who knows? If you add Greece, should use use "EL" like Eurostat or "GR" like the World Bank?
@@ -136,7 +137,7 @@ In these cases, we assume users work with tabular data structures like `data.fra
 
 This document describes a new R package, `dataset`, designed to improve data interoperability for R's `data.frame` objects. The core concept is leveraging R's built-in `attributes()` function to store rich metadata. While `attributes()` allows attaching metadata to any R object, it's often underutilised. The `dataset` package promotes systematic use of attributes, ensuring crucial metadata isn't lost during data manipulation.
 
-Specifically, `dataset` utilizes `person` and extended `bibentry` objects (from the `utils` package) within the attributes. This allows comprehensive descriptive metadata to be associated with the dataset. The package focuses on adhering to either DCTERMS (Dublin Core) or DataCite metadata standards, maintaining and even updating this metadata throughout the R workflow.
+Specifically, `dataset` utilizes `person` and extended `bibentry` objects (from the `utils` package) within the attributes. This allows comprehensive descriptive metadata to be associated with the dataset. The package focuses on adhering to either [DCTERMS](https://www.dublincore.org/specifications/dublin-core/dcmi-terms/) (Dublin Core) or DataCite metadata standards, maintaining and even updating this metadata throughout the R workflow.
 
 `person` objects, crucial for representing agents (creators, contributors, publishers) within the `bibentry`, enable structured metadata and support permanent identifiers like ORCiDs. While acknowledging limitations of `person` objects, their presence in the `utils` package ensures availability for all R users. (An initial attempt to use `RefManageR` objects for bibliographic information was abandoned due to conflicts with modern metadata standards.)
 
@@ -209,7 +210,8 @@ library(dplyr)
 D <- tibble(
   year = c(2020, 2020, 2021, 2021, 2022, 2022),
   geocode = c("USA", "CAN", "USA", "CAN", "USA", "CAN"),
-  country_name = c("United States", "Canada", "United States", "Canada", "United States", "Canada"),
+  country_name = c("United States", "Canada", "United States", 
+                   "Canada", "United States", "Canada"),
   GDP = c(21000, 2000, 22000, 2100, 23000, 2200) # GDP in billions of USD
 )
 
@@ -231,9 +233,9 @@ The `datacube` package emphasizes the distinction between these variable types. 
 
 The ROpenSci `rdflib` package, a wrapper for the Python library of the same name, provides powerful tools for reading and writing graph data. We believe `dataset` and `rdflib` are complementary and should be used together. While there's a small overlap (inspired by an internal `rdflib` function for working with NQuads, a form of RDF), we've avoided making `rdflib` a direct dependency of `dataset`. However, for users working with RDF-annotated graphs, we strongly recommend using `dataset` in conjunction with `rdflib` for importing, exporting, and exchanging data.
 
[…]

[…]
