+++ README.Rmd
@@ -76,6 +76,8 @@ iris_ds <- dataset_df(
 )
 ```
 
+You can read more about the history of this dataset, and some controversy around its association of problematic science in the [Bibentry for FAIR datasets](https://dataset.dataobservatory.eu/articles/bibentry.html) vignette.
+
 It is mandatory to add a `title`, `author` to a dataset, and if the `date` is not specified, the current date will be added. 
 
 As the `dataset_df` at this point is just created, if it is not published yet, the `identifer` receives the default `:tba` value, a `version` of 0.1.0 and the `:unas` (unassigned) `publisher` field.
