+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4008
-Date: 2025-02-16
+Version: 0.3.4009
+Date: 2025-02-19
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
 Authors@R: 
