+++ tests/testthat/test-provenance.R
@@ -1,6 +1,7 @@
 test_that("provenance(x) works", {
   expect_error(provenance(mtcars),
-               regexp="must be a dataset_df object with standardised provenance metadata")
+    regexp = "must be a dataset_df object with standardised provenance metadata"
+  )
   tested <- dataset_df(x = mtcars)
   expect_true(class(provenance(iris_dataset)[1]) == "character")
   expect_output(print(provenance(tested)), "<http://example.com/dataset#>")
