+++ tests/testthat/test-geolocation.R
@@ -3,7 +3,8 @@ test_that("geolocation() works", {
   geolocation(iris_dataset_2) <- "US"
   expect_equal(geolocation(iris_dataset_2), "US")
   expect_message(geolocation(iris_dataset_2, overwrite = FALSE) <- "GB",
-                 regexp = "dataset has already an Geolocation")
+    regexp = "dataset has already an Geolocation"
+  )
 })
 
 test_that("geolocation(x) <- NULL works", {
