+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4010
+Version: 0.3.4011
 Date: 2025-02-20
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
@@ -43,6 +43,7 @@ Imports:
     vctrs (>= 0.5.2)
 RoxygenNote: 7.3.2
 Suggests: 
+    datasets,
     knitr,
     rdflib,
     rmarkdown,
