+++ R/dataset-orange_df.R
@@ -1,0 +1,19 @@
+#' @title Growth of Orange Trees
+#'
+#' @description The Orange data frame has 35 rows and 3 columns of records of
+#'   the growth of orange trees. This is a replication of
+#'   \code{datasets::\link[datasets]{Orange}} as \emph{dataset_df} s3 class.
+#'
+#' @details See \code{datasets::\link[datasets]{Orange}} for details.
+#'
+#' @format
+#' orange_df is a data frame with 35 cases (rows) and 3 variables (columns) named
+#' rowid, tree, age, circumference
+#' @references
+#' Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
+#' Wadsworth & Brooks/Cole.
+#' @source
+#' Draper, N. R. and Smith, H. (1998), Applied Regression Analysis (3rd ed),
+#' Wiley (exercise 24.N). Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects
+#' Models in S and S-PLUS, Springer.
+"orange_df"
