+++ README.md
@@ -25,6 +25,140 @@ release, exchange and reuse. It organizes and formats data frame R
 objects into well-referenced, well-described, interoperable datasets
 into release and reuse ready form.
 
+<!---
+&#10;The primary aim of dataset is create well-referenced, well-described, interoperable datasets from data.frames, tibbles or data.tables that translate well into the W3C DataSet definition within the [Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) in a reproducible manner. The data cube model in itself is is originated in the _Statistical Data and Metadata eXchange_, and it is almost fully harmonized with the Resource Description Framework (RDF), the standard model for data interchange on the web^[RDF Data Cube Vocabulary, W3C Recommendation 16 January 2014  <https://www.w3.org/TR/vocab-data-cube/>, Introduction to SDMX data modeling <https://www.unescap.org/sites/default/files/Session_4_SDMX_Data_Modeling_%20Intro_UNSD_WS_National_SDG_10-13Sep2019.pdf>].
+&#10;--->
+
+You can install the latest CRAN release with
+`install.packages("dataset")`, and the latest development version of
+dataset with `remotes::install_github()`:
+
+``` r
+install.packages("dataset")
+remotes::install_github("dataobservatory-eu/dataset", build = FALSE)
+```
+
+The current version of the `dataset` package is in an early,
+experimental stage. You can follow the discussion of this package on
+[rOpenSci \#553](https://github.com/ropensci/software-review/issues/553)
+about the original scope, that included the datacube data model, and the
+[rOpenSci \#681](https://github.com/ropensci/software-review/issues/681)
+on the new version that moves the data cube data model of SDMX into a
+future downstream package. (See, again, the
+[Motivation](https://dataset.dataobservatory.eu/articles/Motivation.html)
+article.)
+
+Interoperability and future (re)usability depends on the amount and
+quality of the metadata that was generated, recorded, and released
+together with the data. The `dataset` package aims to collect such
+metadata and record them in the least possible intrusive way.
+
+## Semantically richer data frames
+
+Let’s take a simple data.frame from the datasets package. The “Growth of
+Orange Trees” dataset contains 35 rows and 3 columns that record the
+growth of orange trees.
+
+``` r
+library(datasets)
+head(datasets::Orange)
+#>   Tree  age circumference
+#> 1    1  118            30
+#> 2    1  484            58
+#> 3    1  664            87
+#> 4    1 1004           115
+#> 5    1 1231           120
+#> 6    1 1372           142
+```
+
+Following the tidy data principle, we create an unambiguous row
+identifier. Then we go three steps further:
+
+1.  We add more semantic information about the meaning of the variables,
+    for example, to avoid joining numeric variables of the same type
+    (numeric or integer) but different unit of measure (mm vs cm.)
+
+``` r
+library(dataset)
+data("orange_df")
+orange_df
+#> Draper N, Smith H (1998). "Growth of Orange Trees."
+#>    rowid      tree       age        circumference
+#>    <hvn_lbl_> <hvn_lbl_> <hvn_lbl_> <hvn_lbl_>   
+#>  1 orange:1   2 [1]       118        30          
+#>  2 orange:2   2 [1]       484        58          
+#>  3 orange:3   2 [1]       664        87          
[…]
