+++ DESCRIPTION
@@ -31,6 +31,7 @@ LazyData: true
 Imports: 
     assertthat,
     cli,
+    datasets,
     haven,
     ISOcodes,
     labelled,
@@ -43,7 +44,6 @@ Imports:
     vctrs (>= 0.5.2)
 RoxygenNote: 7.3.2
 Suggests: 
-    datasets,
     knitr,
     rdflib,
     rmarkdown,

[…]
