+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4012
-Date: 2025-02-20
+Version: 0.3.4013
+Date: 2025-02-24
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
 Authors@R: 
