+++ data-raw/orange_df.R
@@ -43,7 +43,6 @@ orange_bibentry <- dublincore(
   description = "The Orange data frame has 35 rows and 3 columns of records of the growth of orange trees."
 )
 
-?dublincore
 
 orange_df$rowid <- defined(orange_df$rowid, namespace="https://doi.org/10.5281/zenodo.14917851")
 write.csv(orange_df, file = "orange_df.csv", row.names = F)
