+++ R/description.R
@@ -17,8 +17,9 @@
 #' @examples
 #' description(orange_df)
 #' description(
-#'    iris_dataset,
-#'    overwrite = TRUE) <- "The 'orange' dataset has 35 rows and 3 columns
+#'   iris_dataset,
+#'   overwrite = TRUE
+#' ) <- "The 'orange' dataset has 35 rows and 3 columns
 #'                           of records of the growth of orange trees."
 #' description(iris_dataset)
 #' @family Reference metadata functions
