+++ man/description.Rd
@@ -37,8 +37,9 @@ May be used for technical information. A free text. Similar to
 \examples{
 description(orange_df)
 description(
-   iris_dataset,
-   overwrite = TRUE) <- "The 'orange' dataset has 35 rows and 3 columns
+  iris_dataset,
+  overwrite = TRUE
+) <- "The 'orange' dataset has 35 rows and 3 columns
                           of records of the growth of orange trees."
 description(iris_dataset)
 }

[…]
