### 2025-03-20 ffb9f81
better error handling for incompatible defined vectors
