### 2025-03-20 aaa8eca
error in bind_defined_rows documentation
