+++ man/description.Rd
@@ -37,11 +37,11 @@ May be used for technical information. A free text. Similar to
 \examples{
 description(orange_df)
 description(
-  iris_dataset,
+  orange_df,
   overwrite = TRUE
 ) <- "The 'orange' dataset has 35 rows and 3 columns
                           of records of the growth of orange trees."
-description(iris_dataset)
+description(orange_df)
 }
 \seealso{
 Other Reference metadata functions: 
