### 2025-03-20 a0863fe
new tests and namespace alignment for rowid
