### 2025-03-25 d7d3ab8
improved documentation remove iris better coercion of defined vectors
