### 2025-03-25 b7fcc27
0.3.4018
Many improvements with bind_defined_rows
