+++ man/xsd_convert.Rd
@@ -59,7 +59,7 @@ Convert the numeric, boolean and Date/time columns of a dataset
 \examples{
 
 # Convert data.frame to XML Schema Definition
-xsd_convert(data.frame( a=1:3, b=c("a", "b", "c")))
+xsd_convert(data.frame(a = 1:3, b = c("a", "b", "c")))
 
 # Convert dataset to XML Schema Definition
 xsd_convert(head(dataset_df(orange_df)))
