### 2025-04-22 f84c672
0.3.4019
The iris dataset is removed upon request from rOpenSci reviewers
