+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4018
-Date: 2025-03-25
+Version: 0.3.4019
+Date: 2025-04-22
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
 Authors@R: 
