### 2025-04-22 #681 @antaldaniel
Hi @maelle ,  since the last update I think that I removed the Anderson dataset from all examples and vignettes; the rest of your requests was already ready. I just made a last check now, and commited to main branch (no longer master) Version: 0.3.4019, today's date, is in my opinion has what you requested.
