### 2025-04-22 #681 @maelle
Thank you @antaldaniel! Could you please copy the comments from https://github.com/ropensci/software-review/issues/681#issuecomment-2604134603, quote them and answer below each of them to describe what has been tackled, what hasn't been? [Example of such a response](https://github.com/ropensci/software-review/issues/686#issuecomment-2687367945). The format makes it easier to follow conversations. Thanks!
