### 2025-04-24 a5bfc31
- Added support for generic vector methods: `length()`, `head()`, `tail()`, `as.vector()`, `as.list()`, and subsetting (`[`, `[[`). - Implemented comparison methods (`==`, `<`, `>`, etc.) that operate on the underlying data while maintaining semantic integrity. - Introduced custom `print()` and `format()` methods that summarise metadata (label, unit, definition) in a concise and human-readable manner. - Improved the `summary()` method for `defined` vectors to display variable metadata and integrate seamlessly with base R statistics. - Enhanced the `c()` method to validate compatibility across 
[…]
