+++ R/utils.R
@@ -1,0 +1,4 @@
+#' @keywords internal
+`%||%` <- function(x, y) {
+  if (is.null(x)) y else x
+}
