+++ vignettes/Motivation.Rmd
@@ -20,7 +20,9 @@ knitr::opts_chunk$set(
 library(dataset)
 ```
 
-The `dataset` package aims to enhance the semantic richness of R datasets, particularly tabular data derived from the `data.frame` class. Its goal is to create "tidy" datasets from the outset, promoting reusability and interoperability. Through several iterations, it became clear that this goal is intrinsically linked to the dataset's intended purpose. Consequently, `dataset` has evolved from a single package into a family of related packages.
+The `dataset` package is designed to enhance the **semantic richness of R datasets**, particularly those structured as `data.frame` or `tibble` objects. Its goal is to promote reusability and interoperability by enabling tidy datasets to carry rich metadata — from the start.
+
+Through practical iterations, it became evident that the structure of a dataset cannot be separated from its **purpose**. As a result, the `dataset` package has evolved into a family of interrelated tools that support semantic clarity at both the column and dataset level.
 
 ## Problem statement
 
@@ -32,13 +34,15 @@ data.frame(
 )
 ```
 
-This dataset is tidy. But it certainly could be improved!
+At first glance, this dataset appears tidy — columns are variables, rows are observations. But is it interpretable?
 
--   `geo`: you may figure out that geo means something to do with geography (maybe countries, "LI" standing for Lichtenstein and "SM" for San Marino), but even that, who knows? If you add Greece, should use use "EL" like Eurostat or "GR" like the World Bank?
+-   `geo`: This might refer to geography — but is "LI" Lichtenstein? What about "SM"? Is this ISO alpha-2? Eurostat’s codes? World Bank’s?
 
--   `CPI` can stand for "consumer price index" or "corruption perceptions index". Or anything else!
+-   `CPI`: Is this the Consumer Price Index, or the Corruption Perceptions Index? Something else entirely?
 
--   `GNI` can stand for "Gross National Income" and "Global Nutrition Index", often mentioned in the same context. If there are numbers, do they mean a physical quantity or a currency? Is the currency dollars or euros?
+-   `GNI`: It could mean Gross National Income or Global Nutrition Index — and even if it's GNI, is it measured in dollars, euros, or something else?
+
+A tidy structure doesn't guarantee **semantic clarity**. Without metadata, every column is open to misinterpretation.
 
 ```{r smallcountrydataset}
 options(sciphen = 4)
@@ -75,17 +79,21 @@ small_country_dataset$gdp_capita <- defined(
 )
 ```
 
-Interoperability and future (re)usability depends on the amount and quality of the metadata that was generated, recorded, and released together with the data. The `dataset` package aims to collect such metadata and record them in the least possible intrusive way.
+The **interoperability** and **future reusability** of data depends on the quality and presence of metadata — not just the structure of the data itself. The `dataset` package captures metadata as seamlessly and non-intrusively as possible.
 
--   [x] `Descriptive metadata`, about the dataset as a whole, to improve findability and accessibility; in this regard, follow the DCTERMS and DataCite standards.
--   [x] Metadata about the `structural elements` within the dataset; better definitions of the row identifier (instead of local, encourage global IDs), and utilise global definitions for variables. Include unit of measure to disallow semantically incorrect joins.
--   [x] Add `provenance metadata`, so the next user can understand where each data elements came, and what happened with them.
+It covers:
+
+-   [x] `Descriptive metadata`: Follows DCTERMS and DataCite standards to describe datasets for searchability, citation, and sharing.
+
+-   [x] `Structural metadata`: Uses well-defined variable names, consistent units of measure, and encourages globally resolvable identifiers for rows and values.
+
+-   [x] `Provenance metadata`: Tracks where the data came from and what happened to it — enabling transparent, reproducible data workflows.
 
 ## Limits on reusability
 
[…]
