+++ DESCRIPTION
@@ -32,9 +32,11 @@ Imports:
     assertthat,
     datasets,
     haven,
+    ISOcodes,
     labelled,
     methods,
     pillar,
+    RefManageR,
     rlang,
     tibble,
     utils,
