### 2025-04-24 #681 @antaldaniel
>On the one hand, it'd be interesting to read how dataset compares to other approaches to the same "problem", such as (if I follow correctly)
> https://docs.ropensci.org/frictionless/
https://docs.ropensci.org/datapack/
https://datapackage.org/
https://www.researchobject.org/
> On the other hand, how would an user take advantage of dataset?
To me, it is not clear yet from reading the docs.
Questions I wonder about:
As a data publisher, I create the dataset object, and then, how does it help me document it? How does it help me publish it on a [repository](https://docs.ropensci.org/deposits/)?

This is now explained in detail in the new [vignette](https://dataset.dataobservatory.eu/articles/Motivation.html).

Repositories like Zenodo apply at most, but usually with limitations, the 5-star FAIR data model, which helps with finding data, but does not make the data readily reusable. We apply more the 8-star model, which means that we put metadata about the contents of the data, not only the file.  On Zenodo, you can upload three different datasets about GDP per capita in countries, one in dollars, the other in euros, and yet another on PPP adjusted euros.  These datasets are not very reusable, because if a user accidentally joins them, it will result in logical-semantic errors; for example, euro values will be compared or added to dollar values.  Or, in natural sciences, measures of units may mismatch.

Dataset contain the machine readable definition of each variable, it unit of measure, and other metadata. To my knowledge, apart from rdflib, no other R pakcages help in that. We designed a set of packages that work with rdflib.

> When you mention standard statistical libraries in the README, could you name some?

We aim for full compliance with the GSIM system, which is partly the StatisticaL Data and Metadata eXchange (so our datasets can be exchanged, enriched, etc. with IMF,Eurostat, French national statitical office data) and DDI (the standard for microdata.)

> As a data consumer, how do I create a dataset object (do I get it from an R package? shared in another way)? How can I easily use the information on units when exploring the data, when plotting it?

You use the dataset_df constructor that enriches a data.frame or a tibble object. The print method use this information.

> In short, could you exemplify "release" and "re-use" in a vignette or more, as use cases, potentially using as roles the type of users you mention in the submission under "target audience".

Release means that you can create datasets that you can send to the EU Open Data Portal, which has far more strict requirements than Zenodo. It is meant for people who want to machine-exchange data with professional scientific or statistical data providers, even on API level.

> For instance https://wbdataset.dataobservatory.eu/ is a good example, but it is mentioned in a vignette.
More concrete information like wbdataset should make it to the README to make it clearer what dataset is about (and then be expanded in vignettes).

>Installation instructions
>I'd recommend documenting the two methods of installation (CRAN and GitHub) in distinct chunks so readers could copy-paste the entire code chunk of interest.

Done

>Contributing guide
>The contributing guide does not seem customized.
It mentions a possible "src" folder which is not present. Since you are looking for co-developers, and mentioned one of the articles could be relevant to potential contributors, I'd recommend having some text related to design and wishes for feedback in the contributing guide.

The Guide is rewritten.

>The contributing guide mentions "AppVeyor" which is not used any more as far as I can tell. If AppVeyor is not used any more, please remove the related contributing file.

It is removed

>Project management
From the open issues, which ones are meant to be tackled soon?

All issues tacked and wait for your review.

>Code style
I'd recommend running styler (on R scripts including tests) to make spacing more consistent._

styler is used now with tidyverse settings.

> is_person <- function(p) ifelse (inherits(p, "person"), TRUE, FALSE) in https://github.com/dataobservatory-eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/R/as_dublincore.R#L11
>could be is_person <- function(p) inherits(p, "person") (thanks [lintr](https://masalmon.eu/2024/08/28/lintr-3-steps/) for catching this).
That pattern comes up several times in the codebase (the relevant linter is "redundant_ifelse_linter")._

This issues are resolved.

I hope that all such cases are handled.

>Since dataset imports rlang, you could use the %||% operator from rlang.
For instance https://github.com/dataobservatory-eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/R/agent.R#L24

I prefer to remain as close to base R as possible and later remove some dependencies, possibly even rlang.  So I prefer to use less of rlang not more.  However, for now I created an internal function for the %||% and there are few examples of its use.

>Example dataset
>The iris dataset is very well-known, but it is also infamous because of its eugenics links.
Since having a good example dataset is very important, would you consider replacing it with another one, like maybe the [palmerpenguins](https://allisonhorst.github.io/penguins_paper_distill/rjarticle/penguins.html) one, even if it comes at the cost of adding a (possibly optional) dependency?_

I think that the way rOpenSci handles this issue could improve, because the dataset was never used in eugenics. I find boycotting an innocent author, i.e., Edgar Anderson,  unethical. There are better practices in handling problematic provenance descriptions (because in this case it is the provenance description that is associated with eugenics, not the dataset or its author) and scientific solutions about such problems from digital humanities and digital heritage management could  be introduced.  But this could be the the topic of a peer-reviewed article itself.  So, upon your request, I removed the iris dataset from all vignetes and examples.  It uses now the Orange dataset.

>Tests
Should the line https://github.com/dataobservatory-eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/tests/testthat/test-agent.R#L9 be removed as it is not used in the test?

Solved.

>I don't understand why the iris object needs to be duplicated in lines like https://github.com/dataobservatory-eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/tests/testthat/test-creator.R#L23
>expect_true(is.dataset_df might become a custom expectation and/or rely on expect_s3_class() instead. Same comment for >expect_true(is.subject.
> https://github.com/dataobservatory-eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/tests/testthat/test-datacite.R#L18
>should be
>expect_type(as_datacite(iris_dataset, "list"), "list")
>What is https://github.com/dataobservatory-eu/dataset/blob/master/tests/testthat/test-dataset_prov.bak?
>When using expect_error() as in https://github.com/dataobservatory->eu/dataset/blob/7bf85ac7abe9477d02b429b4d335179d94993a77/tests/testthat/test-dataset_title.R#L3 maybe add a pattern for the error message, just in case another error happens?_

These minor issues are solved and there are many new texts.
