### 2025-04-24 ceb5935
plot method README
