+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# The dataset R Package <a href='https://dataset.dataobservatory.eu/'><img src='man/figures/logo.png' align="right" /></a>
+# The dataset R Package <a href='https://dataset.dataobservatory.eu/'><img src="man/figures/logo.png" align="right"/></a>
 
 <!-- badges: start -->
 
@@ -14,10 +14,11 @@ WIP](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.o
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci/software-review/issues/553)
 [![DOI](https://zenodo.org/badge/DOI/10.32614/CRAN.package.dataset.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
-[![devel-version](https://img.shields.io/badge/devel%20version-0.3.4017-blue.svg)](https://github.com/dataobservatory-eu/dataset)
+[![devel-version](https://img.shields.io/badge/devel%20version-0.3.4021-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 [![Codecov test
 coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset)
+
 <!-- badges: end -->
 
 The aim of the *dataset* package is to make tidy datasets easier to
@@ -25,9 +26,14 @@ release, exchange and reuse. It organizes and formats data frame R
 objects into well-referenced, well-described, interoperable datasets
 into release and reuse ready form.
 
-<!---
-&#10;The primary aim of dataset is create well-referenced, well-described, interoperable datasets from data.frames, tibbles or data.tables that translate well into the W3C DataSet definition within the [Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) in a reproducible manner. The data cube model in itself is is originated in the _Statistical Data and Metadata eXchange_, and it is almost fully harmonized with the Resource Description Framework (RDF), the standard model for data interchange on the web^[RDF Data Cube Vocabulary, W3C Recommendation 16 January 2014  <https://www.w3.org/TR/vocab-data-cube/>, Introduction to SDMX data modeling <https://www.unescap.org/sites/default/files/Session_4_SDMX_Data_Modeling_%20Intro_UNSD_WS_National_SDG_10-13Sep2019.pdf>].
-&#10;--->
+It enhances R’s data frames (`data.frame`, `tibble`) by allowing you to
+attach machine-readable metadata to datasets and variables using R’s
+native `attributes()`. This includes units, labels, definitions,
+provenance, and persistent identifiers. This help making datasets *FAIR*
+— Findable, Accessible, Interoperable, and Reusable — with embedding
+semantics from the start.
+
+## 🔧 Installation
 
 You can install the latest CRAN release with
 `install.packages("dataset")`.
@@ -40,52 +46,29 @@ The latest development version of dataset can be installed with
 `remotes::install_github()`:
 
 ``` r
-remotes::install_github("dataobservatory-eu/dataset", build = FALSE)
+install.packages("pak")
+pak::pak("dataobservatory-eu/dataset")
 ```
 
-The current version of the `dataset` package is in an early,
-experimental stage. You can follow the discussion of this package on
-[rOpenSci \#553](https://github.com/ropensci/software-review/issues/553)
-about the original scope, that included the datacube data model, and the
[…]
