+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4020
-Date: 2025-04-24
+Version: 0.3.4021
+Date: 2025-04-25
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
 Authors@R: 

[…]
