+++ R/defined.R
@@ -192,11 +192,8 @@ new_datetime_defined <- function(x,
   tmp
 }
 
-#' @rdname defined
-#' @export
-as.character.haven_labelled_defined <- function(x, ...) {
-  NextMethod()
-}
+
+## Subsetting ------------------------------------------------------
 
 #' @export
 `[.haven_labelled_defined` <- function(x, i, ...) {
@@ -250,6 +247,9 @@ tail.haven_labelled_defined <- function(x, n = 6L, ...) {
   x[seq.int(to = length(x), length.out = min(n, length(x)))]
 }
 
+
+## Print & Summary --------------------------------------------------
+
 #' @export
 print.haven_labelled_defined <- function(x, ...) {
   has_def <- !is.null(var_definition(x)) && !is.na(var_definition(x)) && nzchar(var_definition(x))
@@ -288,15 +288,6 @@ format.haven_labelled_defined <- function(x, ...) {
   paste0(base, suffix)
 }
 
-#' @export
-as.list.haven_labelled_defined <- function(x, ...) {
-  lapply(seq_along(x), function(i) x[[i]])
-}
-
-#' @export
-as.vector.haven_labelled_defined <- function(x, mode = "any") {
-  as.vector(vec_data(x), mode = mode)
-}
 #' @rdname defined
 #' @param object An R object to be summarised.
 #' @export
@@ -314,14 +305,124 @@ summary.haven_labelled_defined <- function(object, ...) {
   NextMethod()
 }
 
+
+## Coercion ----------------------------------------------------------
+
+#' @export
+as.list.haven_labelled_defined <- function(x, ...) {
+  lapply(seq_along(x), function(i) x[[i]])
+}
+
+#' @export
+as.vector.haven_labelled_defined <- function(x, mode = "any") {
+  as.vector(vec_data(x), mode = mode)
+}
+
+#' @title Strip the class from a defined vector
+#' @description Converts a `defined` vector to a base R numeric or character,
+#' retaining metadata as passive attributes.
+#' @param x A `defined` vector.
+#' @return A base R vector with attributes (`label`, `unit`, etc.) intact.
+#' @seealso [as_numeric()], [as_character()]
+#' @examples
+#' gdp <- defined(c(3897L, 7365L), label = "GDP", unit = "million dollars")
+#' strip_defined(gpd)
+#'
+#' fruits <- defined(c("apple","avocado", "kiwi"),
+#'                    label = "Fruit", unit = "kg")
+#' strip_defined(fruits)
+#' @export
+strip_defined <- function(x) {
+  if (!inherits(x, "haven_labelled_defined")) return(x)
+
+  base_class <- typeof(vec_data(x))  # typically "double" or "integer"
+  class(x) <- base_class
+  x
+}
+
+
+## Numeric vectors --------------------------------------------------------
+as.numeric <- function(x, ...) {
+  UseMethod("as.numeric")
+}
+
+#' @rdname as_numeric
+#' @description Base R's `as.numeric()` does not support custom classes like
+#'   `defined`. Calling `as.numeric()` on a `defined` vector will drop all
+#'   metadata and class information, which equals to
+#'   `as_numeric(x, preserve_attributes = FALSE)`.
+#' @export
+#' @seealso \code{\link{as_numeric}}
+as.numeric.haven_labelled_defined <- function(x, ...) {
+  unclass(vec_data(x))
+}
+
 #' @title Coerce a defined vector to numeric
 #' @param x A vector created with \code{\link{defined}}.
+#' @description `as_numeric()` is the recommended method to convert a `defined`
+#' vector to numeric. It is metadata-aware and ensures that the underlying data
+#' is numeric before coercion.
+#' @details `as_numeric()` allows `preserve_attributes = TRUE` when the
[…]
