### 2025-04-29 #681 @maurolepore
Dear @antaldaniel

Thanks again for your sharing your work with rOpenSci, for you patience, and for incorporating so many of your suggestions so far.

I'm now the handling editor, please direct all  your questions to me.

This week I'm wrapping up my EiC rotation but starting next week I'll be available to work with you to move this towards approval as smoothly and quickly as you and I can make it.

I plan to first digest this thread, then do some preliminary checks, then assign myself as the first reviewer -- which we sometimes do and the editorial board agrees to it in this specific submission. With that deeper dive into the package I'll be more able to share my findings with the editorial board and discuss the next step.

Would you be available to respond to my suggestions between May 12-23? Not a hard commitment but just to sense your availability and decide if this is indeed the right time for you.
