### 2025-04-29 #681 @antaldaniel
Dear @maurolepore , thank you very much.  I think that I have an important deadline on 15 May, so I think that whatever you write I will be able to review, and at least  partly implement by 23 May, most likely starting only on 16 May.

Because the package is already in CRAN, and I fixed a lot of minor bugs and improvements, if it is not against your preferences, I would release the current version (the README makes it very clear that this is an experimental version under review in rOpenSci). The changes that may be bothering those few users who are experimenting with the package are resolved (there are new, and better tested subsetting, printing, etc. methods for the dataframe like and the vector classes -- nothing groundbreaking and conceptual, but if somebody uses the package as it is, it is a significant improvement.)

So is it OK for you if I just release on CRAN these new improved version and wait for your feedback till 12 May, then reply (and whatever possible, implement) by 23 May?
