### 2025-05-04 #681 @maurolepore
This is to mark the end of my EiC rotation and to hand it over to the next EiC with a short summary.

* This submission has been open for a while.
* Lately we've had rapid activity so to use that momentum I'm taking a dual role as editor and reviewer.
* For a more complete summary see https://github.com/ropensci/software-review/issues/681#issuecomment-2839482728
