### 2025-05-12 #681 @maurolepore
Thanks again for your submission. Here's my review. Before you proceed please wait to hear back from me. I need to first discuss with the editorial board. I'll aim to share next steps this week.

I see there is a lot of love and effort invested in this package. With such a large investment I anticipate some of my feedback will likely feel uncomfortable. I appreciate your willingness to recieve feedback. Please read imagining a friendly tone and the best of the intentions. We're ll here helping each other kindly to make reasearch software better.

Thanks for sharing your work with rOpenSci.

## Semantic comment-tags

I tag my comments (ml01, ml02, and so on) to help track them. Please respond in sequence and using
the corresponding tag:

- [ ] A comment starting with a checkbock requires action and I expect a resonse.
* A comment starting with a bullet is something to consider but does not require action. You may respond but I don't expect it.

**Briefly describe any working relationship you have (had) with the package authors.**

- [x] A statement of need: clearly stating problems the software is designed to solve and its target audience in README
- [x] Installation instructions: for the development version of package and any non-standard dependencies in README

  * ml01: The installation instructions mention {remotes} but use {pak}:

  > The latest development version of dataset can be installed with remotes::install_github():

  ```r
  install.packages("pak")
  pak::pak("dataobservatory-eu/dataset")
  ```

- [ ] Vignette(s): demonstrating major functionality that runs successfully locally

  README (not a vignette but I'll cover it here)

  * ml02: The print output of `orange_df` shows the column names are somewhat
  distant from the values.

  * ml03: I'm struggling to understand the meaning of the information in angle
  brackets `<hvn_lbl_>`. I'm sure it'll become clear later in this or other
  docs, but it would be nice to have a clear "first impression". It would be
  nice to compare how the same dataset prints as a tibble, to more clearly show
  the richness that {dataset} adds.

  * ml04: I see an explicit call to `print()` in `print(get_bibentry())`. It seems
  unnecessary.

  * ml05: The text above `plot(orange_df)`says "Print title and labels" but I see
  no effect reflected in the plot.

  - [ ] ml06: Please review the structure of README. I struggle to follow the
  mix between sections with and without code. Consider grouping all the
  code-sections and all the text-sections. A structure that commonly works is an
  intro at the top with the goal and need, then installation instructions and
  code minimal code examples, and finally the rest of the text-sections. Best
  minimize duplication. Links to vignettes seem unnecessary since the website
  shows the Articles tab prominently. Contributing, code of conduct, related
  work, are common.

  VIGNETTES

  - [ ] ml07: Please reconsider the need for the [Article's introduction page](https://dataset.dataobservatory.eu/articles/index.html).
  The default drop down menu that pkgdown provides is common and enough. More
  text adds maintenance cost and may overwhelm the reader. The vignette titles
  should be enough information for the reader to decide if they want to navigate
  to that article, and then the article's introduction would do the rest --
  likely in the first 1-3 sentences. An alternative to not "throw away" the
  introduction to each article while keeping the website clean is to move it
  under a link "More articles...", e.g.:

  Here I see "The Motivation article will replace the New requirements".
  Consider removing this sentence, removing the article "New Requirements", and
  maybe adding a bullet in NEWS.md saying "New article 'Motivation' replaces the
  old article 'New Requirements'". This example that a changelog item appears in
  an article's introduction suggests the text may need a new pass of review, but
  again it's probably best to remove it since it's easier to maintain long term
  and users will still find the articles.

  * ml08: The article [Motivation](https://dataset.dataobservatory.eu/articles/Motivation.htm)
  duplicates information in README. What's in README seems enought. Consider
  removing the article, or replacing the text in README with a link to the
  article.

  * ml09: The article [Create Datasets that are Easy to Share Exchange and Extend](https://dataset.dataobservatory.eu/articles/dataset_df.html)
  seems equivalent to the common "Get started". If you rename the file dataset.Rmd then pkgdown should
  automatically add a "Get started" link to the navbar.

  * ml10: Consider consolidating deduplicating all other articles. Ideally this
  should result in fewer, more focused articles. Also consider moving the less
  important ones under "More articles..." (see above) to reduce the cognitive
  load required to have a first glance at the website. A smooth experience
  on the website suggests a similarly smooth experience with the code.

- [ ] Function Documentation: for all exported functions

  - [ ] ml11: Please review the [Reference index](https://dataset.dataobservatory.eu/reference/index.html).
  The grouping seems not cohesive. For example, to me it makes sense that the
  "Dataset" group includes functions with "dataset" in their name, e.g.
  `dataset_df()`, `as_dataset_df()`. But the fit in this group for the function
  `get_bibentry()` is unclear to me. Instead a function called `get_bibentry()`
  seems to more naturally fit in the group of "Bibliography functions". Consider
  reviewing the groups and evaluating if you need different groups, or different
  function names that more clearly reflect their membership to a group.

  - [ ] ml12: Please document the `@return` value of `bin_defined_rows()`. See
  these recent [pkgcheck results](https://github.com/ropensci/software-review/issues/681#issuecomment-2869185071).

- [ ] Examples: (that run successfully locally) for all exported functions

  - [ ] ml13: Please fix this failing example (if the point is to show an error you may use
  `try()`):

  ```r
  devtools::run_examples()
  #> ℹ Updating dataset documentation
  #> ℹ Loading dataset
  #> ── Running 36 example files ───────────────────────────────────────── dataset ──
  #> ℹ Loading dataset
  #>
  #> > ### Name: as.character.haven_labelled_defined
  #> > ### Title: Coerce to character vector
  #> > ### Aliases: as.character.haven_labelled_defined as_character
  #> > ###  as_character.haven_labelled_defined
  #>
  #> > ###  Examples
  #> >
  #> > x <- defined(c("a", "b", "c"), label = "Letter code")
  #>
  #> > as_character(x)
  #> [1] "a" "b" "c"
  #>
  #> > y <- defined(1:3, label = "Index")
  #>
  #> > as_character(y)
  #> Error in as_character.haven_labelled_defined(y): as_character(): underlying data is not a character vector.
  #> ℹ Loading dataset
  ```

  It's likely what's causing R CMD check to fail on [this recent run of pkgcheck](https://github.com/ropensci/software-review/issues/681#issuecomment-2869185071).

  - [ ] ml14: Please review all examples and ensure the website shows the most relevant output.
  The goal is for a potential user to understand each hepfile before they
  install the package and run the code. The reprex package gives [good advice](https://reprex.tidyverse.org/articles/reprex-dos-and-donts.html#package-philosophy)

  > The ... code ... Should be easy for others to digest, so they don't
  > necessarily have to run it.

  Here are a few examples so you get the idea:

  * [`?orange_df`](https://dataset.dataobservatory.eu/reference/orange_df.html)
  lacks an examples section showing the dataset itself. This would be useful:

  ```r
  library(dataset)

  orange_df
  #> Draper N, Smith H (1998). "Growth of Orange Trees."
  #>   rowid  tree   age circumference
  #>                        The age of circumference
  #>  <hvn_lbl_> <hvn_lbl_> <hvn_lbl_> <hvn_lbl_>
  #> 1 orange:1   2 [1]     118            30
  #> 2 orange:2   2 [1]     484            58
  #> 3 orange:3   2 [1]     664            87
  #> 4 orange:4   2 [1]    1004           115
  #> 5 orange:5   2 [1]    1231           120
  #> 6 orange:6   2 [1]    1372           142
  #> 7 orange:7   2 [1]    1582           145
  #> 8 orange:8   4 [2]     118            33
  #> 9 orange:9   4 [2]     484            69
  #> 10 orange:10  4 [2]     664           111
  #> # ℹ 25 more rows
  ```

  * [`?dataset_df()`](https://dataset.dataobservatory.eu/reference/dataset_df.html)
  shows how to add metadata but not how to see it:

  ```r
  library(dataset)

  my_dataset <- dataset_df(
    country_name = defined(
      c("AD", "LI"),
      definition = "[http://data.europa.eu/bna/c_6c2bb82d](http://data.europa.eu/bna/c_6c2bb82d)",
      namespace = "[https://www.geonames.org/countries/$1/](https://www.geonames.org/countries/$1/)"
    ),
    gdp = defined(
      c(3897, 7365),
      label = "Gross Domestic Product",
      unit = "million dollars",
      definition = "[http://data.europa.eu/83i/aa/GDP](http://data.europa.eu/83i/aa/GDP)"
    )
  )

  print(my_dataset)
  #> Unknown A (:tba). "Untitled Dataset."
  #>   rowid  country_name gdp
  #>                        Gross Dome
  #>  <hvn_lbl_> <hvn_lbl_>   <hvn_lbl_>
  #> 1 eg:1       AD         3897
  #> 2 eg:2       LI         7365

  is.dataset_df(my_dataset)
  #> [1] TRUE
  ```

  This would be useful:

  ```
  str(my_dataset$country_name)
  #>  hvn_lbl_ [1:2] AD, LI
  #>  @ definition: chr "[http://data.europa.eu/bna/c_6c2bb82d](http://data.europa.eu/bna/c_6c2bb82d)"
  #>  @ namespace : chr "[https://www.geonames.org/countries/$1/](https://www.geonames.org/countries/$1/)"
  str(my_dataset$gdp)
  #>  hvn_lbl_ [1:2] 3897 (million dollars), 7365 (million dollars)
  #>  @ label   : chr "Gross Domestic Product"
  #>  @ unit    : chr "million dollars"
  #>  @ definition: chr "[http://data.europa.eu/83i/aa/GDP](http://data.europa.eu/83i/aa/GDP)"
  ```

  * [`?get_bibentry()`](https://dataset.dataobservatory.eu/reference/get_bibentry.html)
  doesn't show its output, neither before nor after calling `set_bibentry()` so
  we can't see what's the effect.

- [x] Community guidelines: including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

- [x] Installation: Installation succeeds as documented.

  * ml15: Great. On a fresh Ubuntu system with R 4.5.0, the package installs
  smoothly from CRAN and GitHub (using `pak()`).

- [x] Functionality: Any functional claims of the software have been confirmed.

  * ml16: The functions I tried worked fine and the tests cover a lot of ground. But
  much of my time was spent in other aspects of this review so I can't claim
  I did this sufficiently.

- [x] Performance: Any performance claims of the software have been confirmed.

  * ml17: I see no performance claims in the computational sense.

- [ ] Automated tests: Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

  * ml18: In general test coverage looks great but some specific files with lots
  of logic may need attention

  ```
  #> R/var_label.R: 35.71%
  #> R/xsd_convert.R: 46.15%
  #> R/agent.R: 47.62%
  #> R/id_to_column.R: 51.72%
  #> R/get_bibentry.R: 71.43%
  #> R/as_dublincore.R: 73.61%
  #> R/publisher.R: 75.00%
  #> R/dataset_df.R: 75.54%
  #> R/utils-idcol_find.R: 76.92%
  #> R/identifier.R: 78.05%
  #> R/dataset_title.R: 78.95%
  ```

  This file has a lot of logic and seems straightforward to test:

  * ml19: Consider making the test titles more specific, e.g

  ```r
  library(testthat)
  devtools::load_all()
  #> ℹ Loading dataset

  # Better
  # More specific. The data is minimal
  test_that("creates objects of S3 class 'dataset_df'", {
    expect_s3_class(as_dataset_df(data.frame()), "dataset_df")

    expect_true(inherits(as_dataset_df(data.frame()), "dataset_df"))
    expect_false(inherits(data.frame(), "dataset_df"))
  })
  #> Test passed 🎊

  # Can improve
  # Not specific. The specific data is irrelevant to this test
  # [https://github.com/dataobservatory-eu/dataset/blob/8768983c36a6f6ee2aadf7637dcf5f0f88ab1e3e/tests/testthat/test-dataset_df.R#L144](https://github.com/dataobservatory-eu/dataset/blob/8768983c36a6f6ee2aadf7637dcf5f0f88ab1e3e/tests/testthat/test-dataset_df.R#L144)
  test_that("as_dataset_df() works", {
    expect_s3_class(as_dataset_df(iris), "dataset_df")
    expect_false(is.dataset_df(mtcars))
  })
  #> Test passed 🥳
  ```

  This makes it easier to understand what's the problem when a test fails. Compare:

  ```
  # Too vague
  ── Failure (test-dataset_df.R:146:3): as_dataset_df() works ───────────────────────────────────────────────────────
  is.dataset_df(mtcars) is not TRUE

  `actual`:    FALSE
  `expected`: TRUE

  # Shows intent
  ── Failure (test-dataset_df.R:151:3): as_dataset_df() creates objects of S3 class 'dataset_df' ────────────────────
  is.dataset_df(mtcars) is not TRUE

  `actual`:    FALSE
  `expected`: TRUE
  ```

  - [ ] ml20: In calls to `expect_error()` please enure to test not just for any error but for the specific one
  you want to test. See the argument `regex` or `class` of `expect_error()` or
  consider using expect_snapshot(error = TRUE)` (or friends) ([example](https://github.com/dataobservatory-eu/dataset/blob/main/tests/testthat/test-defined.R#L240C1-L265C3)).

  ```r
  test_that("new_labelled_defined() throws errors", {
    expect_error(defined(
      x = c(1:3),
      label = c("Numbers", "Numbers"),
      definition = 1,
      namespace = "Iris"
    ))
    expect_error(defined(
      x = c(1:3),
      label = "Numbers",
      definition = 1,
      namespace = "Iris"
    ))
    # ... and more
  })
  ```

  * ml21: Consider removing the name of the function under test from the test
  title *when the test file contains the name of that function* (which should
  be most cases if you follow the most common naming conventions) ([example](https://github.com/dataobservatory-eu/dataset/blob/main/tests/testthat/test-xsd_convert.R)).

  ```
  # The file name already contains the name of the function test-xsd_convert
  ── Failure (test-xsd_convert.R:2:3): works ────────────────────────────────────────────────────────────────────────
  xsd_convert(TRUE) (`actual`) not equal to "fail for demo" (`expected`).

  `actual`:    ""TRUE"^^<xs:boolean>""
  `expected`: "fail for demo"

  # The function name is duplicated information
  ── Failure (test-xsd_convert.R:6:3): xsd_convert(x, idcols) works ─────────────────────────────────────────────────
  xsd_convert(TRUE) (`actual`) not equal to "fail for demo" (`expected`).

  `actual`:    ""TRUE"^^<xs:boolean>""
  `expected`: "fail for demo"
  ```

- [ ] Packaging guidelines: The package conforms to the rOpenSci packaging guidelines.

  * ml22: rOpenSci's guide recommends [markdown syntax in roxygen2 documentation](https://roxygen2.r-lib.org/articles/rd-formatting.html). I guess
  LLMs

  This seems like the kind of task good for an LLM. For example, starting
  from [this](https://github.com/dataobservatory-eu/dataset/blob/main/R/dataset_df.R),
  Google's Gemini did a pretty good job:

  ```
  # https://github.com/dataobservatory-eu/dataset/blob/main/R/dataset_df.R

  #' Create a new dataset_df object
  #'
  #' The `dataset_df` constructor creates objects of this class, which
  #' are semantically rich, modern data frames inherited from
  #' `\link[tibble:tibble]{tibble::tibble}`.
  #'
  #' @details
  #' To check if an object has the class `dataset_df` use `is.dataset_df`.
  #'
  #' `print` is the method to print out the semantically rich data frames
  #' created with the constructor of `dataset_df`.
  #'
  #' `summary` is the method to summarise these semantically rich data frames.
  #'
  #' For more details, please check the `vignette("dataset_df", package = "dataset")`
  #' vignette.
  #'
  #' @param identifier Defaults to `c(eg="http://example.com/dataset#")`, which should be
  #' changed to the permanent identifier of the dataset. For example, if your dataset will be
  #' released with the Digital Object Identifier (DOI) `https;//doi.org/1234`, you should use
  #' a short prefixed identifier like `c(obs="https://doi.org/1234#")`, which will resolve
  #' to the rows being identified as https://doi.org/1234#1...https://doi.org/1234#n.
  #' @param dataset_bibentry A list of bibliographic references and descriptive metadata
  #' about the dataset as a whole created with [`datacite()`] or
  #' [`dublincore()`].
  #' @param var_labels The long, human readable labels of each variable.
  #' @param units The units of measurement for the measured variables.
  #' @param definitions The linked definitions of the variables, attributes, or constants.
  #' @param dataset_subject The subject of the dataset, see [`subject()`].
  #' @param ... The vectors (variables) that should be included in the dataset.
  #' @param x A `dataset_df` object for S3 methods.
  #' @param df A `data.frame` to be converted to `dataset_df`.
  #' @return `dataset_df` is the constructor of this type, it returns an object
  #' inherited from a data frame with semantically rich metadata.
  #' @import vctrs
  #' @import pillar
  #' @examples
  #' my_dataset <- dataset_df(
  #'  country_name = defined(
  #'    c("AD", "LI"),
  #'    definition = "http://data.europa.eu/bna/c_6c2bb82d",
  #'    namespace = "https://www.geonames.org/countries/$1/"
  #'  ),
  #'  gdp = defined(
  #'    c(3897, 7365),
  #'    label = "Gross Domestic Product",
  #'    unit = "million dollars",
  #'    definition = "http://data.europa.eu/83i/aa/GDP"
  #'  )
  #' )
  #'
  #' print(my_dataset)
  #'
  #' is.dataset_df(my_dataset)
  #' @export
  ```

  - [ ] ml23: Please re-run the styler:

    ```r
    styler::style_pkg()
    #> Styling  79  files:
    ...
    #>  R/datacite.R                                    ✔
    #>  R/dataset_df.R                                  ℹ
    #>  R/dataset_title.R                               ✔
    #>  R/dataset_to_triples.R                          ✔
    #>  R/dataset-iris_dataset.R                        ✔
    #>  R/dataset-orange_df.R                           ✔
    #>  R/defined.R                                     ℹ
    #>  R/describe.R                                    ✔
    #>  R/description.R                                 ✔
    #>  R/dublincore.R                                  ℹ
    #>  R/fix_contributor.R                             ℹ
    #>  R/fix_publisher.R                               ✔
    #>  R/geolocation.R                                 ✔
    ...
    #>  tests/testthat/test-dataset_title.R            ✔
    #>  tests/testthat/test-dataset_to_triples.R       ✔
    #>  tests/testthat/test-defined.R                  ℹ
    #>  tests/testthat/test-describe.R                 ✔
    #>  tests/testthat/test-description.R              ✔
    #>  tests/testthat/test-dublincore.R               ℹ
    #>  tests/testthat/test-fix_contributor.R          ℹ
    #>  tests/testthat/test-fix_publisher.R            ℹ
    #>  tests/testthat/test-geolocation.R              ✔
    ...
    #>  vignettes/dataset_df.Rmd                        ✔
    #>  vignettes/defined.Rmd
    #> Warning: When processing vignettes/defined.Rmd: ℹ In index: 7.
    #> Caused by error in `parse_safely()`:
    #> ! <text>:2:7: unexpected 'in'
    #> 1: c(gdp_1, gdp_2)
    #> 2: Error in
    #>        ^
    #> ⚠
    #>  vignettes/Motivation.Rmd                       ✔
    ...
    #> ────────────────────────────────────────
    #> Status    Count   Legend
    #> ✔      70   File unchanged.
    #> ℹ       8   File changed.
    #> ✖       1   Styling threw an error.
    #> ────────────────────────────────────────
    #> Please review the changes carefully!
    ```

    \</details\>

  * ml24: No need to acknowledge in DESCRIPTION. Your kind, thoughtful responses will be
[…]
