### 2025-05-12 #681 @antaldaniel
@maurolepore Thank you very much for the extensive comments, and I will start from here, because I was a bit concerned that you perhaps see a wrong version as all examples ran smoothly for me, as well as checks, and

>This is surprising to me because this https://github.com/ropensci/software-review/issues/681#issuecomment-2828987353 suggests it was removed:
> So, upon your request, I removed the iris dataset from all vignettes and examples. It uses now the Orange dataset.
> Did you remove it from the default main branch (which is the one I'm reviewing)?

... indeed, you should not meet the iris dataset.  The agreement was that dataset will be hidden (after all, it is part of base R, and about 200 unit test use it) but from the examples and vignettes it will be removed. And as far as I see that the iris dataset only appears in the documentation once and in unit tests.

I'll write out all your comments as issues and take a serious look, because when I signalled that I think all issues are resolved (at least from the previous two reviews), I certainly did not see the `ìris` dataset, and any failing examples or tests.  As originally suggested, I will work on this during the week, and try to implement all your suggestions.

The original guidelines for rOpenSci when I first submitted this package for review was to submit preferrably in a conceptual stage, before the first CRAN release, and that was my intention.  So I understand that there are a lot of unpolished solutions present, but the idea was to somehow create a tool that can receive semantically rich datasets into the R ecosystem, and can give it back to such system, mainly those APIs that use SDMX, i.e., the statistical data and metadata exchange standard.
