### 2025-05-17 e396a05
Cover edge cases, and the difftime class
