+++ R/xsd_convert.R
@@ -7,7 +7,12 @@
 #'   (observation) identifiers. If \code{NULL}, it will make a new \code{idcol}
 #'   from [row.names()].
 #' @param ... Further optional parameters for generic method.
-#' @return A serialisation of an R vector or data frame (dataset) in XML.
+#' @return
+#' A character vector of RDF-compatible typed literals. Each element
+#' corresponds to an input value, serialized according to its type
+#' (e.g., `xs:string`, `xs:integer`, `xs:dateTime`).
+#' For data frames or tibbles, each row is converted into a set of
+#' RDF triples, with columns mapped to predicates.
 #' @export
 
 xsd_convert <- function(x, idcol, ...) {
@@ -30,100 +35,84 @@ get_type <- function(t) {
     type <- "xs:date"
   } else if (any(class(t) == "POSIXct")) {
     type <- "xs:dateTime"
+  } else if (any(class(t) == "difftime")) {
+    type <- "xs:duration"
   }
-
   type
 }
 
 #' @rdname xsd_convert
 #' @examples
-#'
 #' # Convert data.frame to XML Schema Definition
 #' xsd_convert(data.frame(a = 1:3, b = c("a", "b", "c")))
 #' @exportS3Method
-#' @export
 xsd_convert.data.frame <- function(x, idcol = NULL, ...) {
-  get_type <- function(t) {
-    if (any(class(t) %in% c("numeric", "double"))) {
-      type <- "xs:decimal"
-    } else if (any(class(t) == "integer")) {
-      type <- "xs:integer"
-    } else if (any(class(t) %in% c("character", "factor"))) {
-      type <- "xs:string"
-    } else if (any(class(t) == "logical")) {
-      type <- "xs:boolean"
-    } else if (any(class(t) == "numeric")) {
-      type <- "xs:decimal"
-    } else if (any(class(t) == "Date")) {
-      type <- "xs:date"
-    } else if (any(class(t) == "POSIXct")) {
-      type <- "xs:dateTime"
-    }
-
-    type
-  }
-
-  convert_cols <- seq_along(x)
-
+  # Identify ID column (or default to row names)
   if (!is.null(idcol)) {
-    ## See utils-idcol_find.R for the internal function
-    convert_cols <- convert_cols[-idcol_find(x = x, idcol = idcol)]
-  }
-
-  convert_column <- function(c) {
-    var_type <- get_type(t = x[[c]])
-    if (!var_type %in% c("codelist", "literal")) {
-      paste0('\"', as.character(x[[c]]), '\"', "^^<", var_type, ">")
-    } else {
-      as.character(x[[c]])
-    }
-  }
-
-  xsd_list <- lapply(convert_cols, function(c) convert_column(c))
-  xsd_dataframe <- as.data.frame(xsd_list)
-
-  idcol <- which(!seq_along(x) %in% convert_cols)
-  if (length(idcol) == 1) {
-    xsd_dataframe <- cbind(x[, idcol], xsd_dataframe)
-    names(xsd_dataframe) <- names(x)
+    id_idx <- idcol_find(x, idcol)
+    ids <- as.character(x[[id_idx]])
+    id_name <- names(x)[id_idx]
   } else {
-    names(xsd_dataframe) <- names(x)
+    ids <- row.names(x)
+    id_name <- ".rowid"
   }
 
-  tmp_df <- x
-  for (j in seq_along(tmp_df)) {
-    tmp_df[, j] <- as.character(tmp_df[, j])
-    tmp_df[, j] <- xsd_dataframe[, j]
-  }
[…]
