+++ tests/testthat/test-xsd_convert.R
@@ -1,6 +1,270 @@
-test_that("xsd_convert(x, idcols) works", {
-  expect_equal(xsd_convert(TRUE), "\"TRUE\"^^<xs:boolean>")
-  expect_equal(xsd_convert(defined(x = 2:4, label = "test")), c('\"2\"^^<xs:integer>', '\"3\"^^<xs:integer>', '\"4\"^^<xs:integer>'))
-  expect_equal(xsd_convert(head(iris)[1, 1]), '\"5.1\"^^<xs:decimal>')
-  expect_equal(unlist(xsd_convert(x = head(iris_dataset), idcol = 1)[1, 2]), c(Sepal.Length = '\"5.1\"^^<xs:decimal>'))
+
+## get_type internal function -------------------------------------
+test_that("get_type returns correct XSD types", {
+  expect_equal(get_type(1.5), "xs:decimal")
+  expect_equal(get_type(1L), "xs:integer")
+  expect_equal(get_type("text"), "xs:string")
+  expect_equal(get_type(TRUE), "xs:boolean")
+  expect_equal(get_type(Sys.Date()), "xs:date")
+  expect_equal(get_type(Sys.time()), "xs:dateTime")
+  expect_equal(get_type(as.difftime(c("0:3:20", "11:23:15"))), "xs:duration")
+})
+
+## Character conversions ------------------------------------------
+test_that("xsd_convert.character handles standard strings", {
+  expect_equal(
+    xsd_convert(c("apple", " banana ", "cherry")),
+    c('"apple"^^<xs:string>', '" banana "^^<xs:string>', '"cherry"^^<xs:string>')
+  )
+})
+
+test_that("xsd_convert.character handles NA values", {
+  result <- xsd_convert(c("apple", NA, "cherry"))
+  expect_equal(result[1], '"apple"^^<xs:string>')
+  expect_true(is.na(result[2]))
+  expect_equal(result[3], '"cherry"^^<xs:string>')
+})
+
+test_that("xsd_convert.character handles empty strings", {
+  expect_equal(xsd_convert(""), '""^^<xs:string>')
+})
+
+test_that("xsd_convert.character handles empty vector", {
+  expect_equal(xsd_convert(character(0)), '""^^<xs:string>')
+})
+
+## Numeric conversions ----------------------------------------------------
+
+test_that("xsd_convert.numeric handles standard numeric values", {
+  expect_equal(
+    xsd_convert(c(1.0, 2.5, 3.0)),
+    c('"1"^^<xs:decimal>', '"2.5"^^<xs:decimal>', '"3"^^<xs:decimal>')
+  )
+})
+
+test_that("xsd_convert.numeric handles NA values", {
+  result <- xsd_convert(c(1.0, NA, 2.0))
+  expect_equal(result[1], '"1"^^<xs:decimal>')
+  expect_true(is.na(result[2]))
+  expect_equal(result[3], '"2"^^<xs:decimal>')
+})
+
+test_that("xsd_convert.numeric handles empty vector", {
+  expect_equal(xsd_convert(numeric(0)), '""^^<xs:decimal>')
+})
+
+## Integer conversions ----------------------------------------------------
+test_that("xsd_convert.integer handles normal integers", {
+  expect_equal(xsd_convert(1:3), c('"1"^^<xs:integer>', '"2"^^<xs:integer>', '"3"^^<xs:integer>'))
+})
+
+test_that("xsd_convert.integer handles NA", {
+  x <- as.integer(c(42, NA))
+  result <- xsd_convert(x)
+  expect_equal(result[1], '"42"^^<xs:integer>')
+  expect_true(is.na(result[2]))
+})
+
[…]

[…]
