+++ R/id_to_column.R
@@ -8,10 +8,10 @@
 #' @examples
 #'
 #' # Example with a dataset_df object:
-#' id_to_column(iris_dataset)
+#' id_to_column(orange_df)
 #'
-#' # Example with a data.frame object:#'
-#' id_to_column(iris, prefix = "eg:iris-o")
+#' # Example with a data.frame object:
+#' id_to_column(Orange, prefix = "orange:")
 #' @export
 id_to_column <- function(x, prefix = "eg:", ids = NULL) {
   is_dataset <- is.dataset_df(x)
