+++ man/strip_defined.Rd
@@ -18,7 +18,7 @@ retaining metadata as passive attributes.
 }
 \examples{
 gdp <- defined(c(3897L, 7365L), label = "GDP", unit = "million dollars")
-strip_defined(gpd)
+strip_defined(gdp)
 
 fruits <- defined(c("apple","avocado", "kiwi"),
                    label = "Fruit", unit = "kg")
