### 2025-05-17 68d03e3
more consistent handling of as_character and as.character
