+++ man/as_character.Rd
@@ -1,11 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/defined.R
-\name{as.character.haven_labelled_defined}
+\name{as.character}
+\alias{as.character}
 \alias{as.character.haven_labelled_defined}
 \alias{as_character}
 \alias{as_character.haven_labelled_defined}
 \title{Coerce to character vector}
 \usage{
+as.character(x, ...)
+
 \method{as.character}{haven_labelled_defined}(x)
 
 as_character(x, ...)
@@ -46,11 +49,9 @@ prevent accidental coercion of non-numeric data. \cr\cr
 preferred method.
 }
 \examples{
-x <- defined(c("a", "b", "c"), label = "Letter code")
-as_character(x)
-
-y <- defined(1:3, label = "Index")
-as_character(y)
+as.character(defined(c("a", "b", "c"), label = "Letter code"))
+as_character(defined(c("a", "b", "c"), label = "Letter code"))
+fruits <- defined(c("apple","avocado", "kiwi"), label = "Fruit", unit = "kg")
 # Keep the metadata, but revert to base R character type:
 as_character(fruits, preserve_attributes = TRUE)
 
