### 2025-05-17 ec1ca7d
improve function documentation
