+++ man/bind_defined_rows.Rd
@@ -7,29 +7,30 @@
 bind_defined_rows(x, y, ..., strict = FALSE)
 }
 \arguments{
-\item{strict}{Logical. If \code{TRUE} (default), all variables must match
-exactly in both datasets, including variable labels, units, definitions,
-namespaces, and the dataset-level row identifier namespace. If
-\code{FALSE}, only user-defined variable semantics are checked strictly,
-while the row identifier column is inherited from \code{x} without
-requiring an exact match.}
+\item{x}{A \code{dataset_df} object.}
+
+\item{y}{A \code{dataset_df} object to bind to \code{x}.}
+
+\item{...}{Optional dataset-level attributes such as \code{title} or \code{creator} to override.}
+
+\item{strict}{Logical. If \code{TRUE} (default), require full semantic compatibility, including rowid.}
+}
+\value{
+A new \code{dataset_df} object with rows from \code{x} and \code{y}, combined semantically.
 }
 \description{
-Add rows of dataset y to dataset x. Metadata (labels, units,
-definitions, namespaces) must match exactly. The combined dataset inherits
-x's citation by default, and creator roles are merged.
+Add rows of dataset \code{y} to dataset \code{x}, validating all semantic metadata.
+Metadata (labels, units, definitions, namespaces) must match exactly.
+Additional dataset-level metadata such as title and creator can be overridden using \code{...}.
 }
 \details{
 This function combines two semantically enriched datasets created with \code{dataset_df()}.
-It ensures that the structure and metadata of the datasets are compatible.
+All variable-level attributes — including labels, units, definitions, and namespaces —
+must match. If \code{strict = TRUE} (the default), the row identifier namespace
+(used in the \code{rowid} column) must also match exactly.
 
-All variable-level attributes — including variable labels, measurement units,
-linked definitions, and namespaces — must match exactly in both datasets.
-
-If \code{strict = TRUE} (the default), the row identifier namespace
-(used in the \code{rowid} column) must also be identical in both datasets.
-If \code{strict = FALSE}, this requirement is relaxed: the resulting dataset
-inherits the row ID structure from \code{x}, even if \code{y} used a different one.
+If \code{strict = FALSE}, row identifiers from \code{y} may differ and will be ignored;
+the output will inherit \code{x}'s row identifier scheme.
 }
 \examples{
 A <- dataset_df(
@@ -43,19 +44,16 @@ B <- dataset_df(
   identifier = c(id = "http://example.org/dataset#")
 )
 
-# This works: same structure and rowid namespace
-bind_defined_rows(A, B)
+bind_defined_rows(A, B)  # succeeds
 
 C <- dataset_df(
   length = defined(c(30, 35), label = "Length", unit = "cm", namespace = "http://example.org"),
   identifier = c(id = "http://another.org/dataset#")
 )
 
-# This throws an error because the rowid namespace differs
 \dontrun{
-bind_defined_rows(A, C, strict = TRUE)
+bind_defined_rows(A, C, strict = TRUE)  # fails: mismatched rowid
 }
 
-# This succeeds because rowid is inherited from A
-bind_defined_rows(A, C, strict = FALSE)
+bind_defined_rows(A, C, strict = FALSE)  # succeeds: rowid inherited
 }
