### 2025-05-17 fbbff7e
documentation improvements
