+++ R/fix_contributor.R
@@ -14,10 +14,12 @@
 #' @return A character string like `{Jane Doe [dtm, ctb]} and {John Smith
 #'   [ctb]}`.
 #' @examples
+#' \dontrun{
 #' fix_contributor(person("Jane", "Doe", role = c("dtm", "ctb")))
 #' fix_contributor(c(person("Jane", "Doe", role = "ctb"),
 #'                   person("John", "Smith")))
 #' fix_contributor(":unas")
+#' }
 #' @keywords internal
 fix_contributor <- function(contributors = NULL) {
 
