+++ NAMESPACE
@@ -37,13 +37,14 @@ S3method(xsd_convert,Date)
 S3method(xsd_convert,POSIXct)
 S3method(xsd_convert,character)
 S3method(xsd_convert,data.frame)
-S3method(xsd_convert,dataset)
+S3method(xsd_convert,dataset_df)
+S3method(xsd_convert,difftime)
 S3method(xsd_convert,factor)
 S3method(xsd_convert,haven_labelled_defined)
 S3method(xsd_convert,integer)
 S3method(xsd_convert,logical)
 S3method(xsd_convert,numeric)
-S3method(xsd_convert,tibble)
+S3method(xsd_convert,tbl_df)
 export("creator<-")
 export("dataset_title<-")
 export("definition_attribute<-")
@@ -130,6 +131,8 @@ importFrom(stats,setNames)
 importFrom(tibble,new_tibble)
 importFrom(utils,bibentry)
 importFrom(utils,citation)
+importFrom(utils,head)
 importFrom(utils,person)
+importFrom(utils,tail)
 importFrom(vctrs,df_list)
 importFrom(vctrs,vec_data)
