+++ tests/testthat/test-dataset_df.R
@@ -5,7 +5,7 @@ test_that("dataset_df() works", {
   expect_equal(get_bibentry(dataset_df(mtcars))$author, person("Author", "Unknown"))
   expect_equal(
     dataset_df(mtcars, identifier = c(mt = "http:/mtcars.com/dataset#"))$rowid,
-    defined(paste0("mt:", 1:nrow(mtcars)),
+    defined(paste0("mt:", seq_along(mtcars$mpg)),
       namespace = c(mt = "http:/mtcars.com/dataset#")
     )
   )
