+++ man/as_factor.Rd
@@ -8,6 +8,8 @@ as_factor(x)
 }
 \arguments{
 \item{x}{A vector created with \code{\link{defined}}.}
+
+\item{...}{Further arguments passed to internal methods (not used).}
 }
 \value{
 A factor vector.
