+++ man/as_numeric.Rd
@@ -15,6 +15,8 @@ as_numeric(x, ...)
 \arguments{
 \item{x}{A vector created with \code{\link{defined}}.}
 
+\item{...}{Further arguments passed to internal methods (not used).}
+
 \item{preserve_attributes}{Defaults to \code{FALSE}, in which case the
 \code{unit}, \code{definition} and \code{namespace} attributes will be
 preserved, but the returned value will otherwise become a base R integer,
