### 2025-05-17 d89c8f3
Add warning if dplyr and tidyr is not installed
