### 2025-05-17 ef6a198
Add warning to package dependency
