### 2025-05-17 3fd13e1
Fix a bug with non-existing contributors throwing an error.
