+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4023
-Date: 2025-05-17
+Version: 0.3.4024
+Date: 2025-05-18
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
 Authors@R: 

[…]
