### 2025-05-22 f9af045
improved bibliography handling
