+++ man/get_variable_concepts.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/var_concept.R
+\name{get_variable_concepts}
+\alias{get_variable_concepts}
+\title{Get concepts for all variables in a dataset_df}
+\usage{
+get_variable_concepts(x)
+}
+\arguments{
+\item{x}{A \code{dataset_df} object.}
+}
+\value{
+A named list of concept URIs for each variable.
+}
+\description{
+Returns a named list of concept URIs (or NULLs) for all
+variables.
+}

[…]
