### 2025-05-22 7787615
introducing the bibrecord class
