+++ vignettes/bibentry.Rmd
@@ -1,318 +1,0 @@
----
-title: "Bibentry for FAIR datasets"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Bibentry for FAIR datasets}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-```{r setup}
-library(dataset)
-library(datasets)
-
-orange_df <- dataset_df(
-  datasets::Orange,
-  identifier = c(orange = "dataset::Orange")
-)
-
-orange_df
-```
-
-*Findability*, *Accessibility*, *Interoperability*, and true *Reuse* in the long run are elusive targets. The dataset package aims to find common ground for new variations of tidy datasets that are truly interoperable and can be reused in a greater circle than usual datasets. These aims are covered in other vignette articles. Now, we focus on findability and specific aspects of accessibility.
-
-The bibentry class and its methods in the utils package offer a well-known and widely utilised object for carrying descriptive metadata and licensing information. Unfortunately, neither the bib entry class nor its rOpenSci modernisation, the [RefManageR](https://docs.ropensci.org/RefManageR/), follow the two most widely used metadata standards for making datasets public in open science or open statistical repositories. After experimenting with both bib entry and `RefManageR`, we decided to use the far more widely used bibentry class to be extended to meet the [DCTERMS](https://www.dublincore.org/specifications/dublin-core/dcmi-terms/) (Dublin Core) and [DataCite](https://datacite-metadata-schema.readthedocs.io/en/4.6/) metadata schema standards used by most repositories. However, in the end of this vignette article we show a combination of our new S3 classes with `RefManageR`.
-
-```{r emptybibentry}
-get_bibentry(orange_df)
-```
-
-
-The two standards have a relatively high level of interoperability but do not have compatible elements. To avoid confusion, we created a `dublincore()` inherited object with a function interface that uses small capitalised snake case parameters and the `datacite()` version with capitalised parameters, similar to the original spelling. These classes and functions add the mandatory metadata fields of the standards to the `bibentry` class (the way it retains compatibility with the class) and allow the addition of the `bibentry` of the dataset to the attributes of a data frame.
-
-<!---
-
-In 1935, American botanist *Edgar Anderson* measured petal and sepal structural dimensions (length and width) for 50 flowers from three Iris species: *Iris setosa*, *Iris versicolor*, and *Iris virginica*. The manageable but non-trivial size (5 variables and 150 total observations) and characteristics of Anderson’s Iris dataset, including linear relationships and multivariate normality, and the fact that it is a balanced dataset that can be subset to three identical sized slices without missing variables made it amenable for introducing a wide range of statistical methods, including data wrangling, visualisation, linear modelling, multivariate analyses, and machine learning. The Iris dataset has been used in statistical education for about 90 years, and it is built into many software packages, including the auto-installed dataset package in R, Python’s scikit-learn machine learning library, and the SAS Sashelp library.
-
-There are some other criticism towards this dataset, that we will discuss in the end of this article, at this point we only focus on the fact that the way it is added to R core is lacking in metadata, making it a great candidate to for upgrading from data.frame to dataset_df.
---->
-
[…]
