### 2025-05-22 c70f1f5
More consistent handling of bibliographic entries
