+++ tests/testthat/test-datacite.R
@@ -22,7 +22,125 @@ test_that("as_datacite() works", {
   expect_equal(as_datacite(iris_dataset, "list")$Description, "The famous (Fisher's or Anderson's) iris data set.")
 })
 
+test_that("datacite() constructor works with structured person objects", {
+  dc_obj <- datacite(
+    Title = "Growth of Orange Trees",
+    Creator = c(
+      person(
+        given = "N.R.",
+        family = "Draper",
+        role = "cre",
+        comment = c(VIAF = "http://viaf.org/viaf/84585260")
+      ),
+      person(
+        given = "H",
+        family = "Smith",
+        role = "cre"
+      )
+    ),
+    Contributor = person(
+      given = "Daniel",
+      family = "Antal",
+      role = "dtm"
+    ),
+    Publisher = "Wiley",
+    Identifier = "https://doi.org/10.5281/zenodo.14917851",
+    PublicationYear = 1998,
+    Language = "en",
+    Format = "application/r-rds",
+    Rights = "CC-BY",
+    Description = "Tree growth dataset over years."
+  )
+
+  expect_s3_class(dc_obj, "datacite")
+  expect_s3_class(dc_obj, "bibentry")
+
+  expect_equal(dc_obj$title, "Growth of Orange Trees")
+
+  expect_equal(dc_obj$author, c(
+    person(given = "N.R.", family = "Draper", role = "cre", comment = c(VIAF = "http://viaf.org/viaf/84585260")),
+    person(given = "H", family = "Smith", role = "cre")
+  ))
+
+  expect_equal(dc_obj$publisher, "Wiley")
+  expect_equal(dc_obj$identifier, "https://doi.org/10.5281/zenodo.14917851")
+  expect_equal(dc_obj$year, "1998")
+  expect_equal(dc_obj$language, "en")
+  expect_equal(dc_obj$format, "application/r-rds")
+  expect_equal(dc_obj$rights, "CC-BY")
+  expect_equal(dc_obj$description, "Tree growth dataset over years.")
+
+  # Contributor should be stored as an attribute (structured)
+  contributor_attr <- attr(dc_obj, "contributor")
+  expect_s3_class(contributor_attr, "person")
+  expect_equal(contributor_attr, person(given = "Daniel", family = "Antal", role = "dtm"))
+})
+
+test_that("print.datacite outputs key fields", {
+  dc <- datacite(
+    Title = "Orange Tree Dataset",
+    Creator = person(given = "Jane", family = "Doe"),
+    Contributor = person(given = "John", family = "Smith", role = "ctb"),
+    Publisher = "FruitLab",
+    Identifier = "https://doi.org/10.1234/oranges",
+    Date = 2023,
+    Language = "en",
+    Description = "A dataset measuring the growth of orange trees."
+  )
+
+  out <- capture.output(print(dc))
+  expect_true(any(grepl("DataCite Metadata Record", out)))
+  expect_true(any(grepl("Jane Doe", out)))
+  expect_true(any(grepl("John Smith", out)))
+  expect_true(any(grepl("FruitLab", out)))
+})
 
 
 
-?datacite
+test_that("datacite_to_triples works with minimal input", {
+  input <- list(title = "Test Dataset")
+  out <- datacite_to_triples(input)
+  expect_type(out, "character")
+  expect_length(out, 1)
+  expect_true(grepl("datacite.org/schema/kernel-4/title", out[1]))
+})
+
+test_that("datacite_to_triples works with full metadata", {
+  input <- list(
+    title = "Test Dataset",
+    creator = "Jane Doe",
+    contributor = "John Smith",
+    identifier = "https://doi.org/10.1234/test",
+    publisher = "Test Publisher",
+    publicationyear = "2024",
[…]
