### 2025-05-22 fa8f698
remove warnings from examples
