+++ data-raw/orange_df.R
@@ -1,25 +1,4 @@
 ## code to prepare `orange_df` dataset goes here
-
-orange_df <- dataset_df(
-  rowid = defined(paste0("orange:", row.names(Orange)),
-    label = "ID in the Orange dataset",
-    namespace = c("orange" = "datasets::Orange")
-  ),
-  tree = defined(Orange$Tree,
-    label = "The number of the tree"
-  ),
-  age = defined(Orange$age,
-    label = "The age of the tree",
-    unit = "days since 1968/12/31"
-  ),
-  circumference = defined(Orange$circumference,
-    label = "circumference at breast height",
-    unit = "milimeter",
-    concept = "https://www.wikidata.org/wiki/Property:P2043"
-  ),
-  dataset_bibentry = orange_bibentry
-)
-
 orange_bibentry <- dublincore(
   title = "Growth of Orange Trees",
   creator = c(
@@ -47,6 +26,27 @@ orange_bibentry <- dublincore(
   language = "en",
   description = "The Orange data frame has 35 rows and 3 columns of records of the growth of orange trees."
 )
+orange_df <- dataset_df(
+  rowid = defined(paste0("orange:", row.names(Orange)),
+    label = "ID in the Orange dataset",
+    namespace = c("orange" = "datasets::Orange")
+  ),
+  tree = defined(Orange$Tree,
+    label = "The number of the tree"
+  ),
+  age = defined(Orange$age,
+    label = "The age of the tree",
+    unit = "days since 1968/12/31"
+  ),
+  circumference = defined(Orange$circumference,
+    label = "circumference at breast height",
+    unit = "milimeter",
+    concept = "https://www.wikidata.org/wiki/Property:P2043"
+  ),
+  dataset_bibentry = orange_bibentry
+)
+
+
 
 orange_df$rowid <- defined(orange_df$rowid,
                            namespace = "https://doi.org/10.5281/zenodo.14917851")
