+++ NAMESPACE
@@ -115,7 +115,6 @@ export(var_label)
 export(var_namespace)
 export(var_unit)
 export(xsd_convert)
-import(methods)
 import(pillar)
 import(vctrs)
 importFrom(assertthat,assert_that)
