+++ README.md
@@ -14,252 +14,104 @@ WIP](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.o
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci/software-review/issues/553)
 [![DOI](https://zenodo.org/badge/DOI/10.32614/CRAN.package.dataset.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
-[![devel-version](https://img.shields.io/badge/devel%20version-0.3.4024-blue.svg)](https://github.com/dataobservatory-eu/dataset)
+[![devel-version](https://img.shields.io/badge/devel%20version-0.3.4026-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 [![Codecov test
 coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset)
 
 <!-- badges: end -->
 
-The aim of the *dataset* package is to make tidy datasets easier to
-release, exchange and reuse. It organizes and formats data frame R
-objects into well-referenced, well-described, interoperable datasets
-into release and reuse ready form.
+# dataset: Semantic Metadata for Datasets in R
 
-It enhances R’s data frames (`data.frame`, `tibble`) by allowing you to
-attach machine-readable metadata to datasets and variables using R’s
-native `attributes()`. This includes units, labels, definitions,
-provenance, and persistent identifiers. This help making datasets *FAIR*
-— Findable, Accessible, Interoperable, and Reusable — with embedding
-semantics from the start.
+The `dataset` package provides tools to create semantically rich and
+interoperable datasets in R. It improves metadata handling by
+introducing new S3 classes—`defined()`, `dataset_df()`, and
+`bibrecord()`—that enhance the behavior of `labelled`, `tibble`, and
+`bibentry` objects to meet the requirements of:
 
-## 🔧 Installation
+- **Statistical Data and Metadata eXchange (SDMX)** standards,
+- **Open Science** metadata practices,
+- **Library and archive metadata** conventions (Dublin Core, DataCite).
 
-You can install the latest CRAN release with
-`install.packages("dataset")`.
+## Motivation
 
-``` r
-install.packages("dataset")
-```
+Many tools exist to help document, describe, or publish datasets in R,
+but most separate the metadata from the data itself. This separation
+increases the risk of losing metadata, misaligning it with the data, or
+making documentation hard to maintain.
 
-The latest development version of dataset can be installed with
-`remotes::install_github()`:
+The `dataset` package addresses this by storing all metadata directly in
+R object attributes. This preserves semantic information as data is
+transformed, combined, or exported, preventing the loss of vital
+documentation and improving reproducibility.
 
-``` r
-install.packages("pak")
-pak::pak("dataobservatory-eu/dataset")
-```
+## Key Features
 
-## 📦 Package Status
+### `defined()`
 
-This package is under active development and currently under review by
-rOpenSci:
+An extended version of `labelled()` vectors. Adds support for:
 
-- [Original submission
-  (#553)](https://github.com/ropensci/software-review/issues/553) —
[…]
