### 2025-05-22 87edbf3
correct tests for defintions / concepts
