### 2025-05-22 0fe50f5
make vctrs::vec_data explicit
