### 2025-05-23 5cb11f7
remove potentially non-ascii characters
