+++ DESCRIPTION
@@ -43,6 +43,7 @@ Imports:
     vctrs
 RoxygenNote: 7.3.2
 Suggests: 
+    jsonld,
     knitr,
     rdflib,
     rmarkdown,
