### 2025-05-23 3a16f9d
lintr and code formatting improvements
