+++ R/bibrecord.R
@@ -23,8 +23,7 @@ bibrecord <- function(
     date = Sys.Date(),
     identifier = NULL,
     subject = NULL,
-    ...
-) {
+    ...) {
   stopifnot(inherits(author, "person") || all(vapply(author, inherits, logical(1), "person")))
 
   # Normalize author roles and ensure vector of class "person"
@@ -81,4 +80,3 @@ print.bibrecord <- function(x, ...) {
 
   invisible(x)
 }
-
