+++ tests/testthat/test-dataset_title.R
@@ -1,6 +1,7 @@
 test_that("dataset_title() works", {
   expect_equal(dataset_title(iris_dataset), "Iris Dataset")
-  expect_error(dataset_title(mtcars), regexp = "must be a dataset object created with dataset")
+  expect_error(dataset_title(mtcars),
+               regexp = "must be a dataset object created with dataset")
   expect_error(dataset_title(mtcars) <- "Error")
 })
 
@@ -12,5 +13,7 @@ test_that("dataset_title() <- value works with overwrite", {
 
 test_that("dataset_title() <- value works without overwrite", {
   iris_dataset_2 <- iris_dataset
-  expect_warning(dataset_title(x = iris_dataset_2, overwrite = FALSE) <- "The Most Famous Iris Dataset")
+  expect_warning(
+    dataset_title(x = iris_dataset_2,
+                  overwrite = FALSE) <- "The Most Famous Iris Dataset")
 })
