+++ R/dataset-orange_df.R
@@ -1,8 +1,8 @@
 #' @title Growth of Orange Trees
 #'
 #' @description
-#' The Orange data frame has 35 rows and 3 columns of records of
-#'  the growth of orange trees. This is a replication of
+#' The Orange data frame has 35 rows and 3 columns of records of\cr
+#'  the growth of orange trees. This is a replication of\cr
 #' \code{datasets::\link[datasets]{Orange}} as \emph{dataset_df} s3 class.
 #'
 #' @details See \code{datasets::\link[datasets]{Orange}} for details.
