+++ tests/testthat/test-var_label.R
@@ -1,17 +1,20 @@
 test_that("var_label() works", {
-  expect_equal(var_label(iris_dataset$Sepal.Length), "Length of the sepal in cm")
+  expect_equal(var_label(iris_dataset$Sepal.Length),
+               "Length of the sepal in cm")
   expect_equal(class(var_label(iris_dataset, unlist = TRUE)), "character")
   test_df <- dataset_df(a = 1:2, b = 3:4)
-  expect_equal(var_label(test_df, unlist = TRUE, null_action = "fill"), c(rowid = "rowid", a = "a", b = "b"))
-  # expect_equal(label_attribute(iris_dataset$Species), "Taxon name within the Iris genus")
+  expect_equal(var_label(test_df, unlist = TRUE, null_action = "fill"),
+               c(rowid = "rowid", a = "a", b = "b"))
 })
 
 test_that("var_label()<-  works", {
   iris_dataset_2 <- iris_dataset
   var_label(orange_df$circumference) <- "circumference (breast height)"
   var_label(iris_dataset_2$Sepal.Length) <- "Length of the sepal in centimeters"
-  expect_equal(var_label(iris_dataset_2$Sepal.Length), "Length of the sepal in centimeters")
-  expect_equal(var_label(orange_df$circumference), "circumference (breast height)")
+  expect_equal(var_label(iris_dataset_2$Sepal.Length),
+               "Length of the sepal in centimeters")
+  expect_equal(var_label(orange_df$circumference),
+               "circumference (breast height)")
 })
 
 test_that("var_label() throws error", {
