+++ R/bibrecord.R
@@ -13,6 +13,11 @@
 #' @param ... Additional fields (e.g., language, format, rights, description).
 #' @importFrom utils bibentry
 #' @return An object of class `bibrecord` and `bibentry`.
+#' bibrecord(title = "Gross domestic product, volumes",
+#'           author = person("Eurosat"),
+#'           publisher = person("Eurostat"),
+#'           identifier = "https://doi.org/10.2908/TEINA011",
+#'           date = as.Date("2025-05-20"))
 #' @export
 bibrecord <- function(
     title,
@@ -69,7 +74,6 @@ bibrecord <- function(
 
 #' @export
 print.bibrecord <- function(x, ...) {
-  cat("\nBibliographic Record (bibrecord + bibentry):\n")
   NextMethod("print")
 
   contributors <- attr(x, "contributor")
