### 2025-05-23 9daabba
Documentation and metadata improvements, spelling, formatting for CRAN
