+++ README.Rmd
@@ -47,7 +47,7 @@ An extended version of `labelled()` vectors. Adds support for:
 -   Concept URIs (standardized definitions)
 -   Namespaces (to support URI expansion)
 
-This ensures that, for example, "GDP" is always associated with a precise concept and unit, avoiding ambiguity across analyses and publications.
+This ensures that, for example, "GDP" is always associated with a precise concept and unit, avoiding ambiguity across analyses and publications. See [Semantically Enriched Vectors with `defined()`](https://dataset.dataobservatory.eu/articles/defined.html)
 
 ### `bibrecord()`
 
@@ -58,7 +58,7 @@ An extension of R's built-in `bibentry()` class, with support for:
 -   Contributor roles (e.g. creator, publisher, data manager)
 -   Subject tagging and geolocation
 
-This makes it easier to produce citations and metadata suitable for repositories like [Zenodo](https://zenodo.org/) or [Dataverse](https://dataverse.org/).
+This makes it easier to produce citations and metadata suitable for repositories like [Zenodo](https://zenodo.org/) or [Dataverse](https://dataverse.org/).  See more in the [Modernising Citation Metadata in R: Introducing `bibrecord`](https://dataset.dataobservatory.eu/articles/bibrecord.html)
 
 ### `dataset_df()`
 
@@ -68,6 +68,8 @@ A semantic wrapper around `data.frame` or `tibble`, aligning with SDMX’s **dat
 -   Observations (rows) can be assigned unique identifiers.
 -   Datasets can carry complete metadata inline (title, creator, description, etc.)
 -   Output can be serialized to linked data formats (N-Triples, RDF, etc.)
+
+See more in the [Why Semantics Matter for R Data Frames](https://dataset.dataobservatory.eu/articles/dataset_df.html)
 
 ## Why Use This?
 
@@ -99,7 +101,7 @@ summary(my_data)
 
 We welcome contributions and discussion!
 
--   Please see our [CONTRIBUTING.md](CONTRIBUTING.md) guide.
+-   Please see our [CONTRIBUTING.md](https://github.com/dataobservatory-eu/dataset/blob/main/CONTRIBUTING.md) guide.
 -   Ideas, bug reports, and feedback are welcome via [GitHub issues](https://github.com/dataobservatory-eu/dataset/issues).
 
 ## 📜 Code of Conduct
