+++ NEWS.md
@@ -1,9 +1,10 @@
-# dataset 0.3.4025
+# dataset 0.3.4027
 
 - The new `bibrecord()` class is handles is the superclass of the `dublincore` and
 `datacite()` classes; these classes have a new print method.
 - Breaking change: the `definition` metadata field in the `defined()` class is 
 changed to the more understandable `concept` name.
+- Many code quality improvements and new tests.
 
 # dataset 0.3.4023
 
@@ -110,7 +111,7 @@ A minor correction to avoid vignettes downloading data from the Eurostat data wa
 
 # dataset 0.1.3.
 
-* Vignette articles started to develop and consult the development plan of the project. See  [From dataset To RDF](https://dataset.dataobservatory.eu/articles/RDF.html), _Export and Publish A dataset Object_, _Datasets with FAIR metadata_, all [comments](https://github.com/dataobservatory-eu/dataset/issues/) are welcome.
+* Vignette articles started to develop and consult the development plan of the project. See  [From dataset To RDF](https://dataset.dataobservatory.eu/articles/rdf.html), _Export and Publish A dataset Object_, _Datasets with FAIR metadata_, all [comments](https://github.com/dataobservatory-eu/dataset/issues/) are welcome.
 * New functions: `download_dataset()`, `datacite()`, and the `dataset()` constructor.
 
 # dataset 0.1.2.
@@ -122,4 +123,4 @@ A minor correction to avoid vignettes downloading data from the Eurostat data wa
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6703765.svg)](https://doi.org/10.5281/zenodo.6703765) First development version release.
 
-* Added the [Motivation of the dataset package](https://dataset.dataobservatory.eu/articles/motivation.html) vignette article.
+* Added the [Motivation of the dataset package](https://dataset.dataobservatory.eu/articles/Motivation.html) vignette article.
