### 2025-05-25 41ec8c0
DataCite bibliographic entries can be serialised to N-Triples
