+++ R/datacite.R
@@ -122,7 +122,11 @@
 #'   Language = "en"
 #' )
 #'
+#' # Bibliographic metadata as bibentry...
 #' as_datacite(orange_df)
+#'
+#' # ... or a list:
+#' as_datacite(orange_df, "list")
 #' @export
 
 datacite <- function(Title,
@@ -160,11 +164,15 @@ datacite <- function(Title,
   Date <- ifelse(is.null(DateList), ":tba", as.character(Date))
   DateList <- ifelse(is.null(DateList), ":tba", as.character(DateList))
   Format <- ifelse(is.null(Format), ":tba", as.character(Format))
-  AlternateIdentifier <- ifelse(is.null(AlternateIdentifier), ":unas", AlternateIdentifier)
-  RelatedIdentifier <- ifelse(is.null(RelatedIdentifier), ":unas", RelatedIdentifier)
+  AlternateIdentifier <- ifelse(is.null(AlternateIdentifier),
+                                ":unas", AlternateIdentifier)
+  RelatedIdentifier <- ifelse(is.null(RelatedIdentifier),
+                              ":unas", RelatedIdentifier)
   Rights <- ifelse(is.null(Rights), ":tba", as.character(Rights))
-  Geolocation <- ifelse(is.null(Geolocation), ":unas", as.character(Geolocation))
-  FundingReference <- ifelse(is.null(FundingReference), ":unas", as.character(FundingReference))
+  Geolocation <- ifelse(is.null(Geolocation), ":unas",
+                        as.character(Geolocation))
+  FundingReference <- ifelse(is.null(FundingReference),
+                             ":unas", as.character(FundingReference))
 
   new_datacite(
     Title = Title,
@@ -239,126 +247,6 @@ new_datacite <- function(Title,
   datacite_object
 }
 
-#' @rdname datacite
-#' @param x A dataset object created with \code{dataset::\link{dataset}}.
-#' @param type A DataCite 4.4  metadata can be returned as a \code{type="list"},
-#'   a \code{type="dataset_df"}, or a \code{type="bibentry"} (default).
-#' @param ... Optional parameters to add to a \code{datacite} object.
-#'   \code{author=person("Jane", "Doe")} adds an author to the citation object
-#'   if \code{type="dataset"}. as_datacite(orange_df, type="list")
-#' @return \code{as_datacite(x, type)} returns the DataCite bibliographical
-#'   metadata of x either as a list, a bibentry object, or a dataset_df object.
-#' @export
-as_datacite <- function(x, type = "bibentry", ...) {
-  citation_author <- person(NULL, NULL)
-
-  is_person <- function(p) ifelse(inherits(p, "person"), TRUE, FALSE)
-
-  arguments <- list(...)
-  if (!is.null(arguments$author)) {
-    if (is_person(arguments$author)) {
-      citation_author <- arguments$author
-    } else {
-      stop("as_datacite(x, ..., author = ): author must be created with utils::person().")
-    }
-  }
-
-  if (!type %in% c("bibentry", "list", "dataset_df")) {
-    warning_message <- "as_datacite(ds, type=...) type cannot be "
-    warning(warning_message, type, ". Reverting to 'bibentry'.")
-    type <- "bibentry"
-  }
-
-  ds_bibentry <- get_bibentry(x)
-  Title <- ds_bibentry$title
-  Creator <- ds_bibentry$author
-  Publisher <- ifelse(is.null(ds_bibentry$publisher), ":unas", as.character(ds_bibentry$publisher))
-  Identifier <- ifelse(is.null(ds_bibentry$identifier), ":tba", as.character(ds_bibentry$identifier))
[…]
