### 2025-05-25 3393ba1
Improve the printing of dataset_df
Improve the examples of the dataset_df class

Improve testing
