### 2025-06-22 #681 @maurolepore
@antaldaniel, any updates about the changes?

It would be really nice to keep the momentum. How about prioritizing README and a single Get started article? Minimal docs that meet our requirements would be best to move fast towards publication. After that we won't block you from adding more articles.
