### 2025-06-23 #681 @maurolepore
Oh, my bad, it seems that README is already much more focused. Thanks! The only tiny suggestion there would be to print the output of `my_data` and of `summary(my_data)` so we can see what it looks like before we even install the package.

Thanks!
