+++ README.md
@@ -34,34 +34,31 @@ into release and reuse ready form.
     into the attributes of a data frame-like object. See for more
     information the [Bibentry for FAIR
     datasets](https://dataset.dataobservatory.eu/articles/bibentry.html)
-    vignette. 2.**Interoperability outside R**: Extending the
-    `haven_labelled` class of the `tidyverse` for consistently labelled
-    categorical variables with linked (standard) definitions and units
-    of measures in our
-    [defined](https://dataset.dataobservatory.eu/articles/defined.html)
-    class; this enables to share exact definitions, units of measures
-    across computers and systems, and increasing the interoperability of
-    the data set from an R data.frame to any standardised statistical or
-    library system.
-2.  **Tidy data tidier, richer**: Offering a new data frame format,
-    `dataset_df` that extends tibbles with semantically rich metadata,
-    ready to be shared on open data exchange platforms and in data
-    repositories. This s3 class is aimed at developers and we are
-    working on several packages that provide interoperability with SDMX
-    statistical data exchange platforms, Wikidata, or the EU Open Data
-    portal. Read more in the [Create Datasets that are Easy to Share
-    Exchange and
-    Extend](https://dataset.dataobservatory.eu/articles/dataset_df.html)
     vignette.
-3.  **R+RDF=global interoperability**: The [From R to
-    RDF](https://dataset.dataobservatory.eu/articles/rdf.html) vignette
-    shows how to leverage the capabilities of the *dataset* package with
-    [rdflib](https://docs.ropensci.org/rdflib/index.html), an
-    R-user-friendly wrapper on ROpenSci to work with the *redland*
-    Python library for performing common tasks on rdf data, such as
-    parsing and converting between formats including rdfxml, turtle,
-    nquads, ntriples, and trig, creating rdf graphs, and performing
-    SPARQL queries.
+
+2.**Interoperability outside R**: Extending the `haven_labelled` class
+of the `tidyverse` for consistently labelled categorical variables with
+linked (standard) definitions and units of measures in our
+[defined](https://dataset.dataobservatory.eu/articles/defined.html)
+class; this enables to share exact definitions, units of measures across
+computers and systems, and increasing the interoperability of the data
+set from an R data.frame to any standardised statistical or library
+system. 3. **Tidy data tidier, richer**: Offering a new data frame
+format, `dataset_df` that extends tibbles with semantically rich
+metadata, ready to be shared on open data exchange platforms and in data
+repositories. This s3 class is aimed at developers and we are working on
+several packages that provide interoperability with SDMX statistical
+data exchange platforms, Wikidata, or the EU Open Data portal. Read more
+in the [Create Datasets that are Easy to Share Exchange and
+Extend](https://dataset.dataobservatory.eu/articles/dataset_df.html)
+vignette. 4. **R+RDF=global interoperability**: The [From R to
+RDF](https://dataset.dataobservatory.eu/articles/rdf.html) vignette
+shows how to leverage the capabilities of the *dataset* package with
[…]
