### 2025-07-22 52336a0
fix_contributors() now needs a valid ORCiD
