### 2025-07-22 b7d5776
Fix a bug with triples
