### 2025-07-22 b89fbc1
syntax error with messages in exception
