+++ R/dublincore.R
@@ -288,7 +288,7 @@ new_dublincore <- function (title,
                           email=creator[[i]]$email, role=creator[[i]]$role, comment=creator[[i]]$comment,
                           first=creator[[i]]$first, last=creator[[i]]$last)
       } else {
-        mesage(i)
+        message(i)
         tmp <- person(given = creator[[i]]$given, middle = creator[[i]]$middle, family=creator[[i]]$family,
                           email=creator[[i]]$email, role=creator[[i]]$role, comment=creator[[i]]$comment,
                           first=creator[[i]]$first, last=creator[[i]]$last)
@@ -380,43 +380,3 @@ fix_publisher <- function(publishers) {
 }
 
 
-#' @keywords internal
-fix_contributor <- function(contributors=NULL) {
-
-  if (is.null(contributors)) return(":unas")
-
-  if ( all(inherits(contributors, "person")) ) {
-    if( length(contributors)>1 ) {
-      return_value <- paste0("{",
-                             paste( vapply(contributors, function(x) {as.character(x)}, character(1)),
-                                    collapse="} and {"),
-                             "}" )
-    } else {
-      return_value <- as.character(contributors)
-    }
-  } else if ( all(inherits(contributors, "list")) ) {
-    if( length(contributors)>1 ) {
-      return_value <- paste0("{",
-                             paste( lapply(contributors, function(x) x$given),
-                                    collapse="} and {"),
-                             "}" )
-    } else {
-      return_value <- paste(unlist(contributors[[1]]), collapse=" ")
-    }
-  } else if (length(contributors)>1) {
-    # several character strings
-    return_value <- paste0("{",
-                           paste( vapply(contributors, function(x) x$given, character(1)),
-                                  collapse="} and {"),
-                           "}" )
-  }  else {
-    return_value <- contributors
-  }
-
-  assertthat::assert_that(is.character(return_value),
-                          msg="Error: fix_contributor(contributors): not character but")
-  assertthat::assert_that(length(return_value)==1, msg="Error: fix_contributor(contributors): not 1" )
-
-  return_value <- gsub("* dtm", " [dtm]", return_value )
-  return_value
-}
