### 2025-07-22 b7b5777
New example dataset for documentation, examples, and unit testing
