+++ R/data-gdp.R
@@ -1,0 +1,16 @@
+#' A Small GDP Dataset
+#'
+#' A subset of data from the Eurostat's GDP and main aggregates -
+#' international data cooperation annual data
+#'
+#' @format ## `gdp`
+#' A data frame with 10 rows and 5 columns:
+#' \describe{
+#'   \item{geo}{Country name}
+#'   \item{year}{Reference time period}
+#'   \item{gdp}{Gross Domestic Product}
+#'   \item{unit}{Unit of values}
+#'   \item{freq}{Frequency}
+#' }
+#' @source <https://doi.org/10.2908/NAIDA_10_GDP>
+"gdp"
