### 2025-07-22 2bc0e41
New n_triple function
