+++ tests/testthat/test-n_triple.R
@@ -1,31 +1,65 @@
+test_that("n_triple and related helpers produce correct N-Triple strings", {
+  # String literal object
+  triple_1 <- n_triple(
+    "http://example.org/show/218",
+    "http://www.w3.org/2000/01/rdf-schema#label",
+    "That Seventies Show"
+  )
 
-test_that("n_triples()", {
-  triple_1 <- n_triple("http://example.org/show/218", "http://www.w3.org/2000/01/rdf-schema#label", "That Seventies Show")
-  triple_2 <- n_triple("http://example.org/show/218", "http://example.org/show/localName",  '"Cette Série des Années Septante"@fr-be')
-  author_person <- person(given="Daniel", family="Antal", comment = c(ORCID = "https://orcid.org/0000-0001-7513-6760"))
-  expect_equal(triple_1, "<http://example.org/show/218> <http://www.w3.org/2000/01/rdf-schema#label> \"That Seventies Show\"^^<http://www.w3.org/2001/XMLSchema#string> .")
-  expect_equal(n_triple(s= "http://example.com/creation",
-                        p= "http://www.w3.org/ns/prov#generatedAtTime",
-                        o= as.POSIXct(10000, origin = "2024-01-01", tz="UTC")),
-               "<http://example.com/creation> <http://www.w3.org/ns/prov#generatedAtTime> \"2024-01-01T03:46:40Z\"^^<xs:dateTime> .")
+  # Language-tagged literal object
+  triple_2 <- n_triple(
+    "http://example.org/show/218",
+    "http://example.org/show/localName",
+    '"Cette Série des Années Septante"@fr-be'
+  )
+
+  # XSD datetime literal from POSIXct
+  triple_time <- n_triple(
+    "http://example.com/creation",
+    "http://www.w3.org/ns/prov#generatedAtTime",
+    as.POSIXct(10000, origin = "2024-01-01", tz = "UTC")
+  )
+
+  # ORCID agent as subject
+  triple_agent <- n_triple(
+    "https://orcid.org/0000-0001-7513-6760",
+    "a",
+    "http://www.w3.org/ns/prov#Agent"
+  )
+
+  # Person object with ORCID
+  author_person <- person(
+    given = "Daniel",
+    family = "Antal",
+    comment = c(ORCID = "https://orcid.org/0000-0001-7513-6760")
+  )
+
+  # Expectations
+  expect_equal(triple_1,
+               '<http://example.org/show/218> <http://www.w3.org/2000/01/rdf-schema#label> "That Seventies Show"^^<http://www.w3.org/2001/XMLSchema#string> .'
+  )
+
+  expect_equal(triple_time,
+               '<http://example.com/creation> <http://www.w3.org/ns/prov#generatedAtTime> "2024-01-01T03:46:40Z"^^<xs:dateTime> .'
+  )
+
   expect_equal(length(n_triples(c(triple_1, triple_2, triple_1))), 2)
   expect_equal(length(n_triples(c(triple_1, triple_2))), 2)
-  expect_equal(n_triple("https://orcid.org/0000-0001-7513-6760", "a",  'http://www.w3.org/ns/prov#Agent'),
-               '<https://orcid.org/0000-0001-7513-6760> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.w3.org/ns/prov#Agent> .')
-  expect_equal(create_iri(23), '\"23\"^^<http://www.w3.org/2001/XMLSchema#double>')
-  expect_equal(create_iri("23"), '\"23\"^^<http://www.w3.org/2001/XMLSchema#string>')
[…]
