### 2025-07-22 276abbc
rewrite conversion to dublincore metadata
