### 2025-07-22 5aaad6f
dataset_to_triples converter completely rewritten.
