+++ R/dataset_to_triples.R
@@ -1,46 +1,143 @@
-#' @title Dataset to triples (three columns)
-#' @description The dataset is converted into a three-column long format with
-#' columns \code{s} for subject, \code{p} for predicate and \code{o} for
-#' object.
-#' @param x An R object that contains the data of the dataset (a data.frame or
-#' inherited from [`data.frame`][base::data.frame()]), for example,
-#' [dataset_df()].
-#' @param idcol The identifier column. If \code{idcol} is \code{NULL} it
-#' attempts to use the \code{row.names(df)} as an \code{idcol}.
-#' @return The long form version of the original dataset, retaining the
-#' attributes and class.
-#' @examples
-#' dataset_to_triples(orange_df)
+#' @title Dataset to triples (three columns or N-Triples)
+#' @description Converts a dataset to RDF-style triples with subject, predicate,
+#'   and object columns. Supports semantic expansion via variable metadata.
+#' @param x A `dataset_df` or `data.frame`.
+#' @param idcol Name or index of the subject column. If NULL, defaults to
+#'   `"rowid"` or rownames.
+#' @param expand_uri Logical; if TRUE, expands URIs using namespaces and
+#'   definitions.
+#' @param format Output format: `"data.frame"` (default) or `"nt"` for
+#'   N-Triples.
+#' @return Either a `data.frame` with columns `s`, `p`, and `o`, or a character
+#'   vector of N-Triple lines.
 #' @export
-
-dataset_to_triples <- function(x, idcol = NULL) {
+#' @examples
+#' # A minimal example with just rowid and geo
+#' data("gdp", package = "dataset")
+#' small_geo <- dataset_df(
+#'   rowid = defined(
+#'     paste0("eg:", seq_along(gdp$geo[1:3])),
+#'     namespace = "http://example.com/dataset#"
+#'   ),
+#'   geo = defined(
+#'     gdp$geo[1:3],
+#'     label = "Geopolitical entity",
+#'     concept = "http://example.com/prop/geo",
+#'     namespace = "https://dd.eionet.europa.eu/vocabulary/eurostat/geo/$1"
+#'   )
+#' )
+#'
+#' # View as triple table
+#' dataset_to_triples(small_geo)
+#'
+#' # View as N-Triples
+#' dataset_to_triples(small_geo, format = "nt")
+dataset_to_triples <- function(x,
+                               idcol = NULL,
+                               expand_uri = TRUE,
+                               format = "data.frame") {
+  stopifnot(is.data.frame(x))
   is_dataset <- inherits(x, "dataset_df")
 
-  if (is_dataset) {
-    new_title <- paste0(dataset_title(x), " [triple form]")
-    DataBibentry <- get_bibentry(x)
-    new_Subject <- subject(x)
-  }
-
+  # Determine subject (s)
   if (is.null(idcol)) {
-    x$new_id_col <- row.names(x)
-    idcol <- which(names(x) == "new_id_col")
-    idcol_pos <- idcol_find(x, idcol)
-    seq_along_cols <- seq_along(x)[-idcol_pos]
-  } else {
-    ## See utils-idcol_find.R for the internal function
-    seq_along_cols <- seq_along(x)[-idcol_find(x, idcol)]
+    if ("rowid" %in% names(x)) {
+      idcol <- "rowid"
+    } else {
+      x <- cbind(rowid = row.names(x))
+      idcol <- "rowid"
+    }
   }
 
-  triple_list <- lapply(seq_along_cols, function(i) {
-    data.frame(
-      s = as.character(x[[idcol]]),
-      p = names(x)[i],
[…]
