### 2025-07-22 d63e087
New n_triple function for individual triples
