+++ R/n_triple.R
@@ -7,16 +7,12 @@
 #' @param triples Concatenated N-Triples created with \code{\link{n_triple}}.
 #' @return A character vector containing unique N-Triple strings.
 #' @examples
-#' triple_1 <- n_triple(
-#'   "http://example.org/show/218",
-#'   "http://www.w3.org/2000/01/rdf-schema#label",
-#'   "That Seventies Show"
-#' )
-#' triple_2 <- n_triple(
-#'   "http://example.org/show/218",
-#'   "http://example.org/show/localName",
-#'   '"Cette Série des Années Septante"@fr-be'
-#' )
+#' triple_1 <- n_triple("http://example.org/show/218",
+#'                       "http://www.w3.org/2000/01/rdf-schema#label",
+#'                       "That Seventies Show")
+#' triple_2 <- n_triple("http://example.org/show/218",
+#'                      "http://example.org/show/localName",
+#'                      '"Cette Série des Années Septante"@fr-be')
 #' n_triples(c(triple_1, triple_2, triple_1))
 #' @export
 
@@ -41,28 +37,38 @@ n_triples <- function(triples) {
 #' o <- "That Seventies Show"
 #' n_triple(s, p, o)
 #' @export
-
 n_triple <- function(s, p, o) {
   s <- create_iri(s)
   p <- create_iri(p)
-  o <- create_iri(o)
 
-  sprintf("%s %s %s .", s, p, o)
+  # Don't convert `o` if it's already quoted with ^^<...>
+  if (grepl('^".+"\\^\\^<.+>$', o) || grepl('^<.+>$', o)) {
+    # o is already a well-formed literal or URI
+  } else {
+    o <- create_iri(o)
+  }
+
+  sprintf('%s %s %s .', s, p, o)
 }
+
 
 #' @keywords internal
 create_iri <- function(x) {
-  if (any(c("list", "data.frame", "tbl", "data.table") %in% class(x))) {
-    stop("Error: create_iri(x) must be a URI, string, integer, double, Date, dateTime, or person.")
+
+  if (length(x) != 1) {
+    stop("create_iri(): input must be a scalar value (length = 1).")
   }
 
-  double_string <- "^^<http://www.w3.org/2001/XMLSchema#double>"
-  integer_string <- "^^<http://www.w3.org/2001/XMLSchema#integer>"
-  character_string <- "^^<http://www.w3.org/2001/XMLSchema#string>"
-  date_string <- "^^<http://www.w3.org/2001/XMLSchema#date>"
-  datetime_string <- "^^<http://www.w3.org/2001/XMLSchema#dateTime>"
+  if (any(c("list", "data.frame", "tbl", "data.table") %in% class(x))) {
+    stop("Error: create_iri(x) must be a scalar URI, string, integer, double, Date, dateTime, or person.")
+  }
 
-  # Person object handling
+  double_string    <- '^^<http://www.w3.org/2001/XMLSchema#double>'
+  integer_string   <- '^^<http://www.w3.org/2001/XMLSchema#integer>'
+  character_string <- '^^<http://www.w3.org/2001/XMLSchema#string>'
+  date_string      <- '^^<http://www.w3.org/2001/XMLSchema#date>'
+  datetime_string  <- '^^<http://www.w3.org/2001/XMLSchema#dateTime>'
+
   if (inherits(x, "person")) {
     if ("isni" %in% tolower(names(x$comment))) {
       x <- paste0("https://isni.org/isni/", x$comment[which(tolower(names(x$comment)) == "isni")])
@@ -75,30 +81,20 @@ create_iri <- function(x) {
       qid <- gsub("https://www.wikidata.org/wiki/", "", qid)
       x <- paste0("https://www.wikidata.org/wiki/", qid)
     } else {
[…]
