+++ man/gdp.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data-gdp.R
+\docType{data}
+\name{gdp}
+\alias{gdp}
+\title{A Small GDP Dataset}
+\format{
+\subsection{\code{gdp}}{
+
+A data frame with 10 rows and 5 columns:
+\describe{
+\item{geo}{Country name}
+\item{year}{Reference time period}
+\item{gdp}{Gross Domestic Product}
+\item{unit}{Unit of values}
+\item{freq}{Frequency}
+}
+}
+}
+\source{
+\url{https://doi.org/10.2908/NAIDA_10_GDP}
+}
+\usage{
+gdp
+}
+\description{
+A subset of data from the Eurostat's GDP and main aggregates -
+international data cooperation annual data
+}
+\keyword{datasets}
