### 2025-07-22 fbaf291
New describe function, needs new example
