+++ R/describe.R
@@ -1,42 +1,37 @@
-#' @title Describe a dataset
-#' @param x A dataset_df object.
-#' @param con A connection, for example, \code{con=tempfile()}.
-#' @return The description of the dataset_df object is written to the
-#' connection in the N-Triples form. If \code{con=NULL}, then the serialisation
-#' takes place in \code{tempfile()} and the contents are printed to the console;
-#' if a file is given, than no output is returned.
-#' @examples
-#'
-#' # See the serialisation on the screen:
-#' describe(orange_df)
-#'
-#' # Save it to a connection:
-#' temporary_connection <- tempfile()
-#' describe(orange_df, con = temporary_connection)
+#' @title Describe a dataset in N-Triples format
+#' @description Writes provenance and Dublin Core metadata of a dataset to a
+#'   file or connection in N-Triples format.
+#' @param x A `dataset_df` object.
+#' @param con A connection or a character string path (e.g. from `tempfile()`).
+#' @return Invisibly returns `x`. Writes N-Triples to `con`.
 #' @export
+describe <- function(x, con) {
+  assertthat::assert_that(
+    is.dataset_df(x),
+    msg = "describe(x, con): x must be a dataset_df object."
+  )
 
-describe <- function(x, con=NULL) {
+  # Collect provenance and DC metadata
+  prov_text <- provenance(x)
+  dc_text <- as_dublincore(x, type = "ntriples")
 
-  if(!inherits(x, "dataset_df")) {
-    stop("Error: describe(x, con) - x most be a a dataset_df object.")
+  prov_text <- if (is.null(prov_text)) character(0) else as.character(prov_text)
+  dc_text <- if (is.null(dc_text)) character(0) else as.character(dc_text)
+
+  ntriples_text <- c(prov_text, dc_text)
+
+  if (!is.character(ntriples_text)) {
+    stop("describe(): expected character vector of N-Triples.")
   }
 
-  ntriples_text <- provenance(x)
-  ntriples_text <- c(ntriples_text, as_dublincore(x, "ntriples"))
-
-  if (is.null(con)) {
-    no_connection <- TRUE
-    con <- tempfile()
-  } else {
-      no_connection <- FALSE
-    }
-
-  writeLines(ntriples_text, con = con)
-
-  if (no_connection) {
-    writeLines(ntriples_text, con = con)
-    readLines(con = con)
+  # Handle connection or file path
+  if (is.character(con)) {
+    con_file <- file(con, open = "w", encoding = "UTF-8")
+    on.exit(close(con_file), add = TRUE)
+    writeLines(ntriples_text, con = con_file)
   } else {
     writeLines(ntriples_text, con = con)
   }
+
+  invisible(x)
 }
