### 2025-07-22 112afa0
Rewrite defined vignette
