+++ vignettes/defined.Rmd
@@ -6,96 +6,90 @@ vignette: >
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
-
-```{r setupdefined, include = FALSE}
+```{r setupdefinedvignette, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 ```
 
-```{r load}
+The `defined()` function creates semantically rich labelled vectors. These are a inherited from haven::labelled and extend the concept by adding:
+
+- A human-readable variable label (available in haven::labelled)
+- A unit of measurement
+- A definition URI linking to external ontologies or concepts
+- (Optionally) a namespace for controlled vocabularies
+
+This structure enhances clarity and semantic precision, particularly useful for statistical metadata, time series, and linked open data.
+
+## Getting Started
+
+```{r setup}
 library(dataset)
+data("gdp")
 ```
 
-The `defined()` function in the `dataset` package allows you to create semantically enriched vectors that retain human-readable metadata — including labels, measurement units, definitions (e.g. URIs), and namespaces.
+We’ll start by wrapping a numeric GDP vector using `defined()`:
 
-This vignette demonstrates how to create, manipulate, and interpret defined vectors, and how they integrate seamlessly into data frames and tidy workflows.
+```{r gdp1}
+gdp_1 = defined(
+    gdp$gdp, 
+    label = "Gross Domestic Product", 
+    unit = "CP_MEUR", 
+    concept = "http://data.europa.eu/83i/aa/GDP")
+```
 
-## Introducting the  `defined` Class
+The `defined()` class builds on labelled vectors by adding rich metadata:
 
-The `defined()` constructor enriches a vector by attaching additional attributes that convey semantic meaning. It builds upon the foundation of labelled vectors and introduces three further metadata elements:
+- label: a description of what the variable represents.
+- unit: a short code for the measurement unit (e.g., CP_MEUR).
+- definition: a URI linking to a concept or standard.
+- namespace: (optional) for classifying coded values (e.g., SDMX codes).
 
--  A `unit of measurement` (e.g. "million dollars")
+This is particularly useful for reproducible research, standard-compliant data, or long-term interoperability. The class is defined with the attributes() function of R, and this guarantees a very wide compatibility. A defined vector can be used even in base R. 
 
--  A `concept`, which can be a textual reference or ideally a URI
+```{r seeattributes}
+attributes(gdp_1)
+```
+Use the `var_label()`, `var_unit()` and `var_defintion()` helper functions to set or retrieve metadata individually:
 
--  A `namespace`, which enables the construction of meaningful, resolvable identifiers for values or categories
+```{r convenience}
+cat("Get the label only: ", var_label(gdp_1), "\n")
+cat("Get the unit only: ", var_unit(gdp_1), "\n")
+cat("Get the concept definition only: ", var_concept(gdp_1), "\n")
+cat("All attributes:\n")
+```
+## Printing and Summary
 
-Let’s inspect the metadata attached to a defined vector representing GDP values:
-
-```{r definedvector}
-gdp_1 <- defined(
-  c(3897, 7365),
-  label = "Gross Domestic Product",
-  unit = "million dollars",
-  concept = "http://data.europa.eu/83i/aa/GDP"
-)
-
-cat("The print method:\n")
+```{r printdefined}
 print(gdp_1)
-cat("And the summary:\n")
+```
+```{r summarydefined}
 summary(gdp_1)
 ```
[…]
