### 2025-07-22 40a95d8
rewrite rdf vignette
