+++ vignettes/rdf.Rmd
@@ -22,108 +22,150 @@ if (!requireNamespace("jsonld", quietly = TRUE)) {
 }
 ```
 
+## From tidy data to RDF triples
+
+This vignette demonstrates how to convert tidy R datasets into semantically enriched RDF triple structures, using the `dataset` and `rdflib` packages. These packages help you annotate variables with machine-readable concept definitions, units, and links to controlled vocabularies.
+
+We’ll start with a small example of a tidy dataset representing countries (`geo`) with unique identifiers (`rowid`) and then show how to transform the dataset into RDF triples using standard vocabularies.
+
 ```{r setup}
 library(dataset)
 library(rdflib)
-library(jsonld)
+data("gdp")
 ```
 
-The **RDF** (Resource Description Framework) annotation significantly enhances the interoperability and exchangeability of datasets in data repositories by leveraging a standardised, machine-readable format for describing and linking data. This vignette shows how to leverage the capabilities of the _dataset_ package with [rdflib](https://docs.ropensci.org/rdflib/index.html), an R-user-friendly wrapper on rOpenSci to work with the _redland_ Python library for performing common tasks on rdf data, such as parsing and converting between formats including rdfxml, turtle, nquads, ntriples, and trig, creating rdf graphs, and performing SPARQL queries.
+## Creating a minimal semantically defined dataset
 
-## Standardised Semantic Framework
-RDF provides a common framework to describe resources and their relationships using triples (subject-predicate-object). This standardisation ensures that data from different systems can be understood in a unified way, regardless of the original source or format.  Notice that this format is a stricter version of the tidy dataset concept, where not only on every observation is in a row, but there are always strictly three columns.
+```{r smallgeo}
+small_geo <- dataset_df(
+  rowid = defined(
+    paste0("eg:", seq_along(gdp$geo[1:3])),
+    namespace = "http://example.com/dataset#"
+  ),
+  geo = defined(
+    gdp$geo[1:3],
+    label = "Geopolitical entity",
+    concept = "http://purl.org/linked-data/sdmx/2009/dimension#refArea",
+    namespace = "https://dd.eionet.europa.eu/vocabulary/eurostat/geo/$1"
+  )
+)
 
-```{r orangedf}
-data(orange_df)
-orange_df[1, ]
 ```
 
-Instead of placing the relevant measurement of an observed flower into the intersection of columns and rows, in the triple format we put them next to each other: 
-
-- the first tree's age is 118 days.
-- the first tree's circumference is 30 mm.
-
-
-```{r triples}
-dataset_to_triples(orange_df[1:2, ])[1:3, ]
+```{r triplesdf}
+triples_df <- dataset_to_triples(small_geo)
+knitr::kable(triples_df)
 ```
-We describe the `dataset_df` datasets in such triplets, where each triplet is a semantic statement: it connects a single observation unit with a single measurement.
 
-##  Enhanced Interoperability
-RDF uses globally unique identifiers (URIs) for resources, ensuring that different datasets can reference the same entities unambiguously. This allows seamless data integration and querying across repositories, even if the datasets come from diverse domains.
+This produces triples like:
 
-Our `defined` class supports this enhanced interoperability. In the example below, an application can look up that the numeric values in your table conform the statistical definition of GDP, and they are expressed in millions of dollars; meaning that you have to multiply them by 1000 if you want to join them with different data expressed in thousands of dollars. 
+```{r createntriples}
+ntriples <- dataset_to_triples(small_geo, format = "nt")
+cat(ntriples, sep = "\n")
+```
 
-```{r defined}
[…]
