### 2025-07-22 fa34c0a
rewrite dataset_df vignette
