+++ vignettes/dataset_df.Rmd
@@ -1,130 +1,126 @@
 ---
-title: "Why Semantics Matter for R Data Frames"
+title: "Create Datasets that are Easy to Share Exchange and Extend"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Why Semantics Matter for R Data Frames}
+  %\VignetteIndexEntry{Create Datasets that are Easy to Share Exchange and Extend}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setupvignette, include = FALSE}
+```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 ```
 
+The `dataset_df()` function helps create semantically rich datasets that are easier to:
+- understand by humans,
+- validate and process by machines,
+- and share across tools, teams, and domains.
+
+This vignette demonstrates how to construct a semantically enriched dataset using a subset of the  *GDP and main aggregates - international data cooperation annual data* dataset created by Eurostat (DOI: <https://doi.org/10.2908/NAIDA_10_GDP>).
+
+While the raw dataset is valid, it can be hard to interpret without external documentation. For example:
+- Countries are encoded in the `geo` variable.
+- Reporting frequency (e.g., `A` for annual) is stored in `freq`
+
+```{r printgdp}
+print(gdp)
+```
+
+## Getting Started
+
 ```{r setup}
 library(dataset)
+data("gdp")
 ```
 
-R users love `data.frame`s and `tibble`s for tidy, rectangular data. But tidy data isn't always **meaningful data**. What does a column labelled `gdp` actually represent? Euros? Millions? Per capita? Current prices? Constant 2010 prices? These questions matter—especially in statistics, open data publishing, and knowledge graph integration.
 
-The `dataset_df` class extends the familiar `data.frame` structure with lightweight, semantically meaningful metadata. It's built for:
+The `dataset_df()` constructor enables two levels of semantic annotation for a tibble `tbl_df` object (created from a data.frame or tibble):
+- Per-variable metadata: label, unit, definition, namespace
+- Whole-dataset metadata: title, author, license, etc.
 
--   **Tidyverse lovers** who want better documentation and safer analysis
+Let’s create a semantically enriched subset:
 
--   **Open science workflows** that need interoperable metadata
-
--   **Semantic web users** who want to export structured RDF data from R
-
-`dataset_df` helps you preserve the meaning of variables, units, identifiers, and dataset-level context.
-
-## From Tidy to Meaningful: An Example
-
-Let's start with a basic data frame and upgrade it to a `dataset_df` with semantically enriched columns using `defined()`:
-
-```{r smallcountries}
+```{r createdataasetdf}
 small_country_dataset <- dataset_df(
-  country_name = defined(c("AD", "LI"),
+  geo = defined(
+    gdp$geo,
     label = "Country name",
-    concept = "http://data.europa.eu/bna/c_6c2bb82d",
-    namespace = "https://www.geonames.org/countries/$1/"
+    definition = "http://purl.org/linked-data/sdmx/2009/dimension#refArea", 
+    namespace = "https://dd.eionet.europa.eu/vocabulary/eurostat/geo/$1"), 
+  year = defined(
+    gdp$year, 
+    label = "Reference Period (Year)", 
+    definition = "http://purl.org/linked-data/sdmx/2009/dimension#refPeriod"),
+  gdp = defined(
+    gdp$gdp, 
+    label = "Gross Domestic Product", 
+    unit = "CP_MEUR", 
+    definition = "http://data.europa.eu/83i/aa/GDP"),
+  unit = defined(
+    gdp$unit,
+    label = "Unit of Measure", 
[…]
