+++ tests/testthat/test-dataset_to_triples.R
@@ -1,5 +1,143 @@
-test_that("dataset_to_triples works()", {
-  expect_equal(class(dataset_to_triples(iris_dataset)), "data.frame")
-  expect_equal(ncol(dataset_to_triples(iris_dataset)), 3)
-  expect_equal(nrow(dataset_to_triples(head(iris_dataset, 3))), dim(head(iris_dataset, 3))[1] * dim(head(iris_dataset, 3))[2])
+test_that("dataset_to_triples expands subject URIs correctly", {
+  data("gdp", package = "dataset")
+
+  small_country_dataset <- dataset_df(
+    geo = defined(
+      gdp$geo,
+      label = "Country name",
+      concept = "http://dd.eionet.europa.eu/vocabulary/eurostat/geo/",
+      namespace = "https://www.geonames.org/countries/$1/"
+    ),
+    year = defined(
+      gdp$year,
+      label = "Reference Period (Year)",
+      concept = "http://purl.org/linked-data/sdmx/2009/dimension#refPeriod"
+    ),
+    gdp = defined(
+      gdp$gdp,
+      label = "Gross Domestic Product",
+      unit = "CP_MEUR",
+      concept = "http://data.europa.eu/83i/aa/GDP"
+    ),
+    unit = gdp$unit,
+    freq = defined(
+      gdp$freq,
+      label = "Frequency",
+      concept = "http://purl.org/linked-data/sdmx/2009/code"
+    ),
+    dataset_bibentry = dublincore(
+      title = "Small Country Dataset",
+      creator = person("Jane", "Doe"),
+      publisher = "Example Inc.",
+      datasource = "https://doi.org/10.2908/NAIDA_10_GDP",
+      rights = "CC-BY",
+      coverage = "Andorra, Lichtenstein and San Marino"
+    )
+  )
+
+  triples <- dataset_to_triples(small_country_dataset)
+
+  # Extract actual s values
+  s_values <- unique(triples$s)
+  # Expect all s values to begin with the namespace prefix
+  expect_true(all(grepl("^http://example\\.com/dataset#", s_values)))
+
+ # Check that s is URI-expanded
+  expect_false(any(s_values %in% gdp$geo))  # none should be raw codes like "AD"
 })
+
+
+test_that("triples_column_generate generates correct RDF triples", {
+  s <- c("http://example.com/obs/1", "http://example.com/obs/2")
+
+  # Case 1: defined with namespace and concept
+  x <- defined(
+    c("AD", "FR"),
+    label = "Country",
+    concept = "http://example.com/prop/geo",
+    namespace = "https://eurostat.europa.eu/geo/$1"
+  )
+  triples <- triples_column_generate(s, x, "geo")
+
+  expect_equal(triples$p[1], "http://example.com/prop/geo")
+  expect_equal(triples$o[2], "https://eurostat.europa.eu/geo/FR")
+
+  # Case 2: numeric without namespace
+  x <- c(1.23, 4.56)
+  triples <- triples_column_generate(s, x, "value")
+  expect_true(all(grepl('^".+"\\^\\^<xs:decimal>$', triples$o)))
+})
+
+test_that("dataset_to_triples generates valid N-Triples", {
+  data("gdp", package = "dataset")
+
+  small_country_dataset <- dataset_df(
+    geo = defined(
+      gdp$geo,
+      label = "Country name",
+      concept = "http://example.com/prop/geo",
+      namespace = "https://www.geonames.org/countries/$1/"
+    ),
+    year = defined(
+      gdp$year,
+      label = "Reference Period (Year)",
[…]
