+++ tests/testthat/test-describe.R
@@ -1,15 +1,26 @@
-test_that("describe() works", {
-  temporary_con <- tempfile()
-  expect_null(describe(orange_df, temporary_con)) # not returned
-  expect_equal(readLines(temporary_con)[1],
-               "<http://example.com/dataset_prov.nt> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.w3.org/ns/prov#Bundle> .")
-  expect_error(describe(list(a=1)),
-               regexp = "x most be a a dataset_df object")
-  expect_equal(describe(orange_df)[1],
-               "<http://example.com/dataset_prov.nt> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.w3.org/ns/prov#Bundle> .")
+test_that("describe() writes provenance and metadata to file", {
+  data("gdp", package = "dataset")
+
+  test_ds <- dataset_df(
+    rowid = defined(c("eg:1", "eg:2"), namespace = "http://example.com/dataset#"),
+    geo = defined(
+      gdp$geo[1:2],
+      label = "Country",
+      concept = "http://example.com/prop/geo",
+      namespace = "https://eionet.europa.eu/geo/$1"
+    ),
+    dataset_bibentry = dublincore(
+      title = "Example Dataset",
+      creator = person("John", "Doe")
+    )
+  )
+
+  tmpfile <- tempfile(fileext = ".nt")
+  describe(x=test_ds, con=tmpfile)
+
+  lines <- readLines(tmpfile)
+  expect_true(any(grepl("prov#", lines)))
+  expect_true(any(grepl("ttp://purl.org/dc/terms/title", lines)))
+  expect_true(length(lines) > 3)
 })
-
-
-
-
 
