+++ man/triples_to_ntriples.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dataset_to_triples.R
+\name{triples_to_ntriples}
+\alias{triples_to_ntriples}
+\title{Internal: Convert triple data.frame to N-Triples format}
+\usage{
+triples_to_ntriples(df)
+}
+\arguments{
+\item{df}{A data.frame with columns \code{s}, \code{p}, and \code{o}.}
+}
+\value{
+A character vector of N-Triple lines.
+}
+\description{
+Turns a data.frame with \code{s}, \code{p}, \code{o} columns into N-Triples strings.
+}
+\keyword{internal}
