+++ DESCRIPTION
@@ -33,7 +33,6 @@ Imports:
     ISOcodes,
     labelled,
     pillar,
-    rlang,
     tibble,
     utils,
     vctrs
