+++ man/describe.Rd
@@ -18,3 +18,20 @@ Invisibly returns \code{x}. Writes N-Triples to \code{con}.
 Writes provenance and Dublin Core metadata of a dataset to a
 file or connection in N-Triples format.
 }
+\examples{
+test_ds <- dataset_df(
+   rowid = defined(c("eg:1", "eg:2"),
+   namespace = "http://example.com/dataset#"),
+   geo = defined(
+     gdp$geo[1:2],
+     label = "Country",
+     concept = "http://example.com/prop/geo",
+     namespace = "https://eionet.europa.eu/geo/$1"
+   ),
+   dataset_bibentry = dublincore(
+      title = "Example Dataset",
+     creator = person("John", "Doe")
+    )
+  )
+describe(test_ds, con=tempfile())
+}
