+++ vignettes/introduction.Rmd
@@ -1,0 +1,55 @@
+---
+title: "An introduction to the dataset package"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{An introduction to the dataset package}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(dataset)
+```
+
+In a series of vignette tutorials, we will create a small dataset about the Gross Domestic Product (GDP) of three small countries: Andorra, Liechtenstein, and San Marino.
+
+## Remain Tidy
+
+Most data scientists using R have encountered the concept of a tidy dataset and the tidyverse, which establishes clear rules about what constitutes observations and measured variables within a dataset. Messy data is difficult or impossible to analyse and often leads to misunderstandings. Our package aims to help maintain tidy datasets, even when they are subsetted, sliced, or joined with other datasets.
+
+The `dataset` package seeks to make better use of base R’s attribute metadata system and enrich vectors and data frames (tibbles) so that, even after various transformations, misunderstandings and logical errors can be avoided.
+
+We build a small dataset with four years of GDP data for three countries. Adding further data—such as new countries or additional time periods—requires a strict definition of the unit of measure (e.g., is it in millions or thousands of euros, euros or dollars?). It must also be clear whether we are discussing nominal or real GDP, at market prices or PPP-adjusted. The vignette on using the defined S3 vector class (an extension of `labelled::labelled` used in tidyverse) demonstrates how to add unambiguous concept definitions and codebook or unit-of-measure namespaces to numerical, categorical, or character vectors. Retaining such metadata is critically important to avoid accidental, syntactically correct but semantically incorrect concatenation of GDP values measured in euros and market prices with PPP-adjusted dollar figures. In terms of statistical data and its exchange, this allows for correct definitions of measured variables, their metadata attributes, and potential dimensional columns that enable semantically meaningful subsetting ("slicing") of the dataset.
+
+Similarly, datasets often carry descriptive metadata—known as attributes in statistics—that provide information about the table as a whole.
+
+When we add new data to the dataset—for instance, about the Faroe Islands—we must not only ensure the same codebooks, units of measure, and data collection frequencies are used, but also that the provenance is documented. Recording the data sources, who performed the data join, and what procedures were followed is helpful for future use or review. This is what we achieve with a modified tibble: the `dataset_df`.
+
+Base R provides a logical place to store metadata at the data.frame level—i.e., attributes—but this is seldom used and lacks widely adopted practices or helper functions. While many packages aim to facilitate the publication of such metadata (for example, `dataspice` from rOpenSci), these tools typically detach the metadata from the R object. We aimed to create a solution that retains the metadata within the R object (saved as an `.rds` or `.rda` file). Our helper functions were designed to incorporate two standard metadata schemes used in data publication and exchange: **DCTERMS**, an international library standard, and **DataCite**, a standard commonly used by scientific data repositories and open data applications. We created functions that ensure the addition of DCTERMS and DataCite metadata as attributes to an R object.
+
+Statistical data and metadata exchanges have long employed the Resource Description Framework (RDF) to ensure data and metadata remain connected. While official data providers often use RDF, the linkage between R and RDF is only partially supported by the excellent rOpenSci interface to RDFLib. The `rdflib` package supports import and export in all standard RDF serialisations (file formats), but it does not offer helper functions to add the necessary RDF metadata annotations to native R datasets—whether in `data.frame`, `tibble`, `data.table`, or `tsibble` formats.
+
[…]
