### 2025-07-23 1e77c64
Update vignettes
