+++ vignettes/introduction.Rmd
@@ -12,13 +12,20 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-```
 
-```{r setup}
-library(dataset)
+
+if (!requireNamespace("rdflib", quietly = TRUE)) {
+  stop("Please install 'rdflib' to run this vignette.")
+}
 ```
 
 In a series of vignette tutorials, we will create a small dataset about the Gross Domestic Product (GDP) of three small countries: Andorra, Liechtenstein, and San Marino.
+
+```{r setup}
+library(dataset)
+data("gdp")
+print(gdp)
+```
 
 ## Remain Tidy
 
@@ -26,9 +33,65 @@ Most data scientists using R have encountered the concept of a tidy dataset and 
 
 The `dataset` package seeks to make better use of base R’s attribute metadata system and enrich vectors and data frames (tibbles) so that, even after various transformations, misunderstandings and logical errors can be avoided.
 
-We build a small dataset with four years of GDP data for three countries. Adding further data—such as new countries or additional time periods—requires a strict definition of the unit of measure (e.g., is it in millions or thousands of euros, euros or dollars?). It must also be clear whether we are discussing nominal or real GDP, at market prices or PPP-adjusted. The vignette on using the defined S3 vector class (an extension of `labelled::labelled` used in tidyverse) demonstrates how to add unambiguous concept definitions and codebook or unit-of-measure namespaces to numerical, categorical, or character vectors. Retaining such metadata is critically important to avoid accidental, syntactically correct but semantically incorrect concatenation of GDP values measured in euros and market prices with PPP-adjusted dollar figures. In terms of statistical data and its exchange, this allows for correct definitions of measured variables, their metadata attributes, and potential dimensional columns that enable semantically meaningful subsetting ("slicing") of the dataset.
+We build a small dataset with four years of GDP data for three countries. Adding further data—such as new countries or additional time periods—requires a strict concept of the unit of measure (e.g., is it in millions or thousands of euros, euros or dollars?). It must also be clear whether we are discussing nominal or real GDP, at market prices or PPP-adjusted. 
+
+### defined: unambigous vectors
+The vignette on using the defined S3 vector class (an extension of `labelled::labelled` used in tidyverse) demonstrates how to add unambiguous concept concepts and codebook or unit-of-measure namespaces to numerical, categorical, or character vectors. Retaining such metadata is critically important to avoid accidental, syntactically correct but semantically incorrect concatenation of GDP values measured in euros and market prices with PPP-adjusted dollar figures. In terms of statistical data and its exchange, this allows for correct concepts of measured variables, their metadata attributes, and potential dimensional columns that enable semantically meaningful subsetting ("slicing") of the dataset.
+
+```{r gdp1}
+gdp_1 = defined(
+    gdp$gdp, 
+    label = "Gross Domestic Product", 
+    unit = "CP_MEUR", 
+    concept = "http://data.europa.eu/83i/aa/GDP")
+
+gdp_1
+```
+
+
+Further reading: [defined: Semantically Enriched Vectors](https://dataset.dataobservatory.eu/articles/defined.html)
+
+### dataset_df: publishable datasets
 
 Similarly, datasets often carry descriptive metadata—known as attributes in statistics—that provide information about the table as a whole.
+
+
+```{r createdataasetdf}
+small_country_dataset <- dataset_df(
+  geo = defined(
+    gdp$geo,
+    label = "Country name",
+    concept = "http://purl.org/linked-data/sdmx/2009/dimension#refArea", 
+    namespace = "https://dd.eionet.europa.eu/vocabulary/eurostat/geo/$1"), 
+  year = defined(
+    gdp$year, 
[…]
