+++ README.md
@@ -17,7 +17,7 @@ Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci
 [![devel-version](https://img.shields.io/badge/devel%20version-0.3.91-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 [![Codecov test
-coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset)
+coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset/)
 
 <!-- badges: end -->
 
@@ -37,20 +37,13 @@ of:
 - **Open Science** metadata practices,
 - **Library and archive metadata** conventions (Dublin Core, DataCite).
 
-## Motivation
-
 Many tools exist to help document, describe, or publish datasets in R,
 but most separate the metadata from the data itself. This separation
 increases the risk of losing metadata, misaligning it with the data, or
 making documentation hard to maintain. We wanted to create a tooling
 that enables the user from the birth of a dataset till it is potentially
-seriralised with the help of the
+serialised with the help of the
 [rdflib](https://CRAN.R-project.org/package=rdflib) package.
-
-The `dataset` package addresses this by storing all metadata directly in
-R object attributes. This preserves semantic information as data is
-transformed, combined, or exported, preventing the loss of vital
-documentation and improving reproducibility.
 
 ## Key Features
 
@@ -65,21 +58,49 @@ An extended version of `labelled()` vectors. Adds support for:
 
 ``` r
 library(dataset)
+data(gdp)
 ```
+
+The following small dataset contains the gross domestic product (GDP)
+data of three small countries:
 
 ``` r
-data(orange_df)
-print(orange_df$age)
-#> orange_df$age: The age of the tree
-#> Measured in days since 1968/12/31 
-#>  [1]  118  484  664 1004 1231 1372 1582  118  484  664 1004 1231 1372 1582  118
-#> [16]  484  664 1004 1231 1372 1582  118  484  664 1004 1231 1372 1582  118  484
-#> [31]  664 1004 1231 1372 1582
+print(gdp)
+#> # A tibble: 10 × 5
+#>    geo    year   gdp unit    freq 
+#>    <chr> <int> <dbl> <chr>   <chr>
+#>  1 AD     2020 2355. CP_MEUR A    
+#>  2 AD     2021 2594. CP_MEUR A    
+#>  3 AD     2022 2884. CP_MEUR A    
+#>  4 AD     2023 3120. CP_MEUR A    
+#>  5 LI     2020 5430. CP_MEUR A    
+#>  6 LI     2021 6424. CP_MEUR A    
+#>  7 LI     2022 6759. CP_MEUR A    
+#>  8 SM     2020 1265. CP_MEUR A    
+#>  9 SM     2021 1461. CP_MEUR A    
+#> 10 SM     2022 1612. CP_MEUR A
 ```
 
-This ensures that, for example, “GDP” is always associated with a
-precise concept and unit, avoiding ambiguity across analyses and
-publications. See [defined: Semantically Enriched
+The `defined` vector class, an extension of the `labelled::labelled`
+class, allows to add machine-readable concept definitions and namespaces
+to reveal the coding of the variables. This will allow `AD` to be
+resolved as <https://dd.eionet.europa.eu/vocabulary/eurostat/geo/AD>
+
+``` r
+geo <- defined(
[…]
