### 2025-07-25 8ec5297
0.3.92
Improvements in the vignettes.
