+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.91
+Version: 0.3.92
 Date: 2025-07-22
 Language: en-GB
 Authors@R: 

[…]
