### 2025-08-01 #681 @maurolepore
@antaldaniel, can you please help me check the boxed for the comments that required action? Please the section "Semantic comment-tags" in [my review-comment](https://github.com/ropensci/software-review/issues/681#issuecomment-2870424492).

From skimming the thread it seems you already addressed ml06 and ml25, but you would really help me you responded to each of the check-box comments, maybe summarizing the decisions we agreed.

I would expect an answer strcuctured like this:

- ml06: ...
- ml07: ...
- ml11: ...
- ml12: ...
- ml13: ...
- ml14: ...
- ml20: ...
- ml25: ...
- ml26: ...

Responses to any other comment is welcome but optional.

Sorry to bother you with such a minor formatting detail. It really helps me keep track of each item by running a search for "mlxx".
