### 2025-08-05 a64b560
Multiple subjects give a meaningful error
