### 2025-08-05 c3d7832
bibliography to RDF changes
