+++ man/expand_triples.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dataset_to_triples.R
+\name{expand_triples}
+\alias{expand_triples}
+\title{Internal: Expand multi-valued DC fields to RDF triples}
+\usage{
+expand_triples(dataset_id, predicate_uri, values)
+}
+\arguments{
+\item{dataset_id}{The subject URI}
+
+\item{predicate_uri}{The RDF predicate URI}
+
+\item{values}{A scalar, character vector, or list (e.g., person objects)}
+}
+\value{
+A character vector of RDF triples
+}
+\description{
+Converts scalar or vector fields into RDF triples.
+}
+\keyword{internal}
