### 2025-08-05 8964cf4
New tests for dublincore_to_triples()
