### 2025-08-05 d5a41f8
Rewrite documentation
