+++ vignettes/introduction.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "An introduction to the dataset package"
+title: "An Introduction to the dataset Package"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{An introduction to the dataset package}
+  %\VignetteIndexEntry{An Introduction to the dataset Package}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -13,132 +13,130 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-
 if (!requireNamespace("rdflib", quietly = TRUE)) {
   stop("Please install 'rdflib' to run this vignette.")
 }
 ```
 
-In a series of vignette tutorials, we will create a small dataset about the Gross Domestic Product (GDP) of three small countries: Andorra, Liechtenstein, and San Marino.
+## Overview
 
-```{r setup}
+The `dataset` package enriches R’s native data structures with machine-readable metadata. It allows variables and datasets to carry semantic definitions — such as URIs, labels, units, and provenance — which makes them suitable for long-term reuse, FAIR-compliant publishing, and integration into semantic web systems.
+
+Unlike most metadata packages that attach metadata after the fact, `dataset` follows a **semantic early-binding** approach: metadata is embedded as soon as the data is created.
+
+This vignette provides a high-level introduction. For details on key components, see the following:
+- `vignette("defined", package = "dataset")`: Semantic vectors with `defined()`
+- `vignette("dataset_df", package = "dataset")`: Structuring and metadata with `dataset_df()`
+- `vignette("rdf", package = "dataset")`: Exporting to RDF and Linked Data
+- `vignette("bibrecord", package = "dataset")`: Creating rich citation metadata using `bibrecord()`
+
+## Why Extend Tidy Data?
+
+Hadley Wickham (2014) defines tidy data with three principles:
+
+- Each variable forms a column  
+- Each observation forms a row  
+- Each observational unit forms a table  
+
+This structure is ideal for analysis, but lacks **semantic clarity**, particularly when an analyst is working in a realistic, but not ideal scenario with several datasets received from various internet services. For example, two datasets might both contain a column named `gdp`, but one might be in euros and the other in dollars. Without metadata, tools cannot detect this mismatch.
+
+The `dataset` package addresses this by allowing you to define variables explicitly, and to store dataset-level metadata within a tidy tibble.
+
+## Example: Defining Semantically Rich Vectors
+
+Semantically rich vectors are vectors in a data.frame that contain richer semantics than a simple column name; a long-form human-readable title; a machine- and human-readable variable definition; and if needed, an external resource that contains the codebook.
+
+```{r}
 library(dataset)
-data("gdp")
-print(gdp)
-```
 
-## Remain Tidy
+gdp <- defined(
+  c(2355, 2592, 2884),
+  label = "Gross Domestic Product",
+  unit = "CP_MEUR",
+  concept = "http://data.europa.eu/83i/aa/GDP"
+)
 
-Most data scientists using R have encountered the concept of a tidy dataset and the tidyverse, which establishes clear rules about what constitutes observations and measured variables within a dataset. Messy data is difficult or impossible to analyse and often leads to misunderstandings. Our package aims to help maintain tidy datasets, even when they are subsetted, sliced, or joined with other datasets.
+geo <- defined(
+  rep("AD", 3),
+  label = "Geopolitical Entity",
+  concept = "http://dd.eionet.europa.eu/vocabulary/eurostat/geo/",
+  namespace = "https://www.geonames.org/countries/$1/"
+)
 
[…]
