+++ vignettes/design.Rmd
@@ -1,0 +1,154 @@
+---
+title: "dataset: Semantically Enriched, Standards-Aligned Datasets in R"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{dataset: Semantically Enriched, Standards-Aligned Datasets in R}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Abstract 
+
+```{r setup}
+library(dataset)
+```
+
+> "A dataset is an identifiable collection of data available for access or download in one or more formats." — ISO/IEC 20546
+
+The dataset package enriches R's data objects with machine-readable metadata. It embeds semantic definitions and provenance at both the variable and dataset level, offering users interoperable, standards-aligned outputs. It follows a semantic early-binding design: metadata is attached during dataset creation, not retrofitted later. This ensures that context and meaning are preserved from exploration to publication, and some documentation can be automated.
+
+This article outlines the design philosophy behind the dataset package, including its theoretical underpinnings, structure, relation to other R tools, and example workflows. It is intended as a long-form complement to the package vignettes.
+
+## Introduction and Motivation
+
+>"The principles of tidy data provide a standard way to organise data values within a dataset." — Wickham (2014)
+
+Wickham's article refers to a particular manifestation of a dataset within R: a data.frame or tibble object. However, R users often conflate this tangible object that lives in their computer memory and in the R ecosystem with the broader concept of a dataset. In reality, the definition of a dataset refers to its abstract structure and semantics—elements that can be realized in many forms across databases, APIs, or file systems.
+
+The original tidy workflow was conceived for interactive sessions where analysts had full context of their data's source and meaning, and provided a pragmatic approach to work with several datasets: keep the original variable names in the identifier row, or place dimensions and attributes before measured variables.  This design is effective when the analyst is the sole user, but insufficient when data is shared, reused institutionally, or published for open use. In such cases, more semantics is needed: not just tidy structure, but explicit definitions of units, codes, contributors, and provenance. Also, we need to be at least aware that institutional statistical providers use a somewhat more complicated format that resembles the tidy data, but with higher dimensionality: the datacube model.
+
+Tidy data assumes that well-named variables and consistent columnar structure will carry enough meaning for analysis. But when working with datasets from different sources—such as Eurostat and OECD—conflicts emerge. For instance, a variable named geo might encode countries using different standards (ISO codes vs Eurostat codes), and economic indicators like GDP may appear in different currencies or base years. These differences may not be apparent until after erroneous joins or analyses have occurred.
+
+To bridge this semantic gap, the dataset package integrates machine-readable metadata directly into R data objects, preserving this information from creation to export.
+
+The original tidy workflow was conceived for interactive sessions where analysts had full context of their data's source and meaning. This design is effective when the analyst is the sole user, but insufficient when data is shared, reused institutionally, or published for open use. In such cases, more semantics is needed: not just tidy structure, but explicit definitions of units, codes, contributors, and provenance.
+
+Tidy data principles (Wickham, 2014) dramatically improved the consistency of R's data wrangling workflows:
+-	Each variable forms a column
+-	Each observation forms a row
+-	Each observational unit forms a table.
+
+However, this tidy structure often lacks semantics. Consider:
+
+```{r}
+data.frame(
+  geo = c("LI", "SM"),
+  CPI = c("0.8", "0.9"),
+  GNI = c("8976", "9672")
+)
+```
+
[…]
