### 2025-08-05 cc7cea0
Bug fixes and documentation improvements since 0.3.92
