+++ R/provenance.R
@@ -11,13 +11,14 @@
 #'
 #' ## add a statement:
 #'
-#'  provenance(orange_df) <- n_triple(
+#' provenance(orange_df) <- n_triple(
 #'   "https://doi.org/10.5281/zenodo.10396807",
 #'   "http://www.w3.org/ns/prov#wasInformedBy",
-#'   "http://example.com/source#1")
+#'   "http://example.com/source#1"
+#' )
 #' @export
 provenance <- function(x) {
-  if(!is.dataset_df(x)) {
+  if (!is.dataset_df(x)) {
     stop("provenance(x): x must be a dataset_df object with standardised provenance metadata.")
   }
   attr(x, "prov")
@@ -25,8 +26,7 @@ provenance <- function(x) {
 
 #' @rdname provenance
 #' @export
-`provenance<-` <- function(x,  value) {
-
+`provenance<-` <- function(x, value) {
   if (!is.dataset_df(x)) {
     stop("provenance(x)<- : x must be a dataset object created with dataset_df() or as_dataset_df().")
   }
@@ -47,7 +47,7 @@ default_provenance <- function(dataset_id = "http://example.com/dataset#",
 
   agent_triples <- c()
   if (!is.null(author)) agent_triples <- c(agent_triples, prov_author(author))
-  if (!is.null(dtm))    agent_triples <- c(agent_triples, prov_author(dtm))
+  if (!is.null(dtm)) agent_triples <- c(agent_triples, prov_author(dtm))
 
   if (is.null(generated_at_time)) {
     generated_at_time <- Sys.time()

[…]
