+++ tests/testthat/test-defined.R
@@ -1,4 +1,4 @@
-test_that("labelled_defined() works", {
+test_that("defined() constructor creates expeced vectors", {
   z <- defined(c(1, 1, 1, 0, 0, 0),
     label = "",
     labels = c("F" = 0, "M" = 1, "_N" = 99),
@@ -50,7 +50,7 @@ test_that("labelled_defined() works", {
   )
 })
 
-test_that("labelled_defined() works", {
+test_that("defined() attributes can be retrieved", {
   sepal_length <- defined(iris$Sepal.Length,
     labels = NULL,
     label = "Sepal length",
@@ -186,12 +186,14 @@ test_that("as.list.haven_labelled_defined preserves metadata", {
 test_that("labelled_defined() throws error", {
   expect_error(var_unit(sepal_length) <- c("cm", "mm"))
   expect_error(var_unit(sepal_length) <- 1)
+
+  # formal argument "concept" matched by multiple actual arguments
   expect_error(defined(
     x = iris$Species,
     label = "Taxon name within the Iris genus",
     concept = 1,
     concept = "Iris"
-  ))
+  ), regexp = "matched by multiple arguments")
   expect_error(defined(
     x = iris$Species,
     label = "Taxon name within the Iris genus",
@@ -266,8 +268,9 @@ test_that("new_labelled_defined() throws errors", {
 
 test_that("iris_dataset() prints", {
   expect_output(str(iris_dataset),
-                "https://doi.org/10.5281/zenodo.10396807",
-                ignore.case = FALSE)
+    "https://doi.org/10.5281/zenodo.10396807",
+    ignore.case = FALSE
+  )
   expect_output(print(iris_dataset), "Iris Dataset.", ignore.case = FALSE)
 })
 
@@ -302,8 +305,10 @@ test_that("c() works", {
 
 test_that("summary.haven_labelled_defined() works ", {
   sepal_length <- iris_dataset$Sepal.Length
-  expect_output(summary(sepal_length),
-                "Length of the sepal in cm \\(centimeter\\)")
+  expect_output(
+    summary(sepal_length),
+    "Length of the sepal in cm \\(centimeter\\)"
+  )
   expect_equal(names(summary(sepal_length))[1], "Min.")
 })
 
@@ -328,13 +333,17 @@ test_that("as_numeric() returns underlying numeric vector", {
 
 test_that("as_character() returns underlying character vector", {
   fruits <- defined(c("apple", "avocado", "kiwi"),
-                    label = "Fruit", unit = "kg")
-  expect_equal(as_character(fruits, preserve_attributes = FALSE),
-               c("apple", "avocado", "kiwi"))
+    label = "Fruit", unit = "kg"
+  )
+  expect_equal(
+    as_character(fruits, preserve_attributes = FALSE),
+    c("apple", "avocado", "kiwi")
+  )
   expect_equal(attr(as_character(fruits, TRUE), "unit"), "kg")
   expect_type(as_character(fruits), "character")
   expect_error(as_numeric(fruits),
-               regexp = "underlying data is not numeric")
+    regexp = "underlying data is not numeric"
+  )
 })
 
 test_that("as_factor() works with defined vector", {
@@ -351,21 +360,35 @@ test_that("as_factor() works with defined vector", {
 
 
 test_that("c.haven_labelled_defined() works ", {
-  a <- defined(1:3, label = "testlabel", unit = "meter",
-               concept = "testdef", namespace = "http://example.com")
-  b <- defined(4:6, label = "testlabel", unit = "meter",
-               concept = "testdef", namespace = "http://example.com")
-  ab <- defined(1:6, label = "testlabel", unit = "meter",
-                concept = "testdef", namespace = "http://example.com")
[…]
