+++ tests/testthat/test-bibrecord.R
@@ -63,4 +63,3 @@ test_that("bibrecord() works without contributor field", {
   expect_null(attr(rec, "contributor"))
   expect_output(print(rec), "Doe J")
 })
-
