+++ R/get_bibentry.R
@@ -1,17 +1,17 @@
 #' @title Get/set the Bibentry of the object.
-#' @description The \code{\link{dataset_df}} objects contain among their
+#' @description The [dataset_df()] objects contain among their
 #' attributes bibliographic entries which are stored in a
 #' \code{\link[utils:bibentry]{utils::bibentry}} object. Upon creation, these
 #' entries are filled with default values when applicable. \cr \cr To retrieve
 #' the bibentry of a dataset_df object, use \code{get_bibentry}.\cr \cr To
-#' create a new bibentry, use the \code{\link{datacite}} function for an
+#' create a new bibentry, use the[datacite()] function for an
 #' interface and default values according to the DataCite standard, or the
-#' \code{\link{dublincore}} function for the more general Dublin Core
+#' [dublincore()] function for the more general Dublin Core
 #' standard.\cr \cr To change or an entire new bibliographic entry to a
 #' dataset_df object (or any data.frame-like object), use the
 #' \code{`set_bibentry<-`} function (see examples.) For more details, please
 #' check the \code{vignette("bibentry", package="dataset")} vignette.
-#' @param dataset A dataset created with \code{\link{dataset_df}}.
+#' @param dataset A dataset created with [dataset_df()].
 #' @param value A \code{\link[utils:bibentry]{utils::bibentry}} object, or a
 #'   newly initialised bibentry object with DataCite default values for
 #'   unassigned entries.
@@ -21,8 +21,8 @@
 #'   \code{\link[utils]{bibentry}} object of \code{x} from its attributes; the
 #'   \code{`set_bibentry<-`} assignment function sets this attribute to
 #'   \code{value} and invisibly returns \code{x} with the changed attributes. To
-#'   set well-formatted input \code{value}, refer to \code{\link{datacite}} or
-#'   \code{\link{dublincore}} (see Details.)
+#'   set well-formatted input \code{value}, refer to[datacite()] or
+#'   [dublincore()] (see Details.)
 #' @family bibentry functions
 #' @examples
 #' # Get the bibentry of a dataset_df object:
