+++ R/provenance.R
@@ -1,9 +1,9 @@
 #' @title Get or update provenance information
 #' @description Add or update information about the history (provenance) of the dataset.
-#' @param x A dataset created with \code{\link{dataset_df}}.
-#' @param value Use \code{\link{n_triples}} to add further statement values.
+#' @param x A dataset created with [dataset_df()].
+#' @param value Use [n_triples()] to add further statement values.
 #' @return \code{provenance(x)} returns the provenance attributes created by
-#' \code{\link{n_triples}} as a text; \code{provenance(x)<-value} adds the new
+#' [n_triples()] as a text; \code{provenance(x)<-value} adds the new
 #' provenance attributes and returns \code{x} invisibly.
 #' @importFrom utils citation
 #' @examples

[…]
