### 2025-08-05 edf6b44
markdown-style links in roxygen
