+++ man/description.Rd
@@ -10,7 +10,7 @@ description(x)
 description(x, overwrite = FALSE) <- value
 }
 \arguments{
-\item{x}{A dataset object created with \code{dataset::\link{dataset_df}} or
+\item{x}{A dataset object created with \code{\link[=dataset_df]{dataset_df()}} or
 \code{dataset::\link{as_dataset_df}}.}
 
 \item{overwrite}{If the \code{Description} attribute should be overwritten.
