+++ man/rights.Rd
@@ -10,7 +10,7 @@ rights(x)
 rights(x, overwrite = FALSE) <- value
 }
 \arguments{
-\item{x}{A semantically rich data frame object created by  \code{dataset::\link{dataset_df}} or
+\item{x}{A semantically rich data frame object created by  \code{\link[=dataset_df]{dataset_df()}} or
 \code{dataset::\link{as_dataset_df}}.}
 
 \item{overwrite}{If the \code{Rights} attribute should be overwritten. In case it
