+++ man/geolocation.Rd
@@ -11,7 +11,7 @@ geolocation(x, overwrite = TRUE) <- value
 }
 \arguments{
 \item{x}{A semantically rich data frame object created by
-\code{dataset::\link{dataset_df}} or \code{dataset::\link{as_dataset_df}}.}
+\code{\link[=dataset_df]{dataset_df()}} or \code{dataset::\link{as_dataset_df}}.}
 
 \item{overwrite}{If the attributes should be overwritten. In case it is set
 to \code{FALSE}, it gives a message with the current \code{Geolocation}
