+++ man/publisher.Rd
@@ -10,7 +10,7 @@ publisher(x)
 publisher(x, overwrite = TRUE) <- value
 }
 \arguments{
-\item{x}{A dataset object created with \code{dataset::\link{dataset_df}} or
+\item{x}{A dataset object created with \code{\link[=dataset_df]{dataset_df()}} or
 \code{dataset::\link{as_dataset_df}}.}
 
 \item{overwrite}{If the attributes should be overwritten. In case it is set
