### 2025-08-05 25a84b1
Add vignette browser
