+++ R/description.R
@@ -49,11 +49,14 @@ description <- function(x) {
     value <- ":unas"
   }
 
-  if (overwrite || existing_description %in% c("", ":unas", ":tba")) {
+  if (overwrite ||
+      length(existing_description) == 0 ||
+      existing_description %in% c("", ":unas", ":tba")) {
+
     ds_bibentry$description <- as.character(value)
     attr(x, "dataset_bibentry") <- ds_bibentry
   } else {
-    warning("The dataset has already a description: ", existing_description, ".")
+    warning("The dataset already has a description: ", existing_description, ".")
   }
   invisible(x)
 }
