+++ R/description.R
@@ -1,31 +1,30 @@
-#' @title Get/set the Description of the object.
-#' @description Get/set the optional \code{Description} property as an attribute
-#'   to an R object.
-#' @details The \code{Description} is recommended for discovery in DataCite. All
-#'   additional information that does not fit in any of the other categories.
-#'   May be used for technical information. A free text. Similar to
-#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/}{dct:description}.
-#' @param x A dataset object created with [dataset_df()] or
-#'   \code{dataset::\link{as_dataset_df}}.
-#' @param value The \code{Description} as a character set.
-#' @param overwrite If the \code{Description} attribute should be overwritten.
-#'   In case it is set to \code{FALSE}, it gives a message with the current
-#'   \code{Description} property instead of overwriting it. Defaults to
-#'   \code{FALSE}, when it gives a warning at an accidental overwrite attempt.
-#' @return The \code{Description} attribute as a character of length 1 is added
-#'   to \code{x}.
+#' @title Get or set the dataset Description
+#'
+#' @description Get or set the optional `Description` property as an attribute
+#' on a dataset object.
+#'
+#' @details The `Description` is recommended for discovery in DataCite. It
+#' captures additional information that does not fit other metadata categories —
+#' such as technical notes or dataset usage. It is a free-text field. See
+#' [dct:description](https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/).
+#'
+#' @param x A dataset object created with [`dataset_df()`] or
+#'   [`as_dataset_df()`].
+#' @param value The new description, as a character string.
+#' @param overwrite Logical. If `TRUE`, will overwrite any existing description.
+#'   If `FALSE` (default), will warn and keep the existing description.
+#'
+#' @return The `Description` attribute as a character vector of length 1.
+#'
 #' @examples
 #' description(orange_df)
-#' description(
-#'   orange_df,
-#'   overwrite = TRUE
-#' ) <- "The 'orange' dataset has 35 rows and 3 columns
-#'                           of records of the growth of orange trees."
+#' description(orange_df, overwrite = TRUE) <- "This dataset records orange tree growth."
 #' description(orange_df)
+#'
 #' @family Reference metadata functions
 #' @export
 
-#' @export
+
 description <- function(x) {
   assert_that(is.dataset_df(x),
     msg = "description(x): x must be a dataset object created with dataset_df() or as_dataset_df()."
