### 2025-08-06 35c6210
Add design vignette
