+++ vignettes/design.Rmd
@@ -28,26 +28,28 @@ This article outlines the design philosophy behind the dataset package, includin
 
 ## Introduction and Motivation
 
->"The principles of tidy data provide a standard way to organise data values within a dataset." — Wickham (2014)
+> "The principles of tidy data provide a standard way to organise data values within a dataset."  
+> — Wickham (2014)
 
-Wickham's article refers to a particular manifestation of a dataset within R: a data.frame or tibble object. However, R users often conflate this tangible object that lives in their computer memory and in the R ecosystem with the broader concept of a dataset. In reality, the definition of a dataset refers to its abstract structure and semantics—elements that can be realized in many forms across databases, APIs, or file systems.
+Wickham's article refers to a particular manifestation of a dataset within R: a `data.frame` or `tibble` object. However, R users often conflate this tangible object — which exists in computer memory and within the R environment — with the broader concept of a dataset. In reality, a dataset refers to its abstract structure and semantics, which can be realized in many forms across databases, APIs, flat files, or publication platforms.
 
-The original tidy workflow was conceived for interactive sessions where analysts had full context of their data's source and meaning, and provided a pragmatic approach to work with several datasets: keep the original variable names in the identifier row, or place dimensions and attributes before measured variables.  This design is effective when the analyst is the sole user, but insufficient when data is shared, reused institutionally, or published for open use. In such cases, more semantics is needed: not just tidy structure, but explicit definitions of units, codes, contributors, and provenance. Also, we need to be at least aware that institutional statistical providers use a somewhat more complicated format that resembles the tidy data, but with higher dimensionality: the datacube model.
+A dataset, when preserved in a less ephemeral format such as a file or dissemination, becomes a type of digital _resource_ — a broader class that also includes data papers, software packages, and other scholarly outputs. Resources are typically described using metadata and may be serialized using the Resource Description Framework (RDF), which ensures representation that is independent of language, operating system, or schema. This kind of serialization facilitates resource discovery and reuse.
 
-Tidy data assumes that well-named variables and consistent columnar structure will carry enough meaning for analysis. But when working with datasets from different sources—such as Eurostat and OECD—conflicts emerge. For instance, a variable named geo might encode countries using different standards (ISO codes vs Eurostat codes), and economic indicators like GDP may appear in different currencies or base years. These differences may not be apparent until after erroneous joins or analyses have occurred.
+In the case of datasets, which are often subsetted or sliced into smaller components, RDF enables description at the level of **elementary statements** — each corresponding to the intersection of a variable (column) and an observation (row). This allows not just metadata about the dataset as a whole, but also fine-grained, machine-readable annotations at the cell level.
 
-To bridge this semantic gap, the dataset package integrates machine-readable metadata directly into R data objects, preserving this information from creation to export.
+The original tidy workflow was designed for interactive sessions where analysts had full context of their data’s origin and structure. It provided a pragmatic convention: place identifiers and dimensions first, followed by measured variables. 
 
-The original tidy workflow was conceived for interactive sessions where analysts had full context of their data's source and meaning. This design is effective when the analyst is the sole user, but insufficient when data is shared, reused institutionally, or published for open use. In such cases, more semantics is needed: not just tidy structure, but explicit definitions of units, codes, contributors, and provenance.
[…]
