### 2025-08-06 95befa6
rOpenSci style description and test formatting
