+++ man/n_triples.Rd
@@ -7,19 +7,20 @@
 n_triples(triples)
 }
 \arguments{
-\item{triples}{Concatenated N-Triples created with \code{\link[=n_triple]{n_triple()}}.}
+\item{triples}{A character vector of concatenated N-Triples, created with
+\code{\link[=n_triple]{n_triple()}}.}
 }
 \value{
-A character vector containing unique N-Triple strings.
+A character vector of unique N-Triple strings.
 }
 \description{
-Create triple statements to annotate your dataset with standard,
+Create RDF triple statements to annotate your dataset with standard,
 interoperable metadata.
 }
 \details{
-N-Triples is an easy to parse line-based subset of Turtle to serialize
-RDF. See
-\href{https://www.w3.org/TR/rdf12-n-triples/}{RDF 1.2 N-Triples. A line-based syntax for an RDF graph}.
+N-Triples is a line-based serialization format for RDF. It is easy to parse
+and widely supported. For details, see the
+\href{https://www.w3.org/TR/rdf12-n-triples/}{W3C RDF 1.2 N-Triples specification}.
 }
 \examples{
 triple_1 <- n_triple(
@@ -27,10 +28,13 @@ triple_1 <- n_triple(
   "http://www.w3.org/2000/01/rdf-schema#label",
   "That Seventies Show"
 )
+
 triple_2 <- n_triple(
   "http://example.org/show/218",
   "http://example.org/show/localName",
   '"Cette Série des Années Septante"@fr-be'
 )
+
 n_triples(c(triple_1, triple_2, triple_1))
+
 }
