### 2025-08-06 176b4d9
Re-document dataset_df and create a new test suite without iris and new edge cases
