+++ tests/testthat/test-dataset_df.R
@@ -1,246 +1,153 @@
-test_that("dataset_df() works", {
-  expect_s3_class(dataset_df(mtcars), "dataset_df")
-  expect_true(is.dataset_df(dataset_df(mtcars)))
+test_that("dataset_df() creates and validates basic object", {
+  df <- dataset_df(mtcars)
+  expect_s3_class(df, "dataset_df")
+  expect_true(is.dataset_df(df))
   expect_false(is.dataset_df(mtcars))
-  expect_equal(get_bibentry(dataset_df(mtcars))$author, person("Author", "Unknown"))
-  expect_equal(
-    dataset_df(mtcars, identifier = c(mt = "http:/mtcars.com/dataset#"))$rowid,
-    defined(paste0("mt:", seq_along(mtcars$mpg)),
-      namespace = c(mt = "http:/mtcars.com/dataset#")
-    )
-  )
-  my_dataset <- dataset_df(
-    country_name = defined(
-      c("AD", "LI"),
-      concept = "http://data.europa.eu/bna/c_6c2bb82d",
-      namespace = "https://www.geonames.org/countries/$1/"
-    ),
-    gdp = defined(
-      c(3897, 7365),
-      label = "Gross Domestic Product",
-      unit = "million dollars",
-      concept = "http://data.europa.eu/83i/aa/GDP"
-    )
-  )
-  expect_equal(var_label(my_dataset$gdp), "Gross Domestic Product")
+
+  expect_equal(get_bibentry(df)$author, person("Author", "Unknown"))
 })
 
-test_that("dataset_df() works", {
-  orange_bibentry <- dublincore(
+test_that("dataset_df() respects identifier for rowid creation", {
+  df <- dataset_df(mtcars,
+                   identifier = c(mt = "http://mtcars.com/dataset#"))
+  expect_s3_class(df$rowid, "defined")
+  expect_true(all(grepl("^mt:", df$rowid)))
+})
+
+class(df$rowid)
+
+df
+test_that("dataset_df() stores variable metadata", {
+  df <- dataset_df(
+    foo = defined(1:3, label = "Example Var", unit = "count")
+  )
+  expect_equal(var_label(df$foo), "Example Var")
+  expect_equal(var_unit(df$foo), "count")
+})
+
+test_that("dataset_df() creates with full metadata", {
+  bib <- dublincore(
     title = "Growth of Orange Trees",
-    creator = personList(
-      person(
-        given = "N.R.",
-        family = "Draper",
-        role = "cre",
-        comment = c(VIAF = "http://viaf.org/viaf/84585260")
+    creator = person("N.R.", "Draper", role = "cre"),
+    publisher = "Wiley",
+    identifier = "https://doi.org/10.5281/zenodo.10396807",
+    dataset_date = 1998,
+    description = "35 rows of orange growth"
+  )
+
+  df <- dataset_df(
+    tree = defined(Orange$Tree, label = "Tree ID"),
+    age = defined(Orange$age, unit = "days"),
+    circumference = defined(Orange$circumference, label = "Size"),
+    dataset_bibentry = bib
+  )
+
+  expect_equal(dataset_title(df), "Growth of Orange Trees")
+  expect_equal(creator(df)[[1]]$family, "Draper")
+})
+
+test_that("as_dataset_df() converts from data.frame", {
+  df <- as_dataset_df(Orange)
+  expect_s3_class(df, "dataset_df")
+})
+
+test_that("new_dataset() validates input", {
+  expect_error(new_dataset(42, identifier = c(x = "http://example.com")))
+  expect_s3_class(
+    new_dataset(Orange, identifier = c(x = "http://example.com")),
+    "dataset_df"
+  )
+})
+
+test_that("subsetting preserves class and metadata", {
[…]

[…]
