### 2025-08-06 6b6a800
Character coercion: new documentation, edge cases, tests
