+++ man/as_character.Rd
@@ -1,12 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defined.R
-\name{as.character}
-\alias{as.character}
-\alias{as.character.haven_labelled_defined}
+% Please edit documentation in R/as_character.R, R/defined.R
+\name{as_character}
 \alias{as_character}
 \alias{as_character.haven_labelled_defined}
-\title{Coerce to character vector}
+\alias{as.character.haven_labelled_defined}
+\alias{as.character}
+\title{Coerce a defined vector to character}
 \usage{
+as_character(x, ...)
+
+\method{as_character}{haven_labelled_defined}(x, preserve_attributes = FALSE, ...)
+
+\method{as.character}{haven_labelled_defined}(x, ...)
+
 as.character(x, ...)
 
 \method{as.character}{haven_labelled_defined}(x, ...)
@@ -28,8 +34,18 @@ a base R character vector. If false, then the effect will be similar to
 }
 \value{
 A character vector.
+
+A character vector.
 }
 \description{
+\code{as_character()} is the recommended method to convert a \code{\link[=defined]{defined()}}
+vector to a character type. It is metadata-aware and ensures that the underlying
+data is character before coercion.
+
+Base R’s \code{as.character()} method applied to \code{defined} vectors simply strips the
+class and returns the values as a plain character vector.
+This is equivalent to calling \code{\link[=as_character]{as_character()}} with \code{preserve_attributes = FALSE}.
+
 Base R's \code{as.character()} does not support custom classes like
 \code{defined}. Calling \code{as.character()} on a \code{defined} vector will drop all
 metadata and class information, which equals to
@@ -40,6 +56,17 @@ vector to character. It is metadata-aware and ensures that the underlying data
 is character before coercion.
 }
 \details{
+If \code{preserve_attributes = TRUE}, the returned character vector retains semantic
+metadata such as \code{unit}, \code{concept}, and \code{namespace}, though the \code{"defined"} class
+itself is removed. If \code{preserve_attributes = FALSE} (default), a plain character
+vector is returned with all attributes stripped.
+
+For numeric-based \code{defined} vectors, \code{as_character()} throws an informative error
+to avoid accidental coercion of non-character data.
+
+Note: \code{as.character()} (base R) is supported but simply returns the raw values,
+and does not preserve or warn about metadata loss.
+
 \code{as_character()} uses \code{preserve_attributes = TRUE}, the resulting
 vector will retain relevant metadata such as the \code{unit}, \code{concept}, and
 \code{namespace} attributes, but it will no longer be of class \code{defined}. If
@@ -51,6 +78,16 @@ prevent accidental coercion of non-numeric data. \cr\cr
 preferred method.
 }
 \examples{
+# Recommended use
+fruits <- defined(c("apple", "avocado", "kiwi"), label = "Fruit", unit = "kg")
+as_character(fruits, preserve_attributes = TRUE)
+
+# Strip metadata
+as_character(fruits, preserve_attributes = FALSE)
+
+# Equivalent base R fallback
+as.character(fruits)
+
 as.character(defined(c("a", "b", "c"), label = "Letter code"))
 as_character(defined(c("a", "b", "c"), label = "Letter code"))
 fruits <- defined(c("apple", "avocado", "kiwi"), label = "Fruit", unit = "kg")
@@ -61,5 +98,7 @@ as_character(fruits, preserve_attributes = TRUE)
 as_character(fruits, preserve_attributes = FALSE)
 }
 \seealso{
+\code{\link[=strip_defined]{strip_defined()}}, \code{\link[=as_numeric]{as_numeric()}}
+
 \code{\link[=strip_defined]{strip_defined()}}
 }
