### 2025-08-06 5c27b13
numeric coercion: edge cases, documentation, new test suite
