+++ man/as_numeric.Rd
@@ -1,62 +1,61 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defined.R
-\name{as.numeric.haven_labelled_defined}
-\alias{as.numeric.haven_labelled_defined}
+% Please edit documentation in R/as_numeric.R
+\name{as_numeric}
 \alias{as_numeric}
 \alias{as_numeric.haven_labelled_defined}
+\alias{as.numeric.haven_labelled_defined}
 \title{Coerce a defined vector to numeric}
 \usage{
-\method{as.numeric}{haven_labelled_defined}(x, ...)
-
 as_numeric(x, ...)
 
 \method{as_numeric}{haven_labelled_defined}(x, preserve_attributes = FALSE, ...)
+
+\method{as.numeric}{haven_labelled_defined}(x, ...)
 }
 \arguments{
-\item{x}{A vector created with\code{\link[=defined]{defined()}}.}
+\item{x}{A vector created with \code{\link[=defined]{defined()}}.}
 
-\item{...}{Further arguments passed to internal methods (not used).}
+\item{...}{Reserved for future use.}
 
-\item{preserve_attributes}{Defaults to \code{FALSE}, in which case the
-\code{unit}, \code{concept} and \code{namespace} attributes will be
-preserved, but the returned value will otherwise become a base R integer,
-double or numeric value. If false, then the effect will be similar to
-\code{\link{strip_defined}}.}
+\item{preserve_attributes}{Logical. Whether to keep metadata attributes.
+Defaults to \code{FALSE}.}
 }
 \value{
-A numeric vector.
+A numeric vector (either bare or with metadata, depending on the
+\code{preserve_attributes} argument).
 }
 \description{
-Base R's \code{as.numeric()} does not support custom classes like
-\code{defined}. Calling \code{as.numeric()} on a \code{defined} vector will drop all
-metadata and class information, which equals to
-\code{as_numeric(x, preserve_attributes = FALSE)}.
+\code{as_numeric()} is the recommended method to convert a \code{\link[=defined]{defined()}}
+vector to a numeric vector. It ensures the underlying data is numeric and can
+optionally preserve semantic metadata.
 
-\code{as_numeric()} is the recommended method to convert a \code{defined}
-vector to numeric. It is metadata-aware and ensures that the underlying data
-is numeric before coercion.
+Base R's \code{as.numeric()} does not support custom classes like \code{defined()}.
+This method drops all metadata and class information, returning a plain
+numeric vector. It is equivalent to \code{as_numeric(x, preserve_attributes = FALSE)}.
 }
 \details{
-\code{as_numeric()} allows \code{preserve_attributes = TRUE} when the
-resulting vector will retain relevant metadata such as the \code{unit},
-\code{concept}, and \code{namespace} attributes, but it will no longer be of class
-\code{defined}. If \code{preserve_attributes = FALSE} (default), a plain numeric
-vector is returned with all metadata and class dropped.\cr For
-character-based \code{defined} vectors, \code{as_numeric()} will throw an informative
-error to prevent accidental coercion of non-numeric data.
+If \code{preserve_attributes = TRUE}, the returned vector retains the \code{unit},
+\code{concept}, and \code{namespace} attributes, but is no longer of class \code{"defined"}.
+If \code{FALSE} (default), a base numeric vector is returned without metadata.
+
+For character-based \code{defined} vectors, an error is thrown to avoid invalid
+coercion.
 }
 \examples{
[…]
