### 2025-08-06 245d46b
New README, vignettes
