+++ README.Rmd
@@ -22,130 +22,63 @@ rlang::check_installed("here")
 
 <!-- badges: end -->
 
-# dataset: Semantic Metadata for Datasets in R
+# Overview
 
-The `dataset` package provides tools to create semantically rich and interoperable datasets in R. Semantically rich datasets that are easier to: - understand by humans, - validate and process by machines, - and share across tools, teams, and domains.
+The `dataset` package helps you create **semantically rich**, **machine-readable**, and **interoperable datasets** in R. It introduces S3 classes that enhance data frames, vectors, and bibliographic entries with formal metadata structures inspired by:
 
-It improves metadata handling by introducing new S3 classes—`defined()`, `dataset_df()`, and `bibrecord()`—that enhance the behaviour of `labelled`, `tibble`, and `bibentry` objects to meet the requirements of:
+-    SDMX (Statistical Data and Metadata eXchange)
 
--   **Statistical Data and Metadata eXchange (SDMX)** standards,
--   **Open Science** metadata practices,
--   **Library and archive metadata** conventions (Dublin Core, DataCite).
+-   Dublin Core / DataCite metadata standards for scientific and open data repositories
 
-Many tools exist to help document, describe, or publish datasets in R, but most separate the metadata from the data itself. This separation increases the risk of losing metadata, misaligning it with the data, or making documentation hard to maintain. We wanted to create a tooling that enables the user from the birth of a dataset till it is potentially serialised with the help of the [rdflib](https://CRAN.R-project.org/package=rdflib) package.
+-   Open Science publishing practices
 
-## Key Features
+The goal is to reduce metadata loss, improve interoperability, and simplify the transition from tidy datasets to web-ready formats like RDF.
 
-### `defined()`
+## Installation
 
-An extended version of `labelled()` vectors. Adds support for:
+You can install the latest released version of **`dataset`** from [CRAN](https://cran.r-project.org/package=dataset) with:
 
--   Variable labels
--   Units of measure (e.g. "million euros")
--   Concept URIs (standardized definitions)
--   Namespaces (to support URI expansion)
-
-```{r setup}
-library(dataset)
-data(gdp)
+```{r, eval=FALSE}
+install.packages("dataset")
 ```
 
-The following small dataset contains the gross domestic product (GDP) data of three small countries:
+To install the development version from GitHub with `pak` or `remotes`:
+
+```{r, eval=FALSE}
+# install.packages("pak")
+pak::pak("dataobservatory-eu/dataset")
+
+# install.packages("remotes")
+remotes::install_github("dataobservatory-eu/dataset")
+```
+
+
+## Minimal Example
 
 ```{r}
-print(gdp)
-```
-
-The `defined` vector class, an extension of the `labelled::labelled` class, allows to add machine-readable concept definitions and namespaces to reveal the coding of the variables. This will allow `AD` to be resolved as <https://dd.eionet.europa.eu/vocabulary/eurostat/geo/AD>
-
-```{r defined}
-geo <- defined(
-  gdp$geo,
-  label = "Country name",
-  concept = "http://purl.org/linked-data/sdmx/2009/dimension#refArea",
-  namespace = "https://dd.eionet.europa.eu/vocabulary/eurostat/geo/$1"
-)
-
-geo[c(1, 3)]
-```
-
-See [defined: Semantically Enriched Vectors](https://dataset.dataobservatory.eu/articles/defined.html)
-
-### `bibrecord()`
-
-An extension of R's built-in `bibentry()` class, with support for:
-
--   Dublin Core Terms (`dcterms`)
--   DataCite metadata
--   Contributor roles (e.g. creator, publisher, data manager)
[…]
