+++ README.md
@@ -14,211 +14,94 @@ WIP](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.o
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci/software-review/issues/553)
 [![DOI](https://zenodo.org/badge/DOI/10.32614/CRAN.package.dataset.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
-[![devel-version](https://img.shields.io/badge/devel%20version-0.3.92-blue.svg)](https://github.com/dataobservatory-eu/dataset)
+[![devel-version](https://img.shields.io/badge/devel%20version-0.3.93-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 [![Codecov test
 coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset/)
 
 <!-- badges: end -->
 
-# dataset: Semantic Metadata for Datasets in R
+# Overview
 
-The `dataset` package provides tools to create semantically rich and
-interoperable datasets in R. Semantically rich datasets that are easier
-to: - understand by humans, - validate and process by machines, - and
-share across tools, teams, and domains.
+The `dataset` package helps you create **semantically rich**,
+**machine-readable**, and **interoperable datasets** in R. It introduces
+S3 classes that enhance data frames, vectors, and bibliographic entries
+with formal metadata structures inspired by:
 
-It improves metadata handling by introducing new S3 classes—`defined()`,
-`dataset_df()`, and `bibrecord()`—that enhance the behaviour of
-`labelled`, `tibble`, and `bibentry` objects to meet the requirements
-of:
+- SDMX (Statistical Data and Metadata eXchange)
 
-- **Statistical Data and Metadata eXchange (SDMX)** standards,
-- **Open Science** metadata practices,
-- **Library and archive metadata** conventions (Dublin Core, DataCite).
+- Dublin Core / DataCite metadata standards for scientific and open data
+  repositories
 
-Many tools exist to help document, describe, or publish datasets in R,
-but most separate the metadata from the data itself. This separation
-increases the risk of losing metadata, misaligning it with the data, or
-making documentation hard to maintain. We wanted to create a tooling
-that enables the user from the birth of a dataset till it is potentially
-serialised with the help of the
-[rdflib](https://CRAN.R-project.org/package=rdflib) package.
+- Open Science publishing practices
 
-## Key Features
+The goal is to reduce metadata loss, improve interoperability, and
+simplify the transition from tidy datasets to web-ready formats like
+RDF.
 
-### `defined()`
+## Installation
 
-An extended version of `labelled()` vectors. Adds support for:
-
-- Variable labels
-- Units of measure (e.g. “million euros”)
-- Concept URIs (standardized definitions)
-- Namespaces (to support URI expansion)
+You can install the latest released version of **`dataset`** from
+[CRAN](https://cran.r-project.org/package=dataset) with:
 
 ``` r
-library(dataset)
-data(gdp)
+install.packages("dataset")
 ```
 
[…]

[…]
