### 2025-08-06 320b92e
Fix numeric coercion and unexpected behavior of recasting to double vectors. New tests.
