+++ man/vec_cast.double.haven_labelled_defined.Rd
@@ -1,0 +1,35 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/as_numeric.R
+\name{vec_cast.double.haven_labelled_defined}
+\alias{vec_cast.double.haven_labelled_defined}
+\title{Cast defined vector to base numeric (double)}
+\usage{
+\method{vec_cast}{double.haven_labelled_defined}(x, to, ...)
+}
+\arguments{
+\item{x}{A \code{\link[=defined]{defined()}} vector of class \code{haven_labelled_defined}.}
+
+\item{to}{Target type (must be \code{double()}).}
+
+\item{...}{Ignored; reserved for future use.}
+}
+\value{
+A plain numeric (double) vector.
+}
+\description{
+This method implements support for converting \code{defined} vectors to base numeric
+via \code{vctrs} casting, typically triggered by \code{as.numeric()}, type coercion in
+\code{tibble}, or base subsetting.
+}
+\details{
+It allows safe and predictable coercion by returning the underlying numeric
+vector, dropping all semantic metadata (e.g., unit, concept).
+}
+\examples{
+x <- defined(c(10, 20), unit = "kg")
+vec_cast(x, double())
+
+# This enables as.numeric(x) to work:
+as.numeric(x)
+
+}
