### 2025-08-06 7dba014
New rOpenSci style documentation, test for warning.
