### 2025-08-06 783bec5
The defined() constructor creates a haven_labelled_defined s3 object. New testthat cases.
