+++ R/defined.R
@@ -1,54 +1,39 @@
-#' @title Create a semantically well-defined, labelled vector
-#' @description
-#' Creates a semantically well-defined vector enriched with metadata.
-#' `defined()` is an S3 constructor that extends numeric or character vectors
-#' with a human-readable label, unit of measurement, linked concept,
-#' and optional namespace. These objects preserve semantics while behaving
-#' like standard vectors in comparisons, printing, and subsetting.
-#'  The \code{defined} constructor creates the objects of this
-#' class, which are semantically extended vectors inherited from
-#' \code{\link[haven:labelled]{haven::labelled}}.
-#' @details
-#' A `defined` vector is an extension of a base vector with additional
-#' semantic metadata:
+#' Create a semantically well-defined, labelled vector
 #'
-#' - A **label** (`label`): a short human-readable description
-#' - A **unit** (`unit`): e.g., "kg", "hours", "USD"
-#' - A **concept** (`concept`): a URI or textual reference
-#' - A **namespace** (`namespace`): for URI-based observation or value identifiers
+#' `defined()` constructs a vector enriched with semantic metadata such as a
+#' label, unit of measurement, concept URI, and optional namespace. These
+#' vectors behave like base R vectors but retain metadata during subsetting,
+#' comparison, and printing.
 #'
-#' The class inherits from `haven::labelled`, supports typical vector
-#' operations (subsetting, comparisons, printing), and integrates with
-#' tibbles and tidy workflows via custom `format()`, `print()`, and
-#' `as.vector()` methods.
+#' The resulting object inherits from [haven::labelled()] and integrates with
+#' tidyverse workflows, enabling downstream conversion to RDF and other
+#' standards.
 #'
-#' Use `is.defined()` to test if an object is of class `defined`.
-#' Use `as_numeric()` and `as_character()` to coerce to base types.
-#' @param x A vector to label. Must be either numeric (integer or double) or
-#'   character.
-#' @param labels A named vector or `NULL`. The vector should be the same type as
-#'   `x`. Unlike factors, labels don't need to be exhaustive: only a fraction of
-#'   the values might be labelled.
-#' @param label A short, human-readable description of the vector or `NULL`.
-#' @param unit A character string of length one containing the unit of measure
-#'   or `NULL`.
-#' @param concept A character string of length one containing a linked
-#'   concept or `NULL`.
-#' @param namespace A namespace for individual observations or categories or
-#'   `NULL`.
-#' @param ... Further parameters for inheritance, not in use.
-#' @family defined metadata methods and functions
-#' @return The constructor \code{defined} returns a vector with defined value
-#'   labels, a variable label, an optional unit of measurement and linked
-#'   concept.\cr \code{is.defined} returns a logical value, stating if the
-#'   object is of class \code{defined}.
+#' @param x A vector of type character, numeric, Date, factor, or a `labelled`
+#'   object.
+#' @param labels An optional named vector of value labels. Only a subset of
+#'   values may be labelled.
+#' @param label A short human-readable label (string of length 1).
+#' @param unit Unit of measurement (e.g., "kg", "hours"). Must be a string of
+#'   length 1 or `NULL`.
+#' @param concept A URI or concept name representing the meaning of the
+#'   variable.
[…]
