+++ tests/testthat/test-defined.R
@@ -1,35 +1,32 @@
-test_that("defined() constructor creates expeced vectors", {
+test_that("defined() constructor creates expected vectors", {
   z <- defined(c(1, 1, 1, 0, 0, 0),
     label = "",
     labels = c("F" = 0, "M" = 1, "_N" = 99),
-    concept = "https://registry.sdmx.org/sdmx/v2/structure/codelist/SDMX/CL_SEX/2.1#"
+    concept = "https://example.org/sex"
   )
   x <- defined(c(0, 1, 0, 1, 1, 0),
     label = "sex",
     labels = c("F" = 0, "M" = 1, "_N" = 99),
-    concept = "https://registry.sdmx.org/sdmx/v2/structure/codelist/SDMX/CL_SEX/2.1#"
+    concept = "https://example.org/sex"
   )
   v <- defined(c(1, 0),
     label = "sex",
     labels = c("F" = 0, "M" = 1, "_N" = 99),
-    concept = "https://registry.sdmx.org/sdmx/v2/structure/codelist/SDMX/CL_SEX/2.1#"
+    concept = "https://example.org/sex"
   )
   y <- defined(c(1, 1, 1, 0, 0, 0),
     label = "sex",
     labels = c("F" = 0, "M" = 1, "_N" = 99),
-    concept = "https://registry.sdmx.org/sdmx/v2/structure/codelist/SDMX/CL_SEX/2.1#"
+    concept = "https://example.org/sex"
   )
   expect_equal(c(1:3, y), c(1, 2, 3, 1, 1, 1, 0, 0, 0))
-  expect_equal(
-    c("a", "b", y),
-    c("a", "b", as.character(c(1, 1, 1, 0, 0, 0)))
-  )
+  expect_equal(c("a", "b", y), c("a", "b", as.character(y)))
   expect_equal(
     c(x, y),
     defined(c(0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0),
       label = "sex",
       labels = c("F" = 0, "M" = 1, "_N" = 99),
-      concept = "https://registry.sdmx.org/sdmx/v2/structure/codelist/SDMX/CL_SEX/2.1#"
+      concept = "https://example.org/sex"
     )
   )
   expect_equal(
@@ -37,374 +34,168 @@ test_that("defined() constructor creates expeced vectors", {
     defined(c(0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0),
       label = "sex",
       labels = c("F" = 0, "M" = 1, "_N" = 99),
-      concept = "https://registry.sdmx.org/sdmx/v2/structure/codelist/SDMX/CL_SEX/2.1#"
+      concept = "https://example.org/sex"
     )
   )
-  a <- defined(c("a", "b"), label = "c")
-  d <- defined(c("d", "e"), label = "c")
+  a <- defined(c("a", "b"), label = "letters")
+  d <- defined(c("d", "e"), label = "letters")
   expect_equal(
     c(a, d),
-    defined(c("a", "b", "d", "e"),
-      label = "c"
-    )
+    defined(c("a", "b", "d", "e"), label = "letters")
   )
 })
 
 test_that("defined() attributes can be retrieved", {
-  sepal_length <- defined(iris$Sepal.Length,
-    labels = NULL,
-    label = "Sepal length",
-    unit = "centimeters",
-    concept = "https://www.wikidata.org/wiki/Property:P2043"
+  numeric_vec <- defined(1:5,
+    label = "Length",
+    unit = "meters",
+    concept = "http://example.org/length",
+    namespace = "http://example.org/ns"
   )
-  myspecies <- defined(
-    x = iris$Species,
[…]
