### 2025-08-06 00c555b
Update the documentation of the example datasets
