+++ data-raw/orange_df.R
@@ -26,6 +26,7 @@ orange_bibentry <- dublincore(
   language = "en",
   description = "The Orange data frame has 35 rows and 3 columns of records of the growth of orange trees."
 )
+
 orange_df <- dataset_df(
   rowid = defined(paste0("orange:", row.names(Orange)),
     label = "ID in the Orange dataset",
