+++ man/gdp.Rd
@@ -5,26 +5,33 @@
 \alias{gdp}
 \title{A Small GDP Dataset}
 \format{
-\subsection{\code{gdp}}{
-
-A data frame with 10 rows and 5 columns:
+A data frame with 10 rows and 5 variables:
 \describe{
-\item{geo}{Country name}
-\item{year}{Reference time period}
-\item{gdp}{Gross Domestic Product}
-\item{unit}{Unit of values}
-\item{freq}{Frequency}
-}
+\item{geo}{Country name (character)}
+\item{year}{Reference year (integer)}
+\item{gdp}{Gross Domestic Product value (numeric)}
+\item{unit}{Unit of measurement, e.g., "Million EUR" (character)}
+\item{freq}{Observation frequency, e.g., "Annual" (character)}
 }
 }
 \source{
-\url{https://doi.org/10.2908/NAIDA_10_GDP}
+Eurostat (2021). GDP and main aggregates - international data
+cooperation (annual data). \doi{10.2908/NAIDA_10_GDP}
 }
 \usage{
 gdp
 }
 \description{
-A subset of data from the Eurostat's GDP and main aggregates -
-international data cooperation annual data
+A compact sample of GDP and main aggregates from Eurostat's annual
+international cooperation dataset. This data subset contains illustrative
+records for select countries and time periods.
+}
+\details{
+This dataset is intended for examples, tests, and demonstration purposes. It
+reflects simplified GDP data as published by Eurostat. The actual Eurostat
+dataset includes more countries, breakdowns, and metadata.
+}
+\examples{
+head(gdp)
 }
 \keyword{datasets}

[…]
