### 2025-08-06 a10a5b0
coercion to  XSD-typed RDF-compatible literals
