+++ R/xsd_convert.R
@@ -1,18 +1,34 @@
-#' @title Convert to XML Schema Definition (XSD) types
-#' @description Convert the numeric, boolean and Date/time columns of a dataset
-#'   \code{xs:decimal}, \code{xsLboolean}, \code{xs:date} and
-#'   \code{xs:dateTime}.
-#' @param x An object to be coerced to an XLM Schema defined string format.
+#' Convert to XML Schema Definition (XSD) Types
+#'
+#' Converts the numeric, logical, and date/time columns of a dataset into [XML
+#' Schema Definition (XSD)](https://www.w3.org/TR/xmlschema11-2/) compatible
+#' string representations such as `xs:decimal`, `xs:boolean`, `xs:date`, and
+#' `xs:dateTime`.
+#'
+#' @param x An object (e.g., vector, data frame) to be coerced to XSD-typed
+#'   string format.
 #' @param idcol The name or position of the column that contains the row
-#'   (observation) identifiers. If \code{NULL}, it will make a new \code{idcol}
-#'   from [row.names()].
-#' @param ... Further optional parameters for generic method.
-#' @return
-#' A character vector of RDF-compatible typed literals. Each element
-#' corresponds to an input value, serialized according to its type
-#' (e.g., `xs:string`, `xs:integer`, `xs:dateTime`).
-#' For data frames or tibbles, each row is converted into a set of
-#' RDF triples, with columns mapped to predicates.
+#'   (observation) identifiers. If `NULL`, a new identifier column will be
+#'   generated using [row.names()].
+#' @param ... Additional arguments passed to methods.
+#'
+#' @return A character vector of RDF-compatible typed literals. Each element
+#'   corresponds to an input value, serialized according to its type (e.g.,
+#'   `xs:string`, `xs:integer`, `xs:dateTime`).
+#'
+#'   For data frames or tibbles, each row is converted into a set of RDF
+#'   triples, with columns mapped to predicates.
+#'
+#' @examples
+#' # Example usage with a simple data frame
+#' df <- data.frame(
+#'   id = 1:2,
+#'   value = c(3.14, 2.71),
+#'   active = c(TRUE, FALSE),
+#'   date = as.Date(c("2020-01-01", "2020-12-31"))
+#' )
+#' as_xsd(df, idcol = "id")
+#'
 #' @export
 
 xsd_convert <- function(x, idcol, ...) {
@@ -42,9 +58,11 @@ get_type <- function(t) {
 }
 
 #' @rdname xsd_convert
+#'
 #' @examples
-#' # Convert data.frame to XML Schema Definition
+#' # Convert a regular data.frame to XSD-typed RDF-compatible literals
 #' xsd_convert(data.frame(a = 1:3, b = c("a", "b", "c")))
+#'
 #' @exportS3Method
 xsd_convert.data.frame <- function(x, idcol = NULL, ...) {
   # Identify ID column (or default to row names)
@@ -76,17 +94,23 @@ xsd_convert.data.frame <- function(x, idcol = NULL, ...) {
 }
 
 #' @rdname xsd_convert
+#'
 #' @examples
-#' # Convert dataset to XML Schema Definition
+#' # Convert a dataset_df object to XSD-typed RDF-compatible literals
 #' xsd_convert(head(dataset_df(orange_df)))
-#' @export
+#'
 #' @exportS3Method
[…]
