+++ man/xsd_convert.Rd
@@ -14,7 +14,7 @@
 \alias{xsd_convert.POSIXct}
 \alias{xsd_convert.Date}
 \alias{xsd_convert.difftime}
-\title{Convert to XML Schema Definition (XSD) types}
+\title{Convert to XML Schema Definition (XSD) Types}
 \usage{
 xsd_convert(x, idcol, ...)
 
@@ -43,40 +43,88 @@ xsd_convert(x, idcol, ...)
 \method{xsd_convert}{difftime}(x, idcol = NULL, ...)
 }
 \arguments{
-\item{x}{An object to be coerced to an XLM Schema defined string format.}
+\item{x}{An object (e.g., vector, data frame) to be coerced to XSD-typed
+string format.}
 
 \item{idcol}{The name or position of the column that contains the row
-(observation) identifiers. If \code{NULL}, it will make a new \code{idcol}
-from \code{\link[=row.names]{row.names()}}.}
+(observation) identifiers. If \code{NULL}, a new identifier column will be
+generated using \code{\link[=row.names]{row.names()}}.}
 
-\item{...}{Further optional parameters for generic method.}
+\item{...}{Additional arguments passed to methods.}
 }
 \value{
 A character vector of RDF-compatible typed literals. Each element
-corresponds to an input value, serialized according to its type
-(e.g., \code{xs:string}, \code{xs:integer}, \code{xs:dateTime}).
-For data frames or tibbles, each row is converted into a set of
-RDF triples, with columns mapped to predicates.
+corresponds to an input value, serialized according to its type (e.g.,
+\code{xs:string}, \code{xs:integer}, \code{xs:dateTime}).
+
+For data frames or tibbles, each row is converted into a set of RDF
+triples, with columns mapped to predicates.
 }
 \description{
-Convert the numeric, boolean and Date/time columns of a dataset
-\code{xs:decimal}, \code{xsLboolean}, \code{xs:date} and
+Converts the numeric, logical, and date/time columns of a dataset into \href{https://www.w3.org/TR/xmlschema11-2/}{XML Schema Definition (XSD)} compatible
+string representations such as \code{xs:decimal}, \code{xs:boolean}, \code{xs:date}, and
 \code{xs:dateTime}.
 }
 \examples{
-# Convert data.frame to XML Schema Definition
+# Example usage with a simple data frame
+df <- data.frame(
+  id = 1:2,
+  value = c(3.14, 2.71),
+  active = c(TRUE, FALSE),
+  date = as.Date(c("2020-01-01", "2020-12-31"))
+)
+as_xsd(df, idcol = "id")
+
+# Convert a regular data.frame to XSD-typed RDF-compatible literals
 xsd_convert(data.frame(a = 1:3, b = c("a", "b", "c")))
-# Convert dataset to XML Schema Definition
+
+# Convert a dataset_df object to XSD-typed RDF-compatible literals
 xsd_convert(head(dataset_df(orange_df)))
-# Convert characters:
+
+# Convert a tibble to XSD-typed RDF-compatible literals
+library(tibble)
+xsd_convert(tibble(a = 1:3, b = c("x", "y", "z")))
+
+# Convert character vector to XSD-typed literals
 xsd_convert(c("apple", " banana ", "cherry"))
 
-# To handle whitespace:
-xsd_convert(trimws(c("apple", " banana ", "cherry"), "both"))
-# Convert integers or doubles, numbers:
+# Preprocess whitespace before conversion
+xsd_convert(trimws(c("apple", " banana ", "cherry"), which = "both"))
+
+# Convert numeric values (integers or doubles) to XSD typed literals
[…]
