### 2025-08-06 e2df1d9
id_to_column new tests without iris, edge cases. New documentation
