+++ tests/testthat/test-id_to_column.R
@@ -1,59 +1,60 @@
-test_that("id_to_column() works on data.frame", {
-  df <- iris
-
-  # Default behavior (prefix = "eg:")
-  df_with_ids <- id_to_column(df)
-  expect_true("rowid" %in% names(df_with_ids))
-  expect_equal(df_with_ids$rowid[1], "eg:1")
-
-  # Custom prefix
-  df_with_custom_prefix <- id_to_column(df, prefix = "iris:")
-  expect_equal(df_with_custom_prefix$rowid[3], "iris:3")
-
-  # No prefix
-  df_with_no_prefix <- id_to_column(df, prefix = NULL)
-  expect_equal(df_with_no_prefix$rowid[1], "1")
-
-  # Custom IDs
-  ids <- paste0("iris-record-", seq_len(nrow(df)))
-  df_with_custom_ids <- id_to_column(df, prefix = "ex:", ids = ids)
-  expect_equal(df_with_custom_ids$rowid[1], "ex:iris-record-1")
-
-  # Error on bad ID length
-  expect_error(id_to_column(df, ids = c("a", "b")), "ids must be of same lengths")
+test_that("id_to_column works on a basic data.frame", {
+  df <- data.frame(a = 1:3, b = letters[1:3])
+  out <- id_to_column(df)
+  expect_true("rowid" %in% names(out))
+  expect_equal(out$rowid, paste0("eg:", 1:3))
 })
 
-test_that("id_to_column() works on dataset_df and preserves metadata", {
-  creator_person <- person(given = "Jane", family = "Doe", role = "aut")
+test_that("id_to_column replaces existing rowid column", {
+  df <- data.frame(rowid = c("x", "y", "z"), a = 1:3)
+  out <- id_to_column(df, prefix = "custom:")
+  expect_equal(out$rowid, paste0("custom:", 1:3))
+})
 
-  original <- dataset_df(
-    mtcars,
-    dataset_bibentry = dublincore(
-      title = "MTCars Dataset",
-      creator = creator_person,
-      publisher = "Motor Trend",
-      dataset_date = 1974
-    )
+test_that("id_to_column accepts custom ids", {
+  df <- data.frame(a = 1:2)
+  out <- id_to_column(df, prefix = "ex:", ids = c("foo", "bar"))
+  expect_equal(out$rowid, c("ex:foo", "ex:bar"))
+})
+
+test_that("id_to_column errors on mismatched ids length", {
+  df <- data.frame(a = 1:3)
+  expect_error(
+    id_to_column(df, ids = c("a", "b")),
+    "ids must be of same lengths as nrow"
   )
+})
 
-  out <- id_to_column(original, prefix = "mt:")
+test_that("id_to_column handles special characters in row names", {
+  df <- data.frame(a = 1:2)
+  rownames(df) <- c("first row", "second#row!")
+  out <- id_to_column(df)
+  expect_equal(out$rowid, c("eg:first-row", "eg:second-row-"))
+})
 
+test_that("id_to_column works with prefix = NULL", {
+  df <- data.frame(a = 1:2)
+  out <- id_to_column(df, prefix = NULL)
+  expect_equal(out$rowid, c("1", "2"))
+})
+
+test_that("id_to_column preserves dataset_df structure", {
+  df <- dataset_df(a = 1:2)
+  out <- id_to_column(df)
   expect_s3_class(out, "dataset_df")
   expect_true("rowid" %in% names(out))
-
-  # Bibentry preserved
-  b <- get_bibentry(out)
-  expect_s3_class(b, "bibentry")
-  expect_equal(b$title, "MTCars Dataset")
-  expect_equal(b$author, creator_person)
-
-  # Provenance and subject preserved
[…]
