+++ tests/testthat/test-var_label.R
@@ -1,53 +1,106 @@
-test_that("var_label() works", {
-  expect_equal(
-    var_label(iris_dataset$Sepal.Length),
-    "Length of the sepal in cm"
+test_that("var_label() works with defined() objects", {
+  x <- defined(1:5, label = "Defined label")
+  expect_equal(var_label(x), "Defined label")
+
+  var_label(x) <- "Updated label"
+  expect_equal(var_label(x), "Updated label")
+})
+
+test_that("var_label<- works with haven_labelled_defined", {
+  x <- haven::labelled(1:5, labels = c(a = 1, b = 2))
+  x <- defined(x)
+  var_label(x) <- "Labelled defined"
+  expect_equal(var_label(x), "Labelled defined")
+})
+
+test_that("var_label.dataset_df returns expected label list", {
+  df <- dataset_df(
+    x = defined(1:3, label = "Label X"),
+    y = defined(4:6),
+    z = defined(letters[1:3], label = "Label Z")
   )
-  expect_equal(class(var_label(iris_dataset, unlist = TRUE)), "character")
-  test_df <- dataset_df(a = 1:2, b = 3:4)
-  expect_equal(
-    var_label(test_df, unlist = TRUE, null_action = "fill"),
-    c(rowid = "rowid", a = "a", b = "b")
+
+  expect_equal(var_label(df)$x, "Label X")
+  expect_null(var_label(df)$y)
+  expect_equal(var_label(df)$z, "Label Z")
+})
+
+test_that("var_label.dataset_df supports null_action options", {
+  df <- dataset_df(a = defined(1:2, label = "Has Label"), b = 3:4)
+
+  expect_equal(var_label(df, null_action = "fill")$b, "b")
+  expect_true(is.na(var_label(df, null_action = "na")$b))
+  expect_equal(var_label(df, null_action = "empty")$b, "")
+  expect_null(var_label(df, null_action = "keep")$b)
+
+  skipped <- var_label(df, null_action = "skip")
+  expect_false("b" %in% names(skipped))
+})
+
+test_that("var_label.dataset_df returns named vector with unlist = TRUE", {
+  df <- dataset_df(
+    u = defined(1:3, label = "U label"),
+    v = 4:6
   )
+  out <- var_label(df, unlist = TRUE, null_action = "empty")
+  expect_type(out, "character")
+  expect_equal(out[["u"]], "U label")
+  expect_equal(out[["v"]], "")
 })
 
-test_that("var_label()<-  works", {
-  iris_dataset_2 <- iris_dataset
-  var_label(orange_df$circumference) <- "circumference (breast height)"
-  var_label(iris_dataset_2$Sepal.Length) <- "Length of the sepal in centimeters"
-  expect_equal(
-    var_label(iris_dataset_2$Sepal.Length),
-    "Length of the sepal in centimeters"
+test_that("var_label.dataset_df handles recurse = TRUE on packed columns", {
+  packed <- dataset_df(
+    a = defined(1:3, label = "A"),
+    b = tibble::tibble(x = defined(1:3, label = "SubX"), y = 4:6)
   )
-  expect_equal(
-    var_label(orange_df$circumference),
-    "circumference (breast height)"
-  )
-})
 
-test_that("var_label() throws error", {
-  test_df <- dataset_df(a = 1:2, b = 3:4)
-  expect_error(var_label(test_df) <- c("A", "B", "C", "E"))
+  out <- var_label(packed, recurse = TRUE, null_action = "empty")
+  expect_type(out$b, "list")
+  expect_equal(out$b$x, "SubX")
 })
 
 
-test_that("var_label.dataset_df() works", {
[…]
