+++ R/var_label.R
@@ -1,67 +1,54 @@
-#' @title Get / Set a variable label
+#' Get or Set a Variable Label
 #'
-#' @description
-#' Add a human readable, easier to understand label as a metadata attribute to a
-#' variable or vector than the programmatic vector object name, or column name
-#' in the data frame.
+#' Adds or retrieves a human-readable label as a metadata attribute for a
+#' variable or vector. This label is useful for making variables easier to
+#' understand than their programmatic names (e.g., column names).
 #'
 #' @details
-#' See \code{\link[labelled:var_label]{labelled::var_label}} for details about
-#' variable labels.
-#' \cr
-#' See \code{vignette("defined", package = "dataset")} to use comprehensively
-#' with variable labels, namespaces, units of measures, and machine-independent
-#' permanent variable identifiers.
+#' This interface builds on [labelled::var_label()] and is compatible with
+#' the `defined()` infrastructure for semantic metadata (labels, namespaces,
+#' units, and variable identifiers).
 #'
-#' @name var_label
-#' @return \code{var_label()} returns the \code{label} attribute as a character
-#' string. The \code{var_label<-} assignment method allows to add, remove, or
-#' overwrite this attribute on a vector \code{x}. The assignment function returns
-#' the \code{x} vector invisibly.
+#' See [labelled::var_label()] for low-level usage. For a comprehensive
+#' guide to working with variable labels and semantic metadata, see:
+#' `vignette("defined", package = "dataset")`.
+#'
+#' @param x A vector or data frame.
+#' @param value A character string or `NULL` to remove the label.
+#'   When used with `var_labels()` on a data frame, this can also be a named list
+#'   or a character vector of the same length as the number of columns in `x`.
+#' @param ... Further arguments passed to or used by methods.
+#'
+#' @return
+#' - `var_label(x)` returns the `"label"` attribute of `x` as a character string.
+#' - `var_label(x) <- value` sets, removes, or replaces the label attribute of `x`,
+#' returning the updated object invisibly.
 #'
 #' @examples
-#' # Retrieve the label attribute:
+#' # Retrieve the label attribute
 #' var_label(orange_df$circumference)
 #'
-#' # To (re)set the label attribute:
-#' `var_label<-`(orange_df$circumference, "circumference (breast height)")
+#' # Set or update the label attribute
+#' var_label(orange_df$circumference) <- "circumference (breast height)"
 #'
+#' @seealso [labelled::var_label()], [var_labels()], [defined()]
+#' @family defined metadata methods and functions
+#'
+#' @name var_label
 #' @export
 #' @importFrom labelled var_label
 #' @importFrom labelled `var_label<-`
-#'
-#' @param x a vector or a data.frame
-#' @param value a character string or `NULL` to remove the label
-#'  For data frames, with `var_labels()`, it could also be a named list or a
-#'  character vector of same length as the number of columns in `x`.
-#' @param ... Further arguments passed to or used by methods.
-#' @family defined metadata methods and functions
 NULL
 
 #' @rdname var_label
-#' @importFrom labelled `var_label<-`
 #' @export
+#' @importFrom labelled `var_label<-`
 var_label.defined <- function(x, ...) {
   NextMethod()
 }
 
-#' @rdname var_label
-#' @param unlist for data frames, return a named vector instead of a list
-#' @param null_action for data frames, by default `NULL` will be returned for
-#' columns with no variable label. Use "fill" to populate with the column name
-#' instead, "skip" to remove such values from the returned list, "na" to
[…]

[…]
