### 2025-08-06 c338b2e
Improve publisher() documentation, test for edge casee and remove isis_df
