+++ NAMESPACE
@@ -4,7 +4,6 @@ S3method("[",dataset_df)
 S3method("[",haven_labelled_defined)
 S3method("[[",haven_labelled_defined)
 S3method("var_concept<-",default)
-S3method("var_label<-",dataset_df)
 S3method("var_label<-",haven_labelled_defined)
 S3method("var_namespace<-",default)
 S3method("var_unit<-",default)
