### 2025-08-06 620285c
Fixing a bug changing the rights statement metadata.  Improving the documentation and testing for edge cases.
