### 2025-08-06 08b57bd
documenting further dataset_df methods
