### 2025-08-06 d1048fb
Replace var_unit documentation and test for edge cases without iris_df
