+++ R/var_unit.R
@@ -1,41 +1,47 @@
-#' @title Get / Set a unit of measure
+#' Get or Set a Unit of Measure
+#'
+#' Adds or retrieves a unit of measure (UoM) attribute to a vector. Units
+#' provide semantic meaning for numeric or character data — such as currency,
+#' weight, or time — helping prevent incorrect operations like merging values
+#' measured in incompatible units.
+#'
 #' @param x A vector.
-#' @param value A character string or `NULL` to remove the unit of measure.
-#' @param ... Further potential parameters reserved for inherited classes.
-#' @details
-#' The aim of the \code{unit} attribute is to add to the R vector object its
-#' unit of measure (for example, physical units like gram and kilogram or
-#' currency units like dollars or euros), so that they are not concatenated or
-#' joined in a syntactically correct but semantically incorrect way (i.e.,
-#' accidentally concatenating values quoted in dollars and euros from different
-#' subvectors.) This is particularly useful when working with linked open data,
-#' i.e., when joins or concatenations are performed on data arriving from a remote
-#' source.\cr \cr
-#' `get_variable_units()` is identical to `var_unit()`.
-#' \cr
-#' \cr
-#' See \code{vignette("defined", package = "dataset")} to use comprehensively
-#' with variable labels, namespaces, units of measures, and machine-independent
-#' permanent variable identifiers.
-#' @family defined metadata methods and functions
+#' @param value A character string specifying the unit of measure, or `NULL` to
+#'   remove the unit.
+#' @param ... Further arguments for method extensions.
+#'
+#' @details The `"unit"` attribute stores a machine-readable representation of a
+#' unit of measure (e.g., `"kg"`, `"USD"`, `"days"`). This is useful when
+#' working with linked open data or when combining data from multiple sources
+#' where silent mismatches in units could cause errors.
+#'
+#' For full integration with semantic metadata (e.g., labels, concepts,
+#' namespaces), use [defined()] vectors or [dataset_df()] objects.
+#'
+#' `get_variable_units()` is an alias for `var_unit()`.
+#'
+#' See `vignette("defined", package = "dataset")` for end-to-end examples
+#' involving semantic enrichment.
+#'
+#' @return
+#' - `var_unit(x)` returns the `"unit"` attribute as a character string.
+#' - `var_unit(x) <- value` sets, updates, or removes the unit and returns
+#' the modified vector invisibly.
+#'
 #' @examples
-#' # The defined vector class and dataset_df support units of measure attributes:
+#' # Retrieve the unit of measure (if defined)
 #' var_unit(orange_df$circumference)
 #'
-#' # Normally columns of a data.frame do not have a unit attribute:
+#' # Regular data.frame columns have no unit by default
 #' var_unit(mtcars$wt)
 #'
-#' # You can add them with the assignment function:
+#' # Add a unit to a column
 #' var_unit(mtcars$wt) <- "1000 lbs"
 #'
-#' # To remove a unit of measure assign the NULL value:
+#' # Remove the unit
 #' var_unit(mtcars$wt) <- NULL
-#' @return The unit attribute of a vector constructed with[defined()],
-#' or any vector that is enriched with a unit attribute. \cr
-#' \cr
-#' The \code{var_unit<-}
-#' assignment method allows to add, remove, or overwrite this attribute on a vector
-#' \code{x}. The assignment function returns the \code{x} vector invisibly.
+#'
+#' @family defined metadata methods and functions
 #' @export
[…]
