### 2025-08-06 3286e50
Replace the documentation of bibrecord()
