+++ tests/testthat/test-dublincore.R
@@ -197,5 +197,7 @@ test_that("dublincore_to_triples produces valid n-triples format", {
   expect_true(any(grepl("<http://purl.org/dc/terms/subject>", triples)))
 })
 
-dublincore(title = "Small Dataset",
-           creator = person("Jane Doe", role="aut"))
+dublincore(
+  title = "Small Dataset",
+  creator = person("Jane Doe", role = "aut")
+)
