+++ R/var_label.R
@@ -122,10 +122,10 @@ label_attribute <- function(x) {
 #' @examples
 #' # Example: Retrieve variable labels from a dataset_df
 #' df <- dataset_df(
-#'         id = defined(1:3, label = "Observation ID"),
-#'         temp = defined(c(22.5, 23.0, 21.8), label = "Temperature (°C)"),
-#'         site = defined(c("A", "B", "A"))
-#'         )
+#'   id = defined(1:3, label = "Observation ID"),
+#'   temp = defined(c(22.5, 23.0, 21.8), label = "Temperature (°C)"),
+#'   site = defined(c("A", "B", "A"))
+#' )
 #'
 #' # List form (default)
 #' var_label(df)
@@ -157,11 +157,11 @@ var_label.dataset_df <- function(x,
 
     if (is.null(label)) {
       label <- switch(null_action,
-                      keep = NULL,
-                      fill = name,
-                      skip = NULL,
-                      na = NA_character_,
-                      empty = ""
+        keep = NULL,
+        fill = name,
+        skip = NULL,
+        na = NA_character_,
+        empty = ""
       )
     }
 
