+++ R/data-gdp.R
@@ -25,4 +25,3 @@
 #' @examples
 #' head(gdp)
 "gdp"
-
