### 2025-08-06 c079093
Test edge cases. New documentation. Bug fix for changing and retrieving dc:format.
