+++ R/dublincore.R
@@ -1,137 +1,62 @@
-#' @title Add or get Dublin Core metadata
-#' @description Add metadata conforming the
-#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/format/}{DCMI
-#'   Metadata Terms}. to datasets, i.e. structured R data.frame or list objects,
-#'   for an accurate and consistent identification of a resource for citation
-#'   and retrieval purposes.
-#' @details The Dublin Core, also known as the Dublin Core Metadata Element Set
-#'   (DCMES), is a set of fifteen main metadata items for describing digital or
-#'   physical resources, such as datasets or their printed versions. Dublin Core
-#'   has been formally standardized internationally as ISO 15836, as IETF RFC
-#'   5013 by the Internet Engineering Task Force (IETF), as well as in the U.S.
-#'   as ANSI/NISO Z39.85.\cr \cr To provide compatibility with
-#'   \code{\link[utils]{bibentry}}  we try to add \code{dataset_date} parameter
-#'   first as \code{publication_date} metadata field, and as a \code{year}
-#'   field, too. This element can be get or set with
-#'   \code{\link{publication_year}}.
-#' @param x An R object of type data.frame, or inherited data.table, tibble;
-#'   alternatively a well structured R list.
-#' @details The \code{ResourceType} property will be by definition "Dataset".
-#'   The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will
-#'   automatically added to the dataset.
-#' @param title
-#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/title/}{dct:title},
-#'   a name given to the resource.[datacite()] allows the use of
-#'   alternate titles, too. See \code{\link{dataset_title}}.
-#' @param creator An entity primarily responsible for making the resource.
-#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/creator/}{dct:creator}
-#'   Corresponds to \code{Creator} in[datacite()]. See
-#'   \code{\link{creator}}.
-#' @param identifier An unambiguous reference to the resource within a given
-#'   context. Recommended practice is to identify the resource by means of a
-#'   string conforming to an identification system. Examples include
-#'   International Standard Book Number (ISBN), Digital Object Identifier (DOI),
-#'   and Uniform Resource Name (URN). Select and identifier scheme from
-#'   \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/index.html}{registered
-#'   URI schemes maintained by IANA}. More details:
-#'   \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/}{Guidelines
-#'   for using resource identifiers in Dublin Core metadata and IEEE LOM}.
-#'   Similar to \code{Identifier} in[datacite()]. See
-#'   \code{\link{identifier}}.
-#' @param publisher Corresponds to
-#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#publisher}{dct:publisher}
-#'   and Publisher in DataCite. The name of the entity that holds, archives,
-#'   publishes prints, distributes, releases, issues, or produces the resource.
-#'   This property will be used to formulate the citation, so consider the
-#'   prominence of the role. For software, use \code{Publisher} for the code
-#'   repository. If there is an entity other than a code repository, that
-#'   "holds, archives, publishes, prints, distributes, releases, issues, or
-#'   produces" the code, use the property
-#'   Contributor/contributorType/hostingInstitution for the code repository. See
-#'   \code{\link{publisher}}.
-#' @param subject In
[…]
