+++ tests/testthat/test-dublincore.R
@@ -1,203 +1,90 @@
-test_that("new_dublincore() works", {
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre")
-    )$author,
-    person("Jane", "Doe", role = "cre")
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = c(
-        person("Jane", "Doe", role = "cre"),
-        person("Joe", "Doe", role = "cre")
-      )
-    )$author,
-    c(
-      person("Jane", "Doe", role = "cre"),
-      person("Joe", "Doe", role = "cre")
-    )
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre")
-    )$title,
-    "Test"
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre"),
-      publisher = person("My Publisher Inc.", role = "pbl")
-    )$publisher,
-    "My Publisher Inc. [pbl]"
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre"),
-      datasource = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
-    )$datasource,
-    "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre"),
-      dataset_date = 1935
-    )$date,
-    "1935"
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre"),
-      dataset_date = 1935
-    )$year,
-    "1935"
-  )
-
-  expect_equal(
-    new_dublincore(
-      title = "Test",
-      creator = person("Jane", "Doe", role = "cre"),
-      language = "en"
-    )$language,
-    "en"
-  )
-})
-
-test_that("dublincore works", {
-  dct_iris1 <- dublincore(
-    title = "Iris Dataset",
-    creator = c(
-      person(given = "Edgar", family = "Anderson", role = "aut"),
-      person(given = "Jane D", family = "Anderson", role = "cre")
-    ),
-    publisher = person("American Iris Society", role = "pbl"),
-    datasource = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
-    dataset_date = 1935,
-    language = "en",
-    description = "The famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica."
-  )
-  expect_equal(dct_iris1$author, c(
-    person(given = "Edgar", family = "Anderson", role = "aut"),
-    person(given = "Jane D", family = "Anderson", role = "cre")
-  ))
-  expect_equal(dct_iris1$publisher, "American Iris Society")
-  dct_iris <- dublincore(
-    title = "Iris Dataset",
-    creator = person("Edgar", "Anderson", role = "aut"),
-    publisher = person("American Iris Society", role = "pbl"),
-    contributor = person("Daniel", "Antal", role = "dtm"),
-    datasource = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
-    dataset_date = 1935,
-    language = "en",
[…]

[…]
