### 2025-08-06 f59906d
Fix errors with as_dublincore in some edge cases
