### 2025-08-06 0f4c480
as_character must not coerce from character
