### 2025-08-07 2efd508
Rewrite publication_year test, add edge cases, improve documentation for rOpenSci
