### 2025-08-07 6904f75
Re-document, and new tests, edge cases.
