### 2025-08-07 bb3b200
Design principles
