+++ vignettes/design.Rmd
@@ -1,10 +1,13 @@
 ---
-title: "dataset: Semantically Enriched, Standards-Aligned Datasets in R"
+title: "Design Principles & Future Work Semantically Enriched, Standards-Aligned Datasets in R"
 output: rmarkdown::html_vignette
 vignette: >
+  %\VignetteEncoding{UTF-8}
   %\VignetteIndexEntry{dataset: Semantically Enriched, Standards-Aligned Datasets in R}
   %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
+editor_options: 
+  markdown: 
+    wrap: 80
 ---
 
 ```{r, include = FALSE}
@@ -14,42 +17,89 @@ knitr::opts_chunk$set(
 )
 ```
 
-## Abstract 
+## Abstract
 
 ```{r setup}
 library(dataset)
 ```
 
-> "A dataset is an identifiable collection of data available for access or download in one or more formats." — ISO/IEC 20546
+> “A dataset is an identifiable collection of data available for access or
+> download in one or more formats.” — ISO/IEC 20546
 
-The dataset package enriches R's data objects with machine-readable metadata. It embeds semantic definitions and provenance at both the variable and dataset level, offering users interoperable, standards-aligned outputs. It follows a semantic early-binding design: metadata is attached during dataset creation, not retrofitted later. This ensures that context and meaning are preserved from exploration to publication, and some documentation can be automated.
+The `dataset` package enriches R data objects with machine-readable metadata by
+embedding semantic definitions and provenance at both the variable and dataset
+levels. It follows a semantic **early-binding** design: metadata is attached at
+creation time, not retrofitted post hoc. This ensures that meaning and context
+are preserved throughout the data lifecycle — from exploration to publication —
+and enables partial automation of documentation.
 
-This article outlines the design philosophy behind the dataset package, including its theoretical underpinnings, structure, relation to other R tools, and example workflows. It is intended as a long-form complement to the package vignettes.
+This article outlines the design philosophy behind the `dataset` package,
+including its theoretical foundations, structure, relationship to other R tools,
+and example workflows. It serves as a long-form complement to the package
+vignettes.
 
 ## Introduction and Motivation
 
-> "The principles of tidy data provide a standard way to organise data values within a dataset."  
+> “The principles of tidy data provide a standard way to organise data values
+> within a dataset.”  
 > — Wickham (2014)
 
-Wickham's article refers to a particular manifestation of a dataset within R: a `data.frame` or `tibble` object. However, R users often conflate this tangible object — which exists in computer memory and within the R environment — with the broader concept of a dataset. In reality, a dataset refers to its abstract structure and semantics, which can be realized in many forms across databases, APIs, flat files, or publication platforms.
+The `dataset` package extends R’s native data structures by embedding
+machine-readable semantics and provenance directly in tidy data objects. It
+builds on tidy data principles (Wickham, 2014) but introduces a **semantic
+early-binding** approach: metadata is attached when the dataset is created,
+ensuring that context and meaning are preserved through all stages of the
+workflow — including transformation, validation, serialization, and reuse.
 
-A dataset, when preserved in a less ephemeral format such as a file or dissemination, becomes a type of digital _resource_ — a broader class that also includes data papers, software packages, and other scholarly outputs. Resources are typically described using metadata and may be serialized using the Resource Description Framework (RDF), which ensures representation that is independent of language, operating system, or schema. This kind of serialization facilitates resource discovery and reuse.
+While tidyverse tools enforce structural clarity, they are generally agnostic
+about semantics. Variables may be misinterpreted, joined incorrectly, or
+published without context. `dataset` addresses this gap by aligning with
+international metadata standards, supporting RDF export, and providing an
+interface to the W3C Data Cube model.
 
[…]
