+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.95
-Date: 2025-08-06
+Version: 0.3.96
+Date: 2025-08-07
 Language: en-GB
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 
