### 2025-08-07 bbde704
Work with legacy and current language metadata. Test edge cases.
