+++ R/language.R
@@ -1,34 +1,70 @@
-#' @title Get/set the primary language of the dataset
-#' @description Add the optional Language property as an attribute to an R
-#'   object.
-#' @details Language is an optional property in DataCite 4.4; see:
-#'   \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties#9-language}{datacite:Language}\cr
-#'   It is a part of the "core" of the
-#'   \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/}{Dublin
-#'   Core metadata terms}. The language parameter is validated against the
-#'   \code{[ISOcodes]{ISO_639_2}} table.\cr The attribute \code{language} is
-#'   added to the object. It will be exported into DataCite applications in a
-#'   capitalized \code{Lanugage} format.
-#' @param x A semantically rich data frame object created by
-#'   [dataset_df()] or \code{\link{as_dataset_df}}.
-#' @param value The language to be added to the object attributes, added by
-#'   name, or as a 2- or 3-character code for the language. You can add a
-#'   language code or language name, and the parameter is normalized to
-#'   \code{tolower(language)}. (The ISO 639 standard capitalizes language names
-#'   and uses lower case for the codes.)
-#' @param iso_639_code Defaults to \code{ISO 639-3}, alternative is \code{ISO
-#'   639-1}.
-#' @return The Language is added to the \code{x} as \code{ISO 639-1}, the
-#'   Datacite recommendation, or \code{ISO 639-3} used by the Zenodo data
-#'   repository.
+#' @title Set the Primary Language of a Dataset
+#'
+#' @description
+#' Assign the primary language of a semantically rich dataset object using an
+#' ISO 639 language code or full language name. This sets the `language`
+#' attribute in the dataset's metadata.
+#'
+#' @details
+#' This function supports recognition of:
+#' - 2-letter codes (ISO 639-1, e.g., `"en"`, `"fr"`)
+#' - 3-letter codes from both:
+#'   - `Alpha_3_B` (bibliographic, e.g., `"fre"`)
+#'   - `Alpha_3_T` (terminologic, e.g., `"fra"`)
+#' - Full language names (e.g., `"English"`, `"French"`)
+#'
+#' For compatibility with open science repositories and modern metadata
+#' standards, this function **returns the terminologic code** (`Alpha_3_T`)
+#' when available. If `Alpha_3_T` is missing for a language, the legacy
+#' bibliographic code (`Alpha_3_B`) is used as a fallback.
+#'
+#' Full language names (e.g., `"English"`, `"Spanish"`) are matched
+#' case-insensitively against the ISO 639-2 Name field. Exact matches are
+#' attempted first; if none are found, a prefix match is used. For example:
+#' - `"English"` returns `"eng"`
+#' - `"English, Old"` returns `"ang"`
+#'
+#' This means that:
+#' - Both `"fra"` (terminologic) and `"fre"` (bibliographic) will be accepted
+#'   as valid input for French
+#' - The resulting value stored and returned will be `"fra"`
+#'
+#' This behavior aligns with:
+#' - [DataCite Metadata Schema 4.4](https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties#9-language)
+#' - [schema.org](https://schema.org/inLanguage)
[…]
