### 2025-08-07 2bcf5ce
Edge cases of namespace handling, new test suite, rOpenSci documentation
