+++ R/var_namespace.R
@@ -1,33 +1,57 @@
-#' @title Get / Set a namespace of measure
-#' @description
-#' Retain the namespace part of a permanent, global variable identifier which
-#' is independent of the R instance in use.
-#' @details
-#' The namespace attribute is useful when users join or concatenate data from
-#' remote, linked, and open data sources. In such cases, variable identifiers
-#' (labels or names) are often resolved with a common namespace prefix, which,
-#' together with the namespace, forms a URI or IRI permanent identifier for the
-#' variable. Retaining the namespace in such cases allows cross-validation or
-#' success later updates of the vector (as a column of a dataset.)
-#' @param x a vector
-#' @param value a character string or `NULL` to remove the namespace of measure.
-#' @param ... Further potential parameters reserved for inherited classes.
-#' @details
-#'   `get_variable_namespaces()` is identical to `var_namespace()`.
-#' See \code{vignette("defined", package = "dataset")} to use comprehensively
-#' with variable labels, namespaces, units of measures, and machine-independent
-#' permanent variable identifiers.
-#' @examples
-#' qid <- defined(c("Q275912", "Q116196078"),
-#'   namespace = c(wd = "https://www.wikidata.org/wiki/")
-#' )
-#' var_namespace(qid)
+#' @title Get or Set the Namespace of a Variable
 #'
-#' # To remove a namespace
-#' var_namespace(qid) <- NULL
+#' @description
+#' Retrieve or assign the namespace part of a permanent, global variable
+#' identifier, independent of the current R session or instance.
+#'
+#' @details
+#' The `namespace` attribute is useful when working with remote, linked, or
+#' open data sources. Variable identifiers in such datasets are often qualified
+#' with a common namespace prefix. When combined, the prefix and namespace form
+#' a persistent URI or IRI for the variable.
+#'
+#' Retaining the namespace ensures the identifiers remain valid and resolvable
+#' during validation, merging, or future updates of the vector (such as when it
+#' is used as a column in a dataset).
+#'
+#' `get_variable_namespaces()` is an alias for `var_namespace()`.
+#' `namespace_attribute()` and `set_namespace_attribute()` are internal helpers.
+#'
+#' For full usage, see:
+#' `vignette("defined", package = "dataset")` — demonstrating integration of
+#' variable labels, namespaces, units of measure, and machine-independent
+#' identifiers.
+#'
+#' @param x A vector.
+#' @param value A character string specifying the namespace, or `NULL` to
+#'   remove it.
+#' @param ... Additional arguments for method compatibility with other classes.
+#'
+#' @return A character string representing the namespace attribute of a vector
+#'   constructed with [defined()]. Returns the updated object (in setter forms).
+#'
+#' @examples
+#' # Define a vector with a namespace
+#' x <- defined("Q42", namespace = c(wd = "https://www.wikidata.org/wiki/"))
+#'
+#' # Get the namespace
+#' var_namespace(x)
+#' get_variable_namespaces(x)
+#'
+#' # Set the namespace
+#' var_namespace(x) <- "https://example.org/ns/"
+#'
+#' # Remove the namespace
+#' var_namespace(x) <- NULL
+#'
+#' # Use lower-level helpers (not typically used directly)
+#' namespace_attribute(x)
+#' namespace_attribute(x) <- "https://example.org/custom/"
+#'
 #' @family defined metadata methods and functions
-#' @return The namespace attribute of a vector constructed with[defined()].
+#' @rdname var_namespace
 #' @export
+
 var_namespace <- function(x, ...) {
   # rlang::check_dots_used()
   UseMethod("var_namespace")
