### 2025-08-07 e7aee48
Edge cases of namespace handling, new test suite, rOpenSci documentation
