+++ man/publication_year.Rd
@@ -3,40 +3,38 @@
 \name{publication_year}
 \alias{publication_year}
 \alias{publication_year<-}
-\title{Get/set the publication_year of the object.}
+\title{Get or Set the Publication Year of a Dataset Object}
 \usage{
 publication_year(x)
 
 publication_year(x, overwrite = TRUE) <- value
 }
 \arguments{
-\item{x}{A semantically rich data frame object created by
-\code{\link[=dataset_df]{dataset_df()}} or \code{dataset::\link{as_dataset_df}}.}
+\item{x}{A dataset object created by \code{\link[=dataset_df]{dataset_df()}} or
+\code{dataset::as_dataset_df()}.}
 
-\item{overwrite}{If the attributes should be overwritten. In case it is set
-to \code{FALSE}, it gives a message with the current \code{PublicationYear}
-property instead of overwriting it. Defaults to \code{TRUE} when the
-attribute is set to \code{value} regardless of previous setting.}
+\item{overwrite}{Logical. If \code{TRUE} (default), the existing
+\code{publication_year} attribute is replaced with \code{value}. If \code{FALSE}, the
+function returns a message and does not overwrite the existing value.}
 
-\item{value}{The publication_year as a character set.}
+\item{value}{A character string specifying the publication year.}
 }
 \value{
-Returns the \code{year} metadata field of the \code{DataBibentry} of
-the dataset
+The \code{publication_year} attribute as a character string.
 }
 \description{
-Get/set the optional \code{publication_year} property as an
-attribute to an R object.
+Access or assign the optional \code{publication_year} attribute to a semantically
+rich dataset object.
 }
 \details{
-The \code{PublicationYear} is the year when the data was or will be
-made publicly available in \code{YYYY} format. See
-\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties#publicationyearadditional-guidance}{Publication
-Year: DataCite Additional Guidance}.
+The \code{publication_year} represents the year when the dataset was or will be
+made publicly available, in \code{YYYY} format. For additional context, see
+\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties#publicationyearadditional-guidance}{DataCite: Publication Year – Additional Guidance}.
 }
 \examples{
 publication_year(orange_df)
-publication_year(orange_df) <- 1998
+publication_year(orange_df) <- "1998"
+
 }
 \seealso{
 Other Reference metadata functions: 
