+++ man/var_namespace.Rd
@@ -8,7 +8,7 @@
 \alias{get_namespace_attribute}
 \alias{set_namespace_attribute}
 \alias{namespace_attribute<-}
-\title{Get / Set a namespace of measure}
+\title{Get or Set the Namespace of a Variable}
 \usage{
 var_namespace(x, ...)
 
@@ -25,40 +25,57 @@ set_namespace_attribute(x, value)
 namespace_attribute(x) <- value
 }
 \arguments{
-\item{x}{a vector}
+\item{x}{A vector.}
 
-\item{...}{Further potential parameters reserved for inherited classes.}
+\item{...}{Additional arguments for method compatibility with other classes.}
 
-\item{value}{a character string or \code{NULL} to remove the namespace of measure.}
+\item{value}{A character string specifying the namespace, or \code{NULL} to
+remove it.}
 }
 \value{
-The namespace attribute of a vector constructed with\code{\link[=defined]{defined()}}.
+A character string representing the namespace attribute of a vector
+constructed with \code{\link[=defined]{defined()}}. Returns the updated object (in setter forms).
 }
 \description{
-Retain the namespace part of a permanent, global variable identifier which
-is independent of the R instance in use.
+Retrieve or assign the namespace part of a permanent, global variable
+identifier, independent of the current R session or instance.
 }
 \details{
-The namespace attribute is useful when users join or concatenate data from
-remote, linked, and open data sources. In such cases, variable identifiers
-(labels or names) are often resolved with a common namespace prefix, which,
-together with the namespace, forms a URI or IRI permanent identifier for the
-variable. Retaining the namespace in such cases allows cross-validation or
-success later updates of the vector (as a column of a dataset.)
+The \code{namespace} attribute is useful when working with remote, linked, or
+open data sources. Variable identifiers in such datasets are often qualified
+with a common namespace prefix. When combined, the prefix and namespace form
+a persistent URI or IRI for the variable.
 
-\code{get_variable_namespaces()} is identical to \code{var_namespace()}.
-See \code{vignette("defined", package = "dataset")} to use comprehensively
-with variable labels, namespaces, units of measures, and machine-independent
-permanent variable identifiers.
+Retaining the namespace ensures the identifiers remain valid and resolvable
+during validation, merging, or future updates of the vector (such as when it
+is used as a column in a dataset).
+
+\code{get_variable_namespaces()} is an alias for \code{var_namespace()}.
+\code{namespace_attribute()} and \code{set_namespace_attribute()} are internal helpers.
+
+For full usage, see:
+\code{vignette("defined", package = "dataset")} — demonstrating integration of
+variable labels, namespaces, units of measure, and machine-independent
+identifiers.
 }
 \examples{
-qid <- defined(c("Q275912", "Q116196078"),
-  namespace = c(wd = "https://www.wikidata.org/wiki/")
-)
-var_namespace(qid)
+# Define a vector with a namespace
+x <- defined("Q42", namespace = c(wd = "https://www.wikidata.org/wiki/"))
 
-# To remove a namespace
-var_namespace(qid) <- NULL
+# Get the namespace
+var_namespace(x)
+get_variable_namespaces(x)
+
+# Set the namespace
+var_namespace(x) <- "https://example.org/ns/"
+
+# Remove the namespace
+var_namespace(x) <- NULL
+
+# Use lower-level helpers (not typically used directly)
+namespace_attribute(x)
+namespace_attribute(x) <- "https://example.org/custom/"
+
 }
 \seealso{
 Other defined metadata methods and functions: 
