+++ tests/testthat/test-language.R
@@ -56,4 +56,3 @@ test_that("language() rejects non-dataset_df input", {
     regexp = "language\\(x\\)<- value: x must be a dataset object"
   )
 })
-
