### 2025-08-08 930cbc4
1. Adding a contributor() helper with tests 2. Debugging datacite metadata conversion
