+++ tests/testthat/test-datacite.R
@@ -1,145 +1,83 @@
-test_that("datacite() works", {
-  iris_datacite <- datacite(
-    Title = "Iris Dataset",
-    Creator = person(family = "Anderson", given = "Edgar", role = "aut"),
-    Publisher = "American Iris Society",
-    PublicationYear = 1935,
-    Geolocation = "US",
+test_that("datacite() creates a datacite object with minimal input", {
+  dc <- datacite(
+    Title = "A Simple Dataset",
+    Creator = person("John", "Doe", role = "cre"),
+    Publisher = "Test Publisher",
+    Date = "2024-01-01",
     Language = "en"
   )
-  expect_true(is.datacite(iris_datacite))
-  expect_equal(iris_datacite$language, "en")
-  expect_equal(iris_datacite$geolocation, "US")
-  expect_equal(iris_datacite$rights, ":tba")
-  expect_equal(iris_datacite$subject, "data sets")
+
+  expect_s3_class(dc, "datacite")
+  expect_s3_class(dc, "bibentry")
+  expect_equal(dc$title, "A Simple Dataset")
+  expect_equal(dc$publisher, "Test Publisher")
+  expect_equal(dc$year, "2024")
+  expect_equal(dc$language, "en")
+  expect_equal(dc$format, ":tba")
 })
 
-test_that("as_datacite() works", {
-  expect_type(as_datacite(iris_dataset, "list"), "list")
-  expect_true(is.dataset_df(as_datacite(iris_dataset, "dataset_df")))
-  expect_warning(as_datacite(x = iris_dataset, type = "data.frame"))
-  expect_equal(as_datacite(iris_dataset, "list")$FundingReference, ":unas")
-  expect_equal(as_datacite(iris_dataset, "list")$Description, "The famous (Fisher's or Anderson's) iris data set.")
-})
-
-test_that("datacite() constructor works with structured person objects", {
-  dc_obj <- datacite(
-    Title = "Growth of Orange Trees",
-    Creator = c(
-      person(
-        given = "N.R.",
-        family = "Draper",
-        role = "cre",
-        comment = c(VIAF = "http://viaf.org/viaf/84585260")
-      ),
-      person(
-        given = "H",
-        family = "Smith",
-        role = "cre"
-      )
-    ),
-    Contributor = person(
-      given = "Daniel",
-      family = "Antal",
-      role = "dtm"
-    ),
-    Publisher = "Wiley",
-    Identifier = "https://doi.org/10.5281/zenodo.14917851",
-    PublicationYear = 1998,
-    Language = "en",
-    Format = "application/r-rds",
-    Rights = "CC-BY",
-    Description = "Tree growth dataset over years."
-  )
-
-  expect_s3_class(dc_obj, "datacite")
-  expect_s3_class(dc_obj, "bibentry")
-
-  expect_equal(dc_obj$title, "Growth of Orange Trees")
-
-  expect_equal(dc_obj$author, c(
-    person(given = "N.R.", family = "Draper", role = "cre", comment = c(VIAF = "http://viaf.org/viaf/84585260")),
-    person(given = "H", family = "Smith", role = "cre")
-  ))
-
-  expect_equal(dc_obj$publisher, "Wiley")
-  expect_equal(dc_obj$identifier, "https://doi.org/10.5281/zenodo.14917851")
-  expect_equal(dc_obj$year, "1998")
-  expect_equal(dc_obj$language, "en")
-  expect_equal(dc_obj$format, "application/r-rds")
-  expect_equal(dc_obj$rights, "CC-BY")
-  expect_equal(dc_obj$description, "Tree growth dataset over years.")
-
-  # Contributor should be stored as an attribute (structured)
-  contributor_attr <- attr(dc_obj, "contributor")
[…]
