+++ NAMESPACE
@@ -49,6 +49,7 @@ S3method(xsd_convert,integer)
 S3method(xsd_convert,logical)
 S3method(xsd_convert,numeric)
 S3method(xsd_convert,tbl_df)
+export("contributor<-")
 export("creator<-")
 export("dataset_title<-")
 export("description<-")
@@ -75,6 +76,7 @@ export(as_factor)
 export(as_numeric)
 export(bibrecord)
 export(bind_defined_rows)
+export(contributor)
 export(creator)
 export(datacite)
 export(dataset_df)
