+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.96
-Date: 2025-08-07
+Version: 0.3.97
+Date: 2025-08-08
 Language: en-GB
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 

[…]
