### 2025-08-08 38082a3
structured subjects bug fix
