+++ tests/testthat/test-subject.R
@@ -1,3 +1,52 @@
+
+test_that("subject() errors for non-dataset_df", {
+  expect_error(
+    subject(mtcars),
+    "subject\\(x\\): x must be a dataset"
+  )
+})
+
+test_that("subject() returns from attribute when present", {
+  df <- dataset_df(data.frame(x = 1:3))
+  subject(df) <- subject_create("Biodiversity")
+  s <- subject(df)
+  expect_s3_class(s, "subject")
+  expect_equal(s$term, "Biodiversity")
+})
+
+test_that("subject() falls back to bibentry$subject when attr is missing", {
+  df <- dataset_df(data.frame(x = 1:3))
+  # set via character → bibentry$subject set to term + attr(subject) set
+  subject(df) <- "Forestry"
+  # remove attr to force fallback
+  attr(df, "subject") <- NULL
+  got <- subject(df)
+  expect_equal(got, "Forestry")
+})
+
+test_that("subject() returns default subject for a fresh dataset_df", {
+  df <- dataset_df(data.frame(x = 1:3))
+  s <- subject(df)
+  expect_s3_class(s, "subject")
+  expect_equal(s$term, default_subject$term)  # e.g., "Data sets"
+})
+
+
+test_that("subject() messages and returns NULL when subject truly missing", {
+  df <- dataset_df(data.frame(x = 1:3))
+  # Nuke both attr and bibentry entry to simulate a broken object
+  attr(df, "subject") <- NULL
+  be <- get_bibentry(df)
+  be$subject <- NULL
+  attr(df, "dataset_bibentry") <- be
+
+  expect_message(out <- subject(df), "No subject is recorded\\.")
+  expect_null(out)
+})
+
+
+# --- subject_create / new_Subject ---
+
 test_that("subject_create() creates a valid subject object", {
   s <- subject_create(
     term = "Climate Data",
@@ -6,7 +55,6 @@ test_that("subject_create() creates a valid subject object", {
     valueURI = "https://id.loc.gov/authorities/subjects/sh2001000174",
     prefix = "lcsh:"
   )
-
   expect_s3_class(s, "subject")
   expect_equal(s$term, "Climate Data")
   expect_equal(s$subjectScheme, "LCSH")
@@ -15,7 +63,7 @@ test_that("subject_create() creates a valid subject object", {
   expect_equal(s$prefix, "lcsh:")
 })
 
-test_that("subject_create() handles NULL values gracefully", {
+test_that("subject_create() handles NULL term gracefully", {
   s <- subject_create(term = NULL)
   expect_s3_class(s, "subject")
   expect_equal(s$term, ":tba")
@@ -31,8 +79,7 @@ test_that("subject_create() handles multiple terms", {
   expect_equal(s[[2]]$term, "Agriculture")
 })
 
-test_that("new_Subject() constructs subject with and
-          without classificationCode", {
+test_that("new_Subject() constructs with/without classificationCode", {
   s1 <- new_Subject("Environment")
   expect_s3_class(s1, "subject")
   expect_equal(s1$term, "Environment")
@@ -42,44 +89,26 @@ test_that("new_Subject() constructs subject with and
   expect_equal(s2$classificationCode, "06")
 })
 
+# --- is.subject ---
+
 test_that("is.subject() returns TRUE for subject objects", {
   s <- subject_create("Health")
   expect_true(is.subject(s))
 })
 
-test_that("is.subject() returns FALSE for non-subject objects", {
+test_that("is.subject() returns FALSE for non-subject", {
   expect_false(is.subject(list(term = "NotClassed")))
   expect_false(is.subject("Just a string"))
 })
 
[…]

[…]
