### 2025-08-08 24ede96
formerly internal set_var_labels is var_labels, exported, and works similar to labelled::labelled.
