+++ tests/testthat/test-fix_publisher.R
@@ -5,9 +5,9 @@ test_that("fix_publisher() works", {
     fix_publisher(publishers = person("Jane", "Doe", role = "pbl")),
     "Jane Doe"
   )
-  expect_equal(fix_publisher(publishers = person("American Iris Society", role = "pbl")), "American Iris Society")
+  expect_equal(fix_publisher(publishers = person("Open Data Institute", role = "pbl")), "Open Data Institute")
   expect_equal(fix_publisher(publishers = c(
     person("Jane", "Doe", role = "pbl"),
-    person("American Iris Society")
-  )), "{Jane Doe} and {American Iris Society}")
+    person("Open Data Institute")
+  )), "{Jane Doe} and {Open Data Institute}")
 })
