### 2025-08-08 817d3f2
For comparability, the removed vignettes are (with .Rbuildignore) present in the vignettes folder with a .bak extension. The iris_dataset.bak is also ignored, just for comparability.
