+++ vignettes/Motivation.Rmd
@@ -1,476 +1,0 @@
----
-title: "Motivation"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Motivation}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-if (!requireNamespace("dplyr", quietly = TRUE) || !requireNamespace("tidyr", quietly = TRUE)) {
-  stop("Please install 'dplyr' and 'tidyr' to run this vignette.")
-}
-
-library(dplyr)
-library(tidyr)
-```
-
-> "A dataset is an identifiable collection of data available for access or download in one or more formats." ISO/IEC 20546
-
-```{r setup}
-library(dataset)
-```
-
-The `dataset` package is designed to enhance the **semantic richness of R datasets**, particularly those structured as `data.frame` or `tibble` objects. Its goal is to promote reusability and interoperability by enabling tidy datasets to carry rich metadata — from the start.
-
-Through practical iterations, it became evident that the structure of a dataset cannot be separated from its **purpose**. As a result, the `dataset` package has evolved into a family of interrelated tools that support semantic clarity at both the column and dataset level.
-
-## Problem statement
-
-```{r example-ambiguity}
-data.frame(
-  geo = c("LI", "SM"),
-  CPI = c("0.8", "0.9"),
-  GNI = c("8976", "9672")
-)
-```
-
-At first glance, this dataset appears tidy — columns are variables, rows are observations. But is it interpretable?
-
--   `geo`: This might refer to geography — but is "LI" Lichtenstein? What about "SM"? Is this ISO alpha-2? Eurostat’s codes? World Bank’s?
-
--   `CPI`: Is this the Consumer Price Index, or the Corruption Perceptions Index? Something else entirely?
-
--   `GNI`: It could mean Gross National Income or Global Nutrition Index — and even if it's GNI, is it measured in dollars, euros, or something else?
-
-A tidy structure doesn't guarantee **semantic clarity**. Without metadata, every column is open to misinterpretation.
-
-```{r smallcountrydataset}
-options(sciphen = 4)
-small_country_dataset <- dataset_df(
-  country_name = defined(
-    c("AD", "LI"),
-    concept = "http://data.europa.eu/bna/c_6c2bb82d",
-    namespace = "https://www.geonames.org/countries/$1/"
-  ),
-  gdp = defined(
-    c(3897, 7365),
-    label = "Gross Domestic Product",
-    unit = "million dollars",
-    concept = "http://data.europa.eu/83i/aa/GDP"
-  ),
-  population = defined(
-    c(77543, 40015),
-    label = "Population",
-    concept = "http://data.europa.eu/bna/c_f2b50efd"
-  ),
-  dataset_bibentry = dublincore(
-    creator = person(given = "Jane", family = "Doe"),
-    title = "Small Country Dataset",
-    publisher = "Reprex"
-  )
-)
-```
-
-```{r percapita}
-small_country_dataset$gdp_capita <- defined(
-  small_country_dataset$gdp * 1e6 / small_country_dataset$population,
-  unit = "dollar",
-  label = "GDP Per Capita"
-)
-```
-
-The **interoperability** and **future reusability** of data depends on the quality and presence of metadata — not just the structure of the data itself. The `dataset` package captures metadata as seamlessly and non-intrusively as possible.
-
-It covers:
-
--   [x] `Descriptive metadata`: Follows DCTERMS and DataCite standards to describe datasets for searchability, citation, and sharing.
-
[…]
