### 2025-08-08 #681 @antaldaniel
@maurolepore I removed all instances of `iris_df` from the tests, too, practically creating a new test suit that also fits better your requirements stated in ml19-ml21. This resulted capturing a few edge cases, for example, structured subject list issues, etc. So generally the entire package works better.

No real new functionality was added, but for semantic clarity, I added to not super important helper functions,
- `dataset_format()` [not to confuse with base R format()]
- - `contributor()`,  a wrapper to person("Jane Doe", role="ctb") type people added to the bibliography, not to be confused with role = 'aut' roles. I separated them because `utils::person()` treats them similarly but the metadata schemas do not.

- For comparability, the removed vignettes are (with .Rbuildignore) present in the vignettes folder with a .bak extension. These are now the new [An Introduction to the dataset Package](https://dataset.dataobservatory.eu/articles/articles/introduction.html) and [Design Principles & Future Work Semantically Enriched, Standards-Aligned Datasets in R](https://dataset.dataobservatory.eu/articles/articles/design.html) articles.
- The `iris_dataset.bak` is also ignored, and stays there for now for for comparability.

I think that this was a bigger revision than I expected but eventually I think it resulted in a much more robust code.
