### 2025-08-08 2744024
Add examples to @examples
