### 2025-08-10 #681 @maurolepore
@antaldaniel, thanks so much for your attention to detail. It really helped.

I noticed the issue was closed, maybe by accident. I reopened it, and it'll be automatically closed once we approve the submission. The package has indeed improved a lot, and it seems we're not far from approval.

Below are a few of my requested changes that seem to still need a little attention.

----

### ml07

> ml07: Removed the separate introduction page and now rely on the default pkgdown dropdown structure ...

But did you do that in a branch other than the default `main` branch?

On main I see this not-so-simple [_pkgdown.yml](https://github.com/dataobservatory-eu/dataset/blob/0e72f6fbcf36ef428afb48982f3c0c343635ed55/_pkgdown.yml#L23-L52).

And the corresponding website doesn't show an Articles dropdown but instead an Articles page:

### ml22 (optional)

- See https://github.com/dataobservatory-eu/dataset/pull/38

### ml23

Thanks, most styler issues are solved but notice the few remaining ones:

```r
Styling  100  files:
 ...
 tests/spelling.R                            ℹ
 ...
 data-raw/prepare_statwales.R                ✔
 vignettes/bibrecord.Rmd                     ✔
 vignettes/dataset_df.Rmd                    ⚠
 vignettes/defined.Rmd                       ⚠
 vignettes/design.Rmd                        ✔
...
───────────────────────────────────────────
Status	Count	Legend
✔ 	97	File unchanged.
ℹ 	1	File changed.
✖ 	2	Styling threw an error.
───────────────────────────────────────────
Please review the changes carefully!

✔ Styled project according to the tidyverse style guide.
Warning messages:
1: When processing vignettes/dataset_df.Rmd: ℹ In index: 10.
Caused by error in `parse_safely()`:
! <text>:2:7: unexpected 'in'
1: rbind(small_country_dataset,feroe_df)
2: Error in
         ^
2: When processing vignettes/defined.Rmd: ℹ In index: 9.
Caused by error in `parse_safely()`:
! <text>:2:7: unexpected 'in'
1: c(gdp_1, gdp_2)
2: Error in
         ^
```

### ml26

Please review once again the most recent output of pkgcheck. Ensure to expand the drop downs. There is a lot of useful information.

On GitHub Actions I see the pkgcheck workflow is failing.

Also I see the R-hub workflow fails, for what seems to be a missing system dependency. You may update the workflow or remove it since the package seems well covered by other workflows.

https://github.com/dataobservatory-eu/dataset/actions/runs/16829920158/job/47674886702#step:5:3117

```
  The following packages have unmet dependencies:
  libraptor2-dev : Depends: libcurl4-gnutls-dev but it is not installable
  E: Unable to correct problems, you have held broken packages.
```
