### 2025-08-13 52e18af
rewrite vignette formatting
