+++ vignettes/dataset_df.Rmd
@@ -7,39 +7,62 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r, include = FALSE}
+```{r setupvignette, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 ```
-```{r setup}
+
+The `dataset` package extends R's native data structures with machine-readable metadata. It follows a *semantic early-binding* approach, which means metadata is embedded as soon as the data is created, making datasets suitable for long-term reuse, FAIR-compliant publishing, and integration into semantic web systems.
+
+In the abstract, and following ISO/IEC 20546, a dataset is an identifiable collection of data available for access or download in one or more formats. This is different from R’s `data.frame` object, which is one specific in-memory structure. While R can already serialise data frames to formats like `.rds`, `.rda`, or `.csv`, these serialisations lack rich, standardised metadata. The `dataset` package aims to bridge that gap by aligning with established metadata standards to produce serialisations that are easier to reuse and interpret.
+
+`dataset_df` works naturally with data structured according to tidy data principles (Wickham, 2014), where each variable is a column, each observation is a row, and each type of observational unit forms a table. It adds an additional semantic layer to these tidy structures, so the meaning of variables and datasets is explicit and machine-readable.
+
+This vignette focuses specifically on the `dataset_df` class and the `dataset_df()` constructor, which you can use to create a semantically enriched data frame. For details on semantically enriched vectors, see `vignette("defined", package = "dataset")`.
+
+## Purpose
+
+The `dataset_df()` function helps you create **semantically rich datasets** that are easier to:
+
+-   understand by humans,
+
+-   validate and process by machines,
+
+-   share across tools, teams, and domains.
+
+This vignette walks you through creating such a dataset using a subset of the *GDP and main aggregates – international data cooperation annual data* dataset from Eurostat\
+(DOI: [https://doi.org/10.2908/NAIDA_10_GDP).](https://doi.org/10.2908/NAIDA_10_GDP).)
+
+## Load example data
+
+```{r loaddata}
 library(dataset)
 data("gdp")
 ```
-
-The `dataset_df()` function helps create semantically rich datasets that are easier to:
-
-- understand by humans,
-- validate and process by machines,
-- and share across tools, teams, and domains.
-
-This vignette demonstrates how to construct a semantically enriched dataset using a subset of the  *GDP and main aggregates - international data cooperation annual data* dataset created by Eurostat (DOI: <https://doi.org/10.2908/NAIDA_10_GDP>).
-
 
 ```{r printgdp}
 print(gdp)
 ```
 
-While the raw dataset is valid, it can be hard to interpret without external documentation. For example:
-- Countries are encoded in the `geo` variable.
-- Reporting frequency (e.g., `A` for annual) is stored in `freq`
+This example dataset is already in tidy format: each row represents a single observation for a country and year, and each column is a variable. `dataset_df` builds on this structure by adding semantic information to the variables and the dataset itself, ensuring that both the shape and the meaning of the data are preserved and unambiguous.
 
-## Getting Started
+While the raw dataset represented in the `gdp` data.frame is valid and tidy, it can be hard to interpret without external documentation. For example:
 
[…]
