+++ _pkgdown.yml
@@ -19,37 +19,36 @@ authors:
     html: <img src="https://reprex.nl/authors/admin/avatar_hud66277e7f64fbf44513e7a4881e2a94f_35557_270x270_fill_lanczos_center_2.png"
       height="24" />
 
+navbar:
+  structure:
+    left: [reference, articles]
+    right: [search, github]
+  components:
+    reference:
+      text: Reference
+      href: reference/index.html
+    articles:
+      text: Articles
+      href: articles/index.html
+    github:
+      icon: fab fa-github
+      href: https://github.com/dataobservatory-eu/dataset/
 
 articles:
   - title: "Introduction"
-    desc: >
-     The dataset package helps R users manage and preserve metadata within
-        tidy datasets, ensuring clarity and semantic consistency even after
-        transformations like slicing or joining. It enriches vectors and
-        data frames with standardised metadata using DCTERMS and DataCite,
-        and supports workflows toward RDF-compatible formats. This ensures
-        data remains both human- and machine-readable across statistical and
-        semantic web contexts.
     contents:
-       - introduction
-       - design
-  - title: "Semantic Enrichment of Variables"
-    desc: >
-     Enriching metadata about individual variables stored as defined vectors.
+      - introduction
+      - design
+  - title: "Core concepts"
     contents:
-       - defined
-  - title: "Semantic Enrichment of Datasets"
-    desc: >
-     Enriching metadata about the dataset as a whole, potential extensions.
+      - defined
+      - dataset_df
+  - title: "Metadata & citation"
     contents:
-       - dataset_df
-       - bibrecord
+      - bibrecord
   - title: "Dataset exchanges"
-    desc: >
-     Making the richer datasets available for other users outside of the
-         R ecosystem.
     contents:
-       - rdf
+      - rdf
 
 reference:
   - title: "Dataset"
