### 2025-08-13 #681 @antaldaniel
@maurolepore Thank you for your patience.

1. I accidentally closed the thread.
2. `ml23` I fixed the styler issue, it was related to the mimicked 'errors' in not evaluated code to show what will not work.  It is now fenced in a way that it does not see it as a style breach, but I think it was a false positive in the two vignettes.
3. `ml07` Thanks for reminding me to the chaotic navbar, indeed, it looks way better. It was an oversight on my part, but I re-read now the rOpenSci style-guide, and tried to align even closer all vignettes. So they are changed in outlook, captioning, and in style. (You did not mention it now, but reviewing how it should look, I wanted to improve it.)
4. `ml26` Thank you for the pull request and the help with actions, that was really very helpful.
5. I also did the the standard rhubv2 checks on many platforms, and I realised that it is failing on ubuntu-release.  It is actually the suggested dependency from rOpenSci, rdflib, that triggered the problem.  rdflib pulls in redland → libraptor2-dev → libcurl4-gnutls-dev. On Ubuntu 22.04, CRAN builds with libcurl4-openssl-dev. This is specifically an rhub issue, now the new actions seems to work fine.
6. I started writing issues to a new release, but they do not affect the current functionality, they will improve NTriples serialisation. Eventually I can also imagine that I will take out the NTriples serialisation to a mini-package for easier maintenance.  I wanted to avoid such a heavy dependency as rdflib, a package that I love but it may be too intimidating for many users who are unfamiliar with RDF. But this is a marginal future development that do not affect the core functionality of the package.

Sincerely thank you again for your patience, I added you as a reviewer to the package.
