### 2025-08-14 dataobservatory-eu/dataset#39 @schneiderpy
Issue with ggplot2 functions
Hi Daniel

It seems there is (or I have) an issue with `ggplot2()` functions:

Here is an error that occured with a dataset, that I made first with frictionless. This dataset was used to enhence meta data with the `dataset` pkg.
I use these new datasets in my R pkg. While creating some examples I encountered an issue, which made my doubt about the use of the dataset pkg (or I don't know how to use it).
Here the error:

```
Error:
! Failed to evaluate glue component {label}
Caused by error in `as.character()`:
! Can't convert `x` <haven_labelled_defined> to <character>.
```

I wanted to convert a variable (dates as character) to date format to use with ggplot2, but I seems that ggplot2 does not support "haven-labelled" variables, which makes the use of the data - more or less - useless.
