### 2025-08-16 8cb2909
Article-groups now appear under a navbar menu
