### 2025-08-16 #681 @maurolepore
- ml07. Check. Thanks! Looks much cleaner to me. Optionally see https://github.com/dataobservatory-eu/dataset/pull/40
- ml22. Check. Thanks! I see this PR is now merged: https://github.com/dataobservatory-eu/dataset/pull/38
- ml23. Check. Thanks! I re-run the styler and I see no issue.
- ml26 Thanks! pkgcheck output ... (sorry need to run to catch a flight. I'll continue later)
