### 2025-08-17 #681 @maurolepore
@antaldaniel thanks again for your response and sorry mixing our conversation with output from the bot. Just in case, and to make sure you don't miss it, here I link my latest response:

https://github.com/ropensci/software-review/issues/681#issuecomment-3193556008

We're so close!
