### 2025-08-18 e4b6e8f
xsd_convert can handle short, Turtle-like, and long, NTriple-like URIs for RDF serialisation
