+++ R/xsd_convert.R
@@ -2,8 +2,8 @@
 #'
 #' Converts R vectors, data frames, and `dataset_df` objects to
 #' [XML Schema Definition (XSD)](https://www.w3.org/TR/xmlschema11-2/)
-#' compatible string representations such as `xs:decimal`, `xs:boolean`,
-#' `xs:date`, and `xs:dateTime`.
+#' compatible string representations such as `xsd:decimal`, `xsd:boolean`,
+#' `xsd:date`, and `xsd:dateTime`.
 #'
 #' This is primarily used for generating RDF-compatible typed literals.
 #'
@@ -17,11 +17,14 @@
 #' @param x An object (vector, data frame, tibble, or `dataset_df`).
 #' @param idcol Column name or position to use as row (observation) identifier.
 #'   If `NULL`, row names are used.
+#' @param shortform Logical. If `TRUE` (default), datatypes are abbreviated with
+#'   the `xsd:` prefix (e.g. `"42"^^<xsd:integer>`). If `FALSE`, datatypes are
+#'   expanded to full URIs (e.g.
+#'   `"42"^^<http://www.w3.org/2001/XMLSchema#integer>`).
 #' @param ... Additional arguments passed to methods.
 #'
-#' @return
-#' A character vector or data frame with values serialized as XSD-compatible
-#' RDF literals.
+#' @return A character vector or data frame with values serialized as
+#' XSD-compatible RDF literals.
 #'
 #' @examples
 #' # Simple data frame with mixed types
@@ -32,47 +35,75 @@
 #'   date   = as.Date(c("2020-01-01", "2020-12-31"))
 #' )
 #'
+#' # Short vs long-form URI:
+#' xsd_convert(120L, shortform = TRUE)
+#' xsd_convert(121L, shortform = FALSE)
+#'
 #' @section Class-specific examples:
 #' ```r
-#' xsd_convert(42L)                   # integer -> xs:integer
-#' xsd_convert(c(TRUE, FALSE, NA))    # logical -> xs:boolean
-#' xsd_convert(Sys.Date())            # Date -> xs:date
-#' xsd_convert(Sys.time())            # POSIXct -> xs:dateTime
-#' xsd_convert(factor("apple"))       # factor -> xs:string
-#' xsd_convert(c("apple", "banana"))  # character -> xs:string
+#' xsd_convert(42L)                   # integer -> xsd:integer
+#' xsd_convert(c(TRUE, FALSE, NA))    # logical -> xsd:boolean
+#' xsd_convert(Sys.Date())            # Date -> xsd:date
+#' xsd_convert(Sys.time())            # POSIXct -> xsd:dateTime
+#' xsd_convert(factor("apple"))       # factor -> xsd:string
+#' xsd_convert(c("apple", "banana"))  # character -> xsd:string
 #' ```
 #' @family RDF and linked data helpers
 #' @export
 
-xsd_convert <- function(x, idcol, ...) {
+xsd_convert <- function(x, idcol = NULL, shortform = TRUE, ...) {
   UseMethod("xsd_convert", x)
 }
 
+
 #' @keywords internal
-get_type <- function(t) {
-  if (any(class(t) %in% c("numeric", "double"))) {
-    type <- "xs:decimal"
+get_type <- function(t, shortform = TRUE) {
+  base_type <- if (any(class(t) %in% c("numeric", "double"))) {
+    "decimal"
   } else if (any(class(t) == "integer")) {
-    type <- "xs:integer"
+    "integer"
   } else if (any(class(t) %in% c("character", "factor"))) {
[…]
