+++ tests/testthat/test-xsd_convert.R
@@ -1,35 +1,49 @@
 ## get_type internal function -------------------------------------
-test_that("get_type returns correct XSD types", {
-  expect_equal(get_type(1.5), "xs:decimal")
-  expect_equal(get_type(1L), "xs:integer")
-  expect_equal(get_type("text"), "xs:string")
-  expect_equal(get_type(TRUE), "xs:boolean")
-  expect_equal(get_type(Sys.Date()), "xs:date")
-  expect_equal(get_type(Sys.time()), "xs:dateTime")
-  expect_equal(get_type(as.difftime(c("0:3:20", "11:23:15"))), "xs:duration")
+test_that("get_type returns correct XSD types (short and long)", {
+  expect_equal(get_type(1.5), "xsd:decimal")
+  expect_equal(get_type(1.5, shortform = FALSE), "http://www.w3.org/2001/XMLSchema#decimal")
+
+  expect_equal(get_type(1L), "xsd:integer")
+  expect_equal(get_type(1L, shortform = FALSE), "http://www.w3.org/2001/XMLSchema#integer")
+
+  expect_equal(get_type("text"), "xsd:string")
+  expect_equal(get_type("text", shortform = FALSE), "http://www.w3.org/2001/XMLSchema#string")
+
+  expect_equal(get_type(TRUE), "xsd:boolean")
+  expect_equal(get_type(TRUE, shortform = FALSE), "http://www.w3.org/2001/XMLSchema#boolean")
+
+  expect_equal(get_type(Sys.Date()), "xsd:date")
+  expect_equal(get_type(Sys.Date(), shortform = FALSE), "http://www.w3.org/2001/XMLSchema#date")
+
+  expect_equal(get_type(Sys.time()), "xsd:dateTime")
+  expect_equal(get_type(Sys.time(), shortform = FALSE), "http://www.w3.org/2001/XMLSchema#dateTime")
+
+  expect_equal(get_type(as.difftime(1, units="secs")), "xsd:duration")
+  expect_equal(get_type(as.difftime(1, units="secs"), shortform = FALSE), "http://www.w3.org/2001/XMLSchema#duration")
 })
+
 
 ## Character conversions ------------------------------------------
 test_that("xsd_convert.character handles standard strings", {
   expect_equal(
     xsd_convert(c("apple", " banana ", "cherry")),
-    c('"apple"^^<xs:string>', '" banana "^^<xs:string>', '"cherry"^^<xs:string>')
+    c('"apple"^^<xsd:string>', '" banana "^^<xsd:string>', '"cherry"^^<xsd:string>')
   )
 })
 
 test_that("xsd_convert.character handles NA values", {
   result <- xsd_convert(c("apple", NA, "cherry"))
-  expect_equal(result[1], '"apple"^^<xs:string>')
+  expect_equal(result[1], '"apple"^^<xsd:string>')
   expect_true(is.na(result[2]))
-  expect_equal(result[3], '"cherry"^^<xs:string>')
+  expect_equal(result[3], '"cherry"^^<xsd:string>')
 })
 
 test_that("xsd_convert.character handles empty strings", {
-  expect_equal(xsd_convert(""), '""^^<xs:string>')
+  expect_equal(xsd_convert(""), '""^^<xsd:string>')
 })
 
 test_that("xsd_convert.character handles empty vector", {
-  expect_equal(xsd_convert(character(0)), '""^^<xs:string>')
+  expect_equal(xsd_convert(character(0)), '""^^<xsd:string>')
 })
 
[…]

[…]
