### 2025-08-18 abc00ef
xs:decimal is abbreviated as xsd::decimal (more widely used in RDF circles)
