+++ man/strip_defined.Rd
@@ -21,7 +21,8 @@ gdp <- defined(c(3897L, 7365L), label = "GDP", unit = "million dollars")
 strip_defined(gdp)
 
 fruits <- defined(c("apple", "avocado", "kiwi"),
-  label = "Fruit", unit = "kg")
+  label = "Fruit", unit = "kg"
+)
 
 strip_defined(fruits)
 }
