### 2025-08-19 dataobservatory-eu/dataset#39 @schneiderpy
Thank you @antaldaniel. Now I have a plot .. (with version 0.3.99). However, I run into another issue (well I dont like the output because I think it's a kind of information loss).
I have a dataset like this:
```
head(py_temperature)
Schneider (2025): Weekly mean temperature data [dataset], https://doi.org/10.5281/zenodo.16729963
  rowid     country   ISO       city         week       avg_temp  holiday
  <defined> <defined> <defined> <defined>    <dttm_dfn> <defined> <defined>
1 obs:1     Paraguay  PY        1 [Asuncion] 2016-01-04 27.8      0
2 obs:2     Paraguay  PY        1 [Asuncion] 2016-01-11 30.3      0
3 obs:3     Paraguay  PY        1 [Asuncion] 2016-01-18 29.9      0
4 obs:4     Paraguay  PY        1 [Asuncion] 2016-01-25 27.3      1
5 obs:5     Paraguay  PY        1 [Asuncion] 2016-02-01 26.6      0
6 obs:6     Paraguay  PY        1 [Asuncion] 2016-02-08 30.1      0
plot <- py_temperature |> ggplot(aes(x = week, y = avg_temp)) + geom_line() + facet_wrap(~ city)
plot
```
As you can see `city` is a factor and with `defined` it has been "refactored". In the plot appear the numbers and not the city name(s). How can I revert this? (I tried as.character(city) but this didn't work)
