### 2025-08-20 dataobservatory-eu/dataset#39 @antaldaniel
Thank you very much for your continued interest. Here are two examples that allow to show the problem.  It is important to understand how you came to this result, but I guessed that somehow like this.

Original tibble, prints nicely with `py_df|> ggplot(aes(x = week, y = avg_temp)) + geom_line() + facet_wrap(~ city)`

```
py_df <- tibble::tibble(
  city     = factor(x = c(1, 2, 1, 2),
                    labels = c("Asuncion", "Leque")),
  week     = as.Date(rep(c("2016-01-04", "2016-01-11"), each = 2)),
  avg_temp = c(27.8, 25.1, 30.3, 26.4),
  holiday  = c(0L, 0L, 0L, 0L)
)
```
The semantically rich example that does not print nicely

```
py_temperature <- dataset_df(
  city = defined(py_df$city, label = "City"),
  week = defined(py_df$week, label = "Week"),
  avg_temp = defined(py_df$avg_temp,
                     label="Average Temperature",
                     unit = "Celsius"),
  identifier = c(obs = "http://example.com/dataset#py_temperature"),
  dataset_bibentry = datacite(Title = "Paraguay Temperature",
                              Creator = person(given="Jane", family="Doe"),
                              Identifier = "https://doi.org/example.134")
)
```

The quick fix is this:
`py_temperature |> ggplot(aes(x = week, y = avg_temp)) + geom_line() + facet_wrap(~ as_factor(city))
`

Also, with labels:

```
py_temperature |> ggplot(aes(x = week, y = avg_temp)) +
  geom_point() +
  geom_text( aes(label=as_factor(city)), nudge_x = 0.3)
```

The bigger fix will be to change the printing behavior of defined, which I will improve, but that is also a bit philosophical question, because in a way `labelled`  and `defined` replace the use of factors, and a character or double vector with labels is a more accurate version of a factor, unless they are `ordered`. I will take a look how labelled handles coercion to/from factor, and what are the downstream consequences of a printing change.

As a matter of fact I am actually happy that only this small tweek is needed, because of course, ggplot2 knows nothing about these new s3 classes and works reasonably well, still.
