### 2025-08-20 0eb239e
handling multiple subjects and relations
