### 2025-08-20 a552cb9
improved serialisation and test
