### 2025-08-20 af13b9d
Handing multiple subjects and relations in the constructors
