+++ R/datacite.R
@@ -131,9 +131,9 @@ datacite <- function(Title,
   AlternateIdentifier <- ifelse(is.null(AlternateIdentifier),
     ":unas", AlternateIdentifier
   )
-  RelatedIdentifier <- ifelse(is.null(RelatedIdentifier),
-    ":unas", RelatedIdentifier
-  )
+  if (is.null(RelatedIdentifier)) {
+    RelatedIdentifie <- ":unas"
+  }
   Rights <- ifelse(is.null(Rights), ":tba", as.character(Rights))
   Geolocation <- ifelse(is.null(Geolocation), ":unas",
     as.character(Geolocation)
@@ -185,6 +185,7 @@ new_datacite <- function(Title,
                          Description,
                          Geolocation,
                          FundingReference) {
+  # bibentry: subject slot must be character
   datacite_object <- bibentry(
     bibtype = "Misc",
     title = Title,
@@ -194,9 +195,15 @@ new_datacite <- function(Title,
     publisher = Publisher,
     date = Date,
     language = Language,
-    subject = Subject$term,
+    subject = if (is.subject(Subject)) Subject$term else as.character(Subject),
     alternateidentifier = AlternateIdentifier,
-    relatedidentifier = RelatedIdentifier,
+    relatedidentifier = if (is.related(RelatedIdentifier)) {
+      RelatedIdentifier$relatedIdentifier
+    } else if (is.null(RelatedIdentifier)) {
+      ":unas"
+    } else {
+      as.character(RelatedIdentifier)
+    },
     format = Format,
     version = Version,
     rights = Rights,
@@ -205,7 +212,17 @@ new_datacite <- function(Title,
     fundingreference = FundingReference
   )
 
-  # Store contributor as structured attribute
+  # keep the structured subject object as an attribute
+  if (!is.null(Subject) && is.subject(Subject)) {
+    attr(datacite_object, "subject") <- Subject
+  }
+
+  # keep the structured related item object as an attribute
+  if (!is.null(RelatedIdentifier) && is.related(RelatedIdentifier)) {
+    attr(datacite_object, "relatedIdentifier") <- RelatedIdentifier
+  }
+
+  # add structured contributor list
   if (!is.null(Contributor)) {
     attr(datacite_object, "contributor") <- Contributor
   }
@@ -231,132 +248,39 @@ is.datacite.datacite <- function(x) inherits(x, "datacite")
 print.datacite <- function(x, ...) {
   cat("DataCite Metadata Record\n")
   cat("--------------------------\n")
-  cat("Title:        ", x$title, "\n")
-  cat("Creator(s):   ", paste(format(x$author), collapse = "; "), "\n")
+
+  pr <- function(label, value) {
+    if (!is.null(value) && length(value) > 0 && any(nzchar(value))) {
+      # %-12s = left-justify label into 12 chars
+      cat(sprintf("%-12s %s\n", paste0(label, ":"), paste(value, collapse = "; ")))
+    }
+  }
+
+  pr("Title",        x$title)
+  pr("Creator(s)",  paste(format(x$author), collapse = "; "))
 
   contributor <- attr(x, "contributor")
   if (!is.null(contributor)) {
-    cat("Contributor(s):", fix_contributor(contributor), "\n")
+    pr("Contributor(s)", fix_contributor(contributor))
   }
 
-  if (!is.null(x$identifier)) cat("Identifier:   ", x$identifier, "\n")
-  if (!is.null(x$publisher)) cat("Publisher:    ", x$publisher, "\n")
[…]

[…]
