+++ man/language.Rd
@@ -99,6 +99,7 @@ Other bibliographic helper functions:
 \code{\link{get_bibentry}()},
 \code{\link{publication_year}()},
 \code{\link{publisher}()},
+\code{\link{relation}()},
 \code{\link{rights}()},
 \code{\link{subject}()}
 }
