+++ man/publication_year.Rd
@@ -47,6 +47,7 @@ Other bibliographic helper functions:
 \code{\link{get_bibentry}()},
 \code{\link{language}},
 \code{\link{publisher}()},
+\code{\link{relation}()},
 \code{\link{rights}()},
 \code{\link{subject}()}
 }
