### 2025-08-20 359cdfd
add unexported functions
