+++ NAMESPACE
@@ -116,6 +116,7 @@ export(publication_year)
 export(publisher)
 export(related_create)
 export(related_item)
+export(relation)
 export(rights)
 export(set_namespace_attribute)
 export(set_unit_attribute)
