### 2025-08-22 #681 @maurolepore
@antaldaniel thanks for checking. No problem from my end. I second what @mpadge said.

Also thanks for your wonderful responses here: https://github.com/dataobservatory-eu/dataset/issues/39#issue-3323379445

Finally thanks for following up on my latest comments. I'll reply shortly.
