+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.992
-Date: 2025-08-20
+Version: 0.4.0
+Date: 2025-08-22
 Language: en-GB
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 

[…]
