### 2025-08-22 7b7e79f
release candidate with fixed URLs and some stylistic changes in the introductory text.
