+++ vignettes/dataset_df.Rmd
@@ -14,23 +14,20 @@ knitr::opts_chunk$set(
 )
 ```
 
-The `dataset` package extends R's native data structures with machine-readable metadata. It follows a *semantic early-binding* approach, which means metadata is embedded as soon as the data is created, making datasets suitable for long-term reuse, FAIR-compliant publishing, and integration into semantic web systems.
+The `dataset` package extends R’s native data structures with machine-readable metadata. It follows a *semantic early-binding* approach: metadata is embedded as soon as the data is created, making datasets suitable for long-term reuse, FAIR-compliant publishing, and integration into semantic web systems.  
 
-In the abstract, and following ISO/IEC 20546, a dataset is an identifiable collection of data available for access or download in one or more formats. This is different from R’s `data.frame` object, which is one specific in-memory structure. While R can already serialise data frames to formats like `.rds`, `.rda`, or `.csv`, these serialisations lack rich, standardised metadata. The `dataset` package aims to bridge that gap by aligning with established metadata standards to produce serialisations that are easier to reuse and interpret.
+In R, a `data.frame` is defined as a tightly coupled collection of variables that share many of the properties of matrices and lists, and it serves as the fundamental data structure for most of R’s modeling software. Users of the R ecosystem often use the term *data frame* interchangeably with *dataset*. However, the standards used in libraries, repositories, and statistical systems for publishing, exchanging, and reusing datasets require metadata that even “tidy” data frames do not provide.  
 
-`dataset_df` works naturally with data structured according to tidy data principles (Wickham, 2014), where each variable is a column, each observation is a row, and each type of observational unit forms a table. It adds an additional semantic layer to these tidy structures, so the meaning of variables and datasets is explicit and machine-readable.
+This vignette introduces the `dataset_df` class and the `dataset_df()` constructor, which extend tidy data frames with a semantic layer. For details on semantically enriched vectors, see `vignette("defined", package = "dataset")`. Readers interested in the underlying ISO and W3C definitions of *dataset* will find them discussed in `vignette("design", package = "dataset")`.  
 
-This vignette focuses specifically on the `dataset_df` class and the `dataset_df()` constructor, which you can use to create a semantically enriched data frame. For details on semantically enriched vectors, see `vignette("defined", package = "dataset")`.
+## Purpose  
 
-## Purpose
+The `dataset_df()` function helps you create **semantically rich datasets** that meet the interoperability, exchange, and reuse requirements of libraries, repositories, and statistical systems. It defines a new S3 class, inherited from the modernised data frame of `tibble::tibble()`, that retains compatibility with existing workflows but is easier to:  
 
-The `dataset_df()` function helps you create **semantically rich datasets** that are easier to:
-
--   understand by humans,
-
--   validate and process by machines,
-
--   share across tools, teams, and domains.
+- understand by humans,  
+- validate and process by machines,  
+- deposit, exchange, and publish,  
+- share across tools, teams, and domains.
 
 This vignette walks you through creating such a dataset using a subset of the *GDP and main aggregates – international data cooperation annual data* dataset from Eurostat\
 (DOI: [https://doi.org/10.2908/NAIDA_10_GDP).](https://doi.org/10.2908/NAIDA_10_GDP).)
