### 2025-08-23 cf5814a
Wrong DOI formatting
