### 2025-08-23 7b38848
Create separate example files for RDF examples
