### 2025-08-23 9b5f110
Handling URIs as URLs for false negative CRAN checks.
