+++ vignettes/rdf.Rmd
@@ -1,57 +1,38 @@
 ---
 title: "From R to RDF"
-output: rmarkdown::html_vignette
+output:
+  rmarkdown::html_vignette:
+    md_extensions: -autolink_bare_uris
 vignette: >
   %\VignetteIndexEntry{From R to RDF}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setupvignette, include = FALSE}
+```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-
-if (!requireNamespace("rdflib", quietly = TRUE)) {
-  stop("Please install 'rdflib' to run this vignette.")
-}
-
-if (!requireNamespace("jsonld", quietly = TRUE)) {
-  stop("Please install 'jsonld' to run this vignette.")
-}
 ```
 
-```{css, echo = FALSE}
-    .small .table{
-      font-size: 10px;
-    }
+```{css, echo=FALSE}
+.smaller .table {
+  font-size: 11px;
+}
+
+.smaller pre,
+.smaller code {
+  font-size: 11px;
+  line-height: 1.2;
+}
 ```
-
-The `dataset` package supports the creation of semantically enriched datasets that align with the principles outlined in *design.Rmd*: preserving meaning across the data lifecycle, enabling interoperability, and ensuring compliance with modern metadata standards. A key part of these design principles is the ability to export data in machine-readable formats that integrate smoothly with the semantic web.
-
-One such format is the Resource Description Framework (RDF), which expresses data as **subject–predicate–object** triples. By converting R data into RDF, we make each observation and its attributes explicitly defined and linkable to external vocabularies. This allows datasets to be easily merged, queried, and interpreted across systems.
-
-## Purpose
-
-The goal of this vignette is to show how a tidy R dataset can be transformed into RDF triples while retaining all semantic annotations added via the `dataset` package. RDF enables:
-
-- **Interoperability** – linking data to controlled vocabularies and ontologies
-- **Precision** – carrying explicit definitions for every variable and observation
-- **Reusability** – producing outputs that are both human- and machine-readable
-
-Exporting to RDF is a natural extension of the semantic early-binding approach described in *design.Rmd*, where metadata is embedded at the moment of dataset creation. This ensures that when data leaves R, it already contains everything needed for FAIR (Findable, Accessible, Interoperable, Reusable) compliance.
 
 ## From tidy data to RDF triples
 
-A tidy dataset, where each variable is a column and each observation is a row, maps cleanly to RDF triples:
+This vignette demonstrates how to convert tidy R datasets into semantically enriched RDF triple structures, using the `dataset` and `rdflib` packages. These packages help you annotate variables with machine-readable concepts, units, and links to controlled vocabularies.
 
-- The **subject** represents an observation or entity (e.g., “Observation #1” or a URI for a country)
-- The **predicate** represents a variable or attribute (e.g., GDP value, reference year, measurement unit)
-- The **object** represents the value, which can be a literal (number, string, date) or a URI pointing to a definition
-
-By structuring data this way, every cell in the dataset becomes an elementary semantic statement that can be validated, linked, and queried via SPARQL.
-
+We’ll start with a small example of a tidy dataset representing countries (`geo`) with unique identifiers (`rowid`) and then show how to transform the dataset into RDF triples using standard vocabularies.
 
 ```{r setup}
 library(dataset)
@@ -61,103 +42,95 @@ data("gdp")
 
[…]
