### 2025-08-23 60a9c56
Handling URIs as URLs for false negative CRAN checks.
