### 2025-08-23 a538b9c
Change the default observation prefix so that CoolURIs work.
