+++ tests/testthat/test-dataset_df.R
@@ -12,7 +12,7 @@ test_that("dataset_df() respects identifier for rowid creation", {
     identifier = c(mt = "http://mtcars.com/dataset#")
   )
   expect_s3_class(df$rowid, "haven_labelled_defined")
-  expect_true(all(grepl("^mt:", df$rowid)))
+  expect_true(all(grepl("^mt", df$rowid)))
 })
 
 
@@ -97,8 +97,8 @@ test_that("print() outputs full APA-style citation and data", {
     )
   )
   expect_output(print(df), "Smith \\(2021\\): Tiny Data \\[dataset\\], https://doi.org/10.1234/example")
-  expect_output(print(df), "1 eg:1\\s+1")
-  expect_output(print(df), "2 eg:2\\s+2")
+  expect_output(print(df), "1 obs1\\s+1")
+  expect_output(print(df), "2 obs2\\s+2")
 })
 
 test_that("dataset_df supports multiple authors in APA formatting", {
@@ -152,3 +152,4 @@ test_that("dataset_df handles empty input gracefully", {
 test_that("dataset_df rejects invalid variable types", {
   expect_error(dataset_df(a = new.env()), "must be vectors")
 })
+
