### 2025-08-23 c88c9e3
Change the default observation prefix so that CoolURIs work.
