+++ R/fix_contributor.R
@@ -1,55 +1,48 @@
-#' @title Format contributor list into citation-friendly string
-#' @description Given a list of contributors (`utils::person` objects), this
-#' function formats them into a LaTeX/BibLaTeX-compatible string, merging roles
-#' per person and formatting names.
+#' Format contributors into a citation string
 #'
-#' Contributors without explicit roles are assumed to have the role `"ctb"`. If
-#' the input is `NULL` or the special string `":unas"`, the function returns
-#' `":unas"`.
+#' Format a list of `utils::person` objects into a compact string, merging roles
+#' per person and normalizing names. Contributors without explicit roles are
+#' assigned `"ctb"`. If `NULL` or `":unas"` is supplied, returns `":unas"`.
 #'
-#' This formatting is intended for metadata serialization or citation exports.
+#' @param contributors A vector or list of `person` objects, or `NULL`, or the
+#'   character string `":unas"`.
 #'
-#' @param contributors A vector of `person` objects, or the character string
-#'   `":unas"`, or `NULL`.
-#' @return A character string like `{Jane Doe [dtm, ctb]} and {John Smith
-#'   [ctb]}`.
+#' @return A single character string, e.g. `"{Jane Doe [dtm, ctb]} and
+#'  {John Smith [ctb]}"`.
+#'
 #' @examples
-#' \dontrun{
 #' fix_contributor(person("Jane", "Doe", role = c("dtm", "ctb")))
-#' fix_contributor(c(
+#' fix_contributor(list(
 #'   person("Jane", "Doe", role = "ctb"),
 #'   person("John", "Smith")
 #' ))
 #' fix_contributor(":unas")
-#' }
+#'
 #' @keywords internal
 fix_contributor <- function(contributors = NULL) {
-  if (is.null(contributors) || (is.character(contributors) && length(contributors) == 1 && contributors[[1]] == ":unas")) {
+  if (is.null(contributors) ||
+      (is.character(contributors) && length(contributors) == 1 && contributors == ":unas")) {
     return(":unas")
   }
 
-  # Normalize single person to a list
+  # Normalize into list of persons
   if (inherits(contributors, "person")) {
     contributors_list <- as.list(contributors)
-  } else if (is.list(contributors) && all(vapply(contributors, function(p) inherits(p, "person"), logical(1)))) {
+  } else if (is.list(contributors) && all(vapply(contributors, inherits, "person", FUN.VALUE = logical(1)))) {
     contributors_list <- contributors
   } else {
     return(":unas")
   }
 
-  # Build identity keys: "given|family"
+  # Group by identity (given|family)
   identity_keys <- vapply(contributors_list, function(p) {
     paste0(p$given %||% "", "|", p$family %||% "")
   }, character(1))
 
-  unique_keys <- unique(identity_keys)
-  merged <- vector("character", length(unique_keys))
-
-  for (i in seq_along(unique_keys)) {
-    key <- unique_keys[i]
+  merged <- vapply(unique(identity_keys), function(key) {
     matches <- contributors_list[identity_keys == key]
 
-    # Collect roles (default to "ctb")
+    # Collect unique roles (default to ctb)
     roles <- unique(unlist(lapply(matches, function(p) {
       if (is.null(p$role) || length(p$role) == 0 || all(!nzchar(p$role))) {
         "ctb"
@@ -59,13 +52,40 @@ fix_contributor <- function(contributors = NULL) {
     })))
 
     base <- matches[[1]]
-    name_parts <- c(base$given, base$family)
-    name_parts <- name_parts[!is.null(name_parts) & nzchar(name_parts)]
[…]

[…]
