### 2025-08-24 #681 @maurolepore
@antaldaniel,

- ml07. Thanks! I now see the expected "Articles" dropdown on the published version of the website, so it seems the main branch is now updated.

- ml26. Thanks! The bot seems to work fine now and all checks pass. On GitHub Actions the pkgcheck and R-CMD-check workflows are green. With all that, I'm not worried about the other workflows.

Also, thanks for your other responses. It all looks good to me, and I'll go ahead and post the review template. After that, I'll touch base with the editorial board to update everyone on the status of this package. Since our process here has been a bit unconventional, we had agreed to discuss once we reached this point to collectively decide if there is anything else we need for the final editorial approval.
