### 2025-08-24 #681 @maurolepore
I'll now touch base with the editorial board to update everyone on the status of this package. Since our process here has been a bit unconventional, we had agreed to discuss once we reached this point to collectively decide if there is anything else we need for the final editorial approval.
