### 2025-08-25 c0fcb6e
internal function for schema.org mapping
