### 2025-08-25 ffec8aa
Allow 'nt' abbreviation to be called 'ntriples' or 'N-Triples'
