### 2025-08-25 c08abc4
future function, not part of package
