+++ data-raw/crosswalk_bib.R
@@ -4,6 +4,48 @@
 #' @param schema One of "schema.org" (default), "dataspice"
 #' @return A named list that matches schema.org/Dataset JSON-LD
 #' @keywords import
+# ---- helpers ---------------------------------------------------------------
+
+# Map R/CRAN-style roles to a compact schema-ish vocabulary
+
+
+# Format utils::person to plain name, stripping trailing "[role]" that format() adds
+clean_person_name <- function(p) {
+  nm <- format(p)
+  sub("\\s*\\[[^]]+\\]$", "", nm)
+}
+
+# Safe extractor for structured subject and relation
+.extract_subject_term <- function(x) {
+  subj <- attr(x, "subject", exact = TRUE)
+  if (!is.null(subj) && inherits(subj, "subject") && length(subj$term)) {
+    subj$term
+  } else if (!is.null(x$subject) && length(x$subject)) {
+    as.character(x$subject)
+  } else {
+    NA_character_
+  }
+}
+
+.extract_related_identifier <- function(x) {
+  rel <- attr(x, "relation", exact = TRUE)
+  if (!is.null(rel) && inherits(rel, "related") && length(rel$relatedIdentifier)) {
+    rel$relatedIdentifier
+  } else if (!is.null(x$relation) && length(x$relation)) {
+    as.character(x$relation)
+  } else {
+    NA_character_
+  }
+}
+
+# ---- main crosswalk --------------------------------------------------------
+
+#' Crosswalk bibliographic metadata to schema.org / dataspice
+#'
+#' @param x A `dublincore` or `datacite` object produced by your helpers
+#' @param target One of "schema.org" or "dataspice"
+#' @return For "schema.org": a named list ready to be JSON‑LD.
+#'         For "dataspice": a list with data frames: $biblio, $creators, $contributors.
 crosswalk_bib <- function(x, target = c("schema.org", "dataspice")) {
   target <- match.arg(target)
 
@@ -11,89 +53,120 @@ crosswalk_bib <- function(x, target = c("schema.org", "dataspice")) {
     stop("Unsupported metadata class")
   }
 
+  # Unified view over both classes
   unified <- list(
     title       = x$title,
     description = x$description %||% NA_character_,
     identifier  = x$identifier %||% NA_character_,
-    creator     = x$author %||% list(),
-    contributor = attr(x, "contributor") %||% list(),
+    creator     = if (is.null(x$author)) list() else {
+      # ensure a list of person
+      if (inherits(x$author, "person")) as.list(x$author) else x$author
+    },
+    contributor = attr(x, "contributor", exact = TRUE) %||% list(),
     publisher   = x$publisher %||% NA_character_,
     date        = x$date %||% x$year %||% NA_character_,
     language    = x$language %||% NA_character_,
-    relation    = attr(x, "relation") %||% x$relation %||% NA_character_,
-    subject     = attr(x, "subject") %||% x$subject %||% NA_character_
+    keywords    = .extract_subject_term(x),
+    relation    = .extract_related_identifier(x)
   )
 
-  if (target == "schema.org") {
+  if (identical(target, "schema.org")) {
     sch <- list(
-      "@context"   = "https://schema.org",
-      "@type"      = "Dataset",
-      "name"       = unified$title,
-      "description"= unified$description,
-      "identifier" = unified$identifier,
-      "publisher"  = list("@type"="Organization","name"=unified$publisher),
-      "datePublished" = unified$date,
-      "inLanguage" = unified$language
+      "@context"      = "https://schema.org",
+      "@type"         = "Dataset",
+      name            = unified$title,
+      description     = unified$description,
+      identifier      = unified$identifier,
[…]
