+++ man/clean_person_name.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fix_contributor.R
+\name{clean_person_name}
+\alias{clean_person_name}
+\title{Remove role suffixes from formatted person names}
+\usage{
+clean_person_name(p)
+}
+\arguments{
+\item{p}{A \code{person} object.}
+}
+\value{
+Character string without role annotations, e.g. \code{"Jane Doe"}.
+}
+\description{
+Remove role suffixes from formatted person names
+}
+\keyword{internal}
