### 2025-08-25 179a181
replace URL that showed failure on CRAN
