+++ vignettes/rdf.Rmd
@@ -48,7 +48,7 @@ small_geo <- dataset_df(
     gdp$geo[1:3],
     label = "Geopolitical entity",
     concept = "http://purl.org/linked-data/sdmx/2009/dimension#refArea",
-    namespace = "https://dd.eionet.europa.eu/vocabulary/eurostat/geo/$1"
+    namespace = "https://www.geonames.org/countries/$1/"
   ), 
   identifier = c(
     obs = "https://dataset.dataobservatory.eu/examples/dataset.html#")
@@ -123,7 +123,6 @@ writeLines(ntriples, con = temp_file)
 rdf_graph <- rdf()
 rdf_parse(rdf_graph, doc = temp_file, format = "ntriples")
 rdf_graph
-
 ```
 
 A simple, serverless scaffolding for publishing `dataset_df` objects on the web (with HTML + RDF exports) is available at    <https://github.com/dataobservatory-eu/dataset-template> with the example
