+++ vignettes/introduction.Rmd
@@ -60,7 +60,7 @@ gdp <- defined(
 geo <- defined(
   rep("AD", 3),
   label = "Geopolitical Entity",
-  concept = "http://dd.eionet.europa.eu/vocabulary/eurostat/geo/",
+  concept = "http://purl.org/linked-data/sdmx/2009/dimension#refArea",
   namespace = "https://www.geonames.org/countries/$1/"
 )
 
@@ -103,7 +103,8 @@ Our packages tries to lower the burden of such retrofitting with early binding a
 You can convert any `dataset_df` object into a tidy 3-column representation (subject–predicate–object) using `dataset_to_triples()`:
 
 ```{r triplesexample}
-triples <- dataset_to_triples(small_dataset, format = "ntriples")
+triples <- dataset_to_triples(small_dataset, 
+                              format = "nt")
 triples
 ```
 
