+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames for Exchange and Reuse
 Version: 0.4.0
-Date: 2025-08-23
+Date: 2025-08-25
 Language: en-GB
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 
