### 2025-08-25 b6f6c1b
codemeta + CRAN comments
