### 2025-08-27 #681 @maurolepore
Thanks @antaldaniel for the update.

From my perspective as an editor I see you're committed to being transparent and to staying on top of CRAN requirements.

Also the editorial board is collectively happy to see that you have responded in positive ways both to my feedback here and that of a real user at https://github.com/dataobservatory-eu/dataset/issues/39. We think the work you have put in deserves a positive outcome and we would love to accept and welcome the dataset package to rOpenSci.

Next I'll instruct the bot to formalize the approval. It'll respond with a checklist for you to follow. That will automatically close this issue but feel free to continue to ask questions are you work through the checklist.

Congratulations!
