### 2025-08-27 #681 @ropensci-review-bot
Approved! Thanks @antaldaniel for submitting and @maurolepore for your reviews! :grin:
