### 2025-08-27 #681 @antaldaniel
Dear @mpadge and @maurolepore, thank you for your patience with me, and I am honoured and thankful.  I will follow, and hopefully not make clumsy mistakes with the package transfer.

It is very lucky that your approval came almost in the same minute as that of CRAN, so the same version was approved in both places. The CRAN issue was somewhat different. The dataset package’s tests already account for CRAN’s URL checks (including requests without a user-agent). The only issue arose from example.com, which always returns 404 by design (RFC 2606). In my view, CRAN should ideally treat example.com differently, because that namesapce is reserved for testing and documentation, and it is by design should not resolve to anything; but I have replaced these with resolvable example URIs.

I will also publish a project for a 0.5.0 next release with plans to integrate better with other rOpenSci packages, and I hope that I will find some contributors for that. And I would also like to develop the [Design Principles & Future Work Semantically Enriched, Standards-Aligned Datasets in R](https://dataset.dataobservatory.eu/articles/design.html) vignette into a journal article.
