### 2025-09-03 #681 @antaldaniel
Thanks! I’d be glad to write a blogpost about dataset. I’m thinking of starting with a short technical introduction for R users, and later developing a longer case study around publishing datasets with some real, new, scientific data that I got from a data partner. I’ll follow the [rOpenSci blog guide](https://blogguide.ropensci.org/?utm_source=chatgpt.com) and look forward to coordinating with @ropensci/blog-editors on timing and scope.
