### 2025-09-04 #681 @steffilazerte
I think that's fine! My only request would be that you add a bit of 'story' to your shorter blog post. We try to ensure that posts of this nature are a bit more than just a tutorial. Consider adding a bit of background motivation for making the package, or about your experiences in the process or review. Thanks!
